% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment_params_to_posterior_trees.R
\name{alignment_params_to_posterior_trees}
\alias{alignment_params_to_posterior_trees}
\title{Creates a posterior of phylogenies from a known alignment.}
\usage{
alignment_params_to_posterior_trees(alignment_params,
  experiment = create_experiment())
}
\arguments{
\item{alignment_params}{parameters to simulate an alignment,
as can be created by \link{create_alignment_params}}

\item{experiment}{a \link{pirouette} experiment,
as can be created by \link{create_experiment}}
}
\value{
a list of phylogenies in the posterior,
  as a \link[ape]{multiphylo}
}
\description{
Ignores the jointly-estimated posterior estimates
}
\author{
Richel J.C. Bilderbeek
}
