% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment_to_posterior.R
\name{alignment_to_posterior_trees}
\alias{alignment_to_posterior_trees}
\title{Creates a posterior from an alignment}
\usage{
alignment_to_posterior_trees(alignment, inference_params)
}
\arguments{
\item{alignment}{a DNA alignment}

\item{inference_params}{parameters for the Bayesian inference,
as can be created by \link{create_inference_params}.
The Bayesian inference is handled by the \link[babette]{babette}
R package, that calls the phylogenetic tool \code{BEAST2}.}
}
\value{
a list of:
\itemize{
  \item \code{trees}: the phylogenies in the posterior,
    as a \link[ape]{multiphylo}
  \item \code{estimates}: the BEAST2 estimates, as a \link{data.frame}
}
}
\description{
Creates a posterior from an alignment
}
\author{
Richel J.C. Bilderbeek
}
