% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bd_phylo_2_l_table}
\alias{bd_phylo_2_l_table}
\title{Convert bd phylo to L table}
\usage{
bd_phylo_2_l_table(phylo)
}
\arguments{
\item{phylo}{a phylogeny of class \link[ape]{phylo}}
}
\value{
the L table
}
\description{
Convert bd phylo to L table. Don't use for mbd.
}
\examples{
phylogeny <- ape::read.tree(text = "((A:2, B:2):1, C:3);")

l_table <- bd_phylo_2_l_table(phylogeny)

library(testthat)
expect_equal("matrix", class(l_table))
expect_true("birth_time" \%in\% colnames(l_table))
expect_true("parent" \%in\% colnames(l_table))
expect_true("id" \%in\% colnames(l_table))
expect_true("death_time" \%in\% colnames(l_table))
}
\author{
Xu Liang, Giovanni Laudanno, Richèl J.C. Bilderbeek
}
