% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_alignment_params.R
\name{check_alignment_params_names}
\alias{check_alignment_params_names}
\title{Checks if the list elements' names match that
of a valid \code{alignment_params}}
\usage{
check_alignment_params_names(alignment_params)
}
\arguments{
\item{alignment_params}{parameters to simulate an alignment,
as can be created by \link{create_alignment_params}}
}
\value{
nothing. Will \link{stop} if not
}
\description{
Will \link{stop} if not.
}
\author{
Richèl J.C. Bilderbeek
}
