% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_beast2_installed.R
\name{check_beast2_installed}
\alias{check_beast2_installed}
\title{Checks if BEAST2 is installed}
\usage{
check_beast2_installed()
}
\value{
nothing
}
\description{
Will \link{stop} if not.
}
\examples{
# Check cleanup by other functions
beastier::check_empty_beaustier_folders()

# Need to check if it is installed, else it will give the desired error
if (beastier::is_beast2_installed()) {
  check_beast2_installed()
}

}
\author{
Richèl J.C. Bilderbeek
}
