% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_error_measure_params.R
\name{check_error_measure_params}
\alias{check_error_measure_params}
\title{Checks if the argument is a valid error_measure parameters structure,
as created by \link{create_error_measure_params}.
Will \link{stop} if not.}
\usage{
check_error_measure_params(error_measure_params)
}
\arguments{
\item{error_measure_params}{parameter set to specify how the
error between the given phylogeny and the Bayesian
posterior is determined.
Use \link{create_error_measure_params} to create such
a parameter set}
}
\value{
nothing. Will \link{stop} if nit
}
\description{
Checks if the argument is a valid error_measure parameters structure,
as created by \link{create_error_measure_params}.
Will \link{stop} if not.
}
\examples{
  library(testthat)

  expect_silent(check_error_measure_params(create_error_measure_params()))
  expect_error(check_error_measure_params("nonsense"))
  expect_error(check_error_measure_params(NA))
  expect_error(check_error_measure_params(NULL))
}
\author{
Richèl J.C. Bilderbeek, Giovanni Laudanno
}
