% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_experiment.R
\name{check_experiment}
\alias{check_experiment}
\title{Checks if the argument is a valid \link{pirouette} experiment.}
\usage{
check_experiment(experiment)
}
\arguments{
\item{experiment}{a \link{pirouette} experiment,
as can be created by \link{create_experiment}}
}
\value{
nothing. Will \link{stop} if not
}
\description{
Will \link{stop} if not.
A valid \link{pirouette} experiment
can be created by \link{create_experiment}.
}
\author{
Richel J.C. Bilderbeek
}
