% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_experiments.R
\name{check_experiments}
\alias{check_experiments}
\title{Checks if the argument is a list of one or more \link{pirouette} experiments.}
\usage{
check_experiments(experiments)
}
\arguments{
\item{experiments}{a list of one or more \link{pirouette} experiments,
as can be created by \link{create_experiment}}
}
\value{
nothing. Will \link{stop} if not
}
\description{
Will \link{stop} if not.
A valid \link{pirouette} experiment
can be created by \link{create_experiment}.
}
\examples{
 experiment <- create_experiment()
 testthat::expect_silent(check_experiments(list(create_experiment())))
}
\author{
Richel J.C. Bilderbeek
}
