% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/check_experiments_all_inference_models_are_unique.R
\name{check_experiments_all_inference_models_are_unique}
\alias{check_experiments_all_inference_models_are_unique}
\title{Check if all experiments have unique inference models.}
\usage{
check_experiments_all_inference_models_are_unique(experiments)
}
\arguments{
\item{experiments}{a list of one or more \link{pirouette} experiments,
as can be created by \link{create_experiment}. If more than one experiment
is provided and a "generative" experiment is part of them, the "generative"
one has to be the first in the list.}
}
\description{
Will \link{stop} if two models have a same site and clock and tree prior.
Note that experiments that differ in their MRCA priors only
are classified being the same.
}
\author{
Richèl J.C. Bilderbeek
}
