% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_inference_model.R
\name{check_inference_model}
\alias{check_inference_model}
\title{Check if \code{x} is a valid inference model}
\usage{
check_inference_model(x)
}
\arguments{
\item{x}{object to be checked to be an inference model,
as can be created by \link{create_inference_model}}
}
\value{
nothing.
  Calls \link{stop} if \code{x} is not a valid inference model
}
\description{
Calls \link{stop} if \code{x} is not a valid inference model,
will do nothing otherwise
}
\author{
Richel J.C. Bilderbeek
}
