% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_inference_model_type_names.R
\name{check_inference_model_type_names}
\alias{check_inference_model_type_names}
\title{Check the model type names}
\usage{
check_inference_model_type_names(model_type_names)
}
\arguments{
\item{model_type_names}{one or more model type names, each
element must equal a value in \link{get_model_type_names}}
}
\value{
nothing
}
\description{
Check the model type names
}
\author{
Richèl J.C. Bilderbeek
}
