% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_inference_param.R
\name{check_inference_param}
\alias{check_inference_param}
\title{Check if the inference parameters are valid.
Throws if these are invalid}
\usage{
check_inference_param(inference_param)
}
\arguments{
\item{inference_param}{one set of parameters for the Bayesian inference,
as can be created by \link{create_inference_param}.
The Bayesian inference is handled by the \link[babette]{babette}
R package, that calls the phylogenetic tool \code{BEAST2}.}
}
\description{
Check if the inference parameters are valid.
Throws if these are invalid
}
\author{
Richel J.C. Bilderbeek
}
