% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_inference_param.R
\name{check_inference_params}
\alias{check_inference_params}
\title{Check if the inference parameters are valid.
Throws if these are invalid}
\usage{
check_inference_params(inference_param)
}
\description{
Check if the inference parameters are valid.
Throws if these are invalid
}
\author{
Richel J.C. Bilderbeek
}
