% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_init_pir_params.R
\name{check_init_pir_params}
\alias{check_init_pir_params}
\title{Check if the \code{pir_params} is initialized}
\usage{
check_init_pir_params(pir_params)
}
\arguments{
\item{pir_params}{the parameters of \link[pirouette]{pirouette}.
They are created by \link{create_pir_params}.}
}
\value{
nothing
}
\description{
Will \link{stop} if not
}
\author{
Richèl J.C. Bilderbeek
}
