% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_model_select_param.R
\name{check_model_select_param}
\alias{check_model_select_param}
\title{Checks if the argument is a valid model selection parameters structure,
as created by \link{create_model_select_param}.
Will \link{stop} if not.}
\usage{
check_model_select_param(model_select_param)
}
\arguments{
\item{model_select_param}{one parameter set to select an inference model
(a combination of site model, clock model and tree prior),
as can be created by \link{create_model_select_param}}
}
\value{
nothing. Will \link{stop} if the object is not
}
\description{
Checks if the argument is a valid model selection parameters structure,
as created by \link{create_model_select_param}.
Will \link{stop} if not.
}
\author{
Richel J.C. Bilderbeek
}
