% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_model_select_params.R
\name{check_model_select_params}
\alias{check_model_select_params}
\title{Checks if the argument is a valid model selection parameters structure,
as created by \link{create_model_select_params}.
Will \link{stop} if not.}
\usage{
check_model_select_params(model_select_params)
}
\arguments{
\item{model_select_params}{parameters to select a model (a combination
of site model, clock model and tree prior),
as can be created by \link{create_model_select_params}}
}
\value{
nothing. Will \link{stop} if the object is not
}
\description{
Checks if the argument is a valid model selection parameters structure,
as created by \link{create_model_select_params}.
Will \link{stop} if not.
}
\author{
Richel J.C. Bilderbeek
}
