% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_mutation_rate.R
\name{check_mutation_rate}
\alias{check_mutation_rate}
\title{Check if the mutation rate is valid}
\usage{
check_mutation_rate(mutation_rate)
}
\arguments{
\item{mutation_rate}{the mutation rate per base pair per time unit.
Use \link{check_mutation_rate} to check if a mutation rate is valid.}
}
\description{
Will \link{stop} if not
}
