% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pir_out.R
\name{check_pir_out}
\alias{check_pir_out}
\title{Checks if the argument is a valid \link{pirouette} parameter set.}
\usage{
check_pir_out(pir_out)
}
\arguments{
\item{pir_out}{the output of \link{pir_run}}
}
\value{
nothing. Will \link{stop} if not
}
\description{
Will \link{stop} if not.
A valid \link{pirouette} parameter set
can be created by \link{create_pir_params}.
}
\details{
A \code{pir_out} contains:

\enumerate{
  \item tree true or twin tree
  \item inference_model generative or candidate
  \item inference_model_weight
  \item site_model the site model name
  \item clock_model the clock model name
  \item tree_prior the tree model name
  \item error_1, error_2, etcetera: inference errors
}
}
\examples{
check_pir_out(create_test_pir_run_output())
}
\author{
Richèl J.C. Bilderbeek
}
