% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pir_out_errors_above_zero.R
\name{check_pir_out_errors_above_zero}
\alias{check_pir_out_errors_above_zero}
\title{Checks that \code{pir_out} errors all are more than zero}
\usage{
check_pir_out_errors_above_zero(pir_out)
}
\arguments{
\item{pir_out}{the output of \link{pir_run}}
}
\description{
Checks that \code{pir_out} errors all are more than zero
}
