% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pir_out.R
\name{check_pir_out_names}
\alias{check_pir_out_names}
\title{Checks if the \code{pir_out} has elements with
the names needed.}
\usage{
check_pir_out_names(pir_out)
}
\arguments{
\item{pir_out}{the output of \link{pir_run}}
}
\value{
nothing. Will \link{stop} if not
}
\description{
Will \link{stop} if not.
}
\author{
Richèl J.C. Bilderbeek
}
