% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pir_params.R
\name{check_pir_params}
\alias{check_pir_params}
\title{Checks if the argument is a valid \link{pirouette} parameter set.}
\usage{
check_pir_params(pir_params)
}
\arguments{
\item{pir_params}{the parameters of \link[pirouette]{pirouette}.
They are created by \link{create_pir_params}.}
}
\value{
nothing. Will \link{stop} if not
}
\description{
Will \link{stop} if not.
A valid \link{pirouette} parameter set
can be created by \link{create_pir_params}.
}
\examples{
if (beautier::is_on_ci()) {
  check_pir_params(create_test_pir_params())
}
}
\author{
Giovanni Laudanno, Richèl J.C. Bilderbeek
}
