% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pir_params.R
\name{check_pir_params_data_types}
\alias{check_pir_params_data_types}
\title{Checks if the \code{pir_params} elements are all of the right data type.}
\usage{
check_pir_params_data_types(pir_params)
}
\arguments{
\item{pir_params}{the parameters of \link[pirouette]{pirouette}.
They are created by \link{create_pir_params}.}
}
\value{
nothing. Will \link{stop} if not
}
\description{
Will \link{stop} if not.
A valid \link{pirouette} parameter set
can be created by \link{create_pir_params}.
}
\author{
Richèl J.C. Bilderbeek
}
