% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pir_paramses.R
\name{check_pir_paramses}
\alias{check_pir_paramses}
\title{Check if all elements in the list of \code{pir_params} are valid.}
\usage{
check_pir_paramses(pir_paramses)
}
\arguments{
\item{pir_paramses}{a list of \link[pirouette]{pirouette} parameters,
each element created by \link{create_pir_params}.}
}
\value{
nothing
}
\description{
Will \link{stop} if not
}
\author{
Richèl J.C. Bilderbeek
}
