% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_reconstructed_phylogeny.R
\name{check_reconstructed_phylogeny}
\alias{check_reconstructed_phylogeny}
\title{Check if the phylogeny is a reconstructed phylogeny.}
\usage{
check_reconstructed_phylogeny(phylogeny)
}
\arguments{
\item{phylogeny}{a phylogeny of class \link[ape]{phylo}}
}
\description{
Will \link{stop} if there are extinct species in the phylogeny
}
