% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_rename_fun.R
\name{check_rename_fun}
\alias{check_rename_fun}
\title{Check if the rename function is a valid filename rename function}
\usage{
check_rename_fun(rename_fun)
}
\description{
Will \link{stop} if not
}
\author{
Richèl J.C. Bilderbeek
}
