% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_result.R
\name{check_result}
\alias{check_result}
\title{Checks if the argument is a valid \link{pirouette} result.}
\usage{
check_result(result)
}
\arguments{
\item{result}{results from measurements. These are:
\itemize{
  \item log_evidence the natural logarithm of the evidence (aka marginal
    likelihood). Can be NA if this is not measured
  \item weight the weight of the model, compared to other (candidate)
    models. This weight will be between 0.0 (there is no evidence for
    this model) to 1.0 (all evidence indicates this is the best model).
    A weight of NA denotes that the weight is not measured
  \item errors a numeric vector of (positive) Bayesian inference errors.
    Will be NA if these are not measured.
}}
}
\value{
nothing. Will \link{stop} if not
}
\description{
Will \link{stop} if not.
A valid \link{pirouette} result
can be created by \link{create_result}.
}
\examples{
  library(testthat)

  expect_silent(check_result(create_test_result()))
  expect_error(check_result("nonsense"))
  expect_error(check_result(NA))
  expect_error(check_result(NULL))
}
\author{
Richel J.C. Bilderbeek
}
