% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_root_sequence.R
\name{check_root_sequence}
\alias{check_root_sequence}
\title{Check if the root sequence is valid}
\usage{
check_root_sequence(root_sequence)
}
\arguments{
\item{root_sequence}{the DNA sequence at the root of the phylogeny.
By default, this will consist out of an equal amount of each letter
Use \link{check_root_sequence} to check if a root sequence is valid.}
}
\description{
Will \link{stop} if not
}
