% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_run_experiments.R
\name{check_run_experiments}
\alias{check_run_experiments}
\title{Checks if the argument is a list of one or more
\link{pirouette} run experiments.}
\usage{
check_run_experiments(run_experiments)
}
\arguments{
\item{run_experiments}{a list of one or more \link{pirouette} run experiments}
}
\value{
nothing. Will \link{stop} if not
}
\description{
Will \link{stop} if not.
}
\examples{
  testthat::expect_silent(
    check_run_experiments(
      list(create_test_run_experiment())
    )
  )
  testthat::expect_error(
    check_run_experiments(
      create_test_run_experiment()
    )
  )
  testthat::expect_error(check_run_experiments("nonsense"))
  testthat::expect_error(check_run_experiments(NA))
  testthat::expect_error(check_run_experiments(NULL))
}
\seealso{
Use \link{check_run_experiment} to check if an object
  is one run experiment
}
\author{
Richel J.C. Bilderbeek
}
