% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_sim_true_alignment_function.R
\name{check_sim_true_alignment_function}
\alias{check_sim_true_alignment_function}
\title{Check if the function to create a true alignment with (from the
true phylogeny) is valid.}
\usage{
check_sim_true_alignment_function(sim_true_alignment_function)
}
\description{
Will \link{stop} if not
}
\seealso{
Use \link{check_sim_twin_alignment_function} to check a function to
generate a twin alignment
}
