% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_sim_twin_alignment_function.R
\name{check_sim_twin_alignment_function}
\alias{check_sim_twin_alignment_function}
\title{Check if the function to create a twin alignment with (from a
true phylogeny and a true alignment) is valid.}
\usage{
check_sim_twin_alignment_function(sim_twin_alignment_function)
}
\description{
Will \link{stop} if not
}
