% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_sim_twin_tree_fun.R
\name{check_sim_twin_tree_fun}
\alias{check_sim_twin_tree_fun}
\title{Check if the \code{sim_twin_tree_fun} is valid}
\usage{
check_sim_twin_tree_fun(sim_twin_tree_fun)
}
\arguments{
\item{sim_twin_tree_fun}{function to simulate a twin tree with.
This function must have one argument called \code{phylogeny}
of type \link[ape]{phylo} and have a return type of type \link[ape]{phylo}
as well.

Some standard functions:\cr
\itemize{
  \item Use \link{create_sim_yule_twin_tree_fun} to use a
    Yule (aka Pure Birth) process
  \item Use \link{create_copy_twtr_from_true_fun} to for a function
    that copies the true tree
  \item Use \link{get_sim_bd_twin_tree_fun} to use a
    Birth-Death process
}}
}
\value{
nothing
}
\description{
Will \link{stop} if not
}
\seealso{
Use \link{check_sim_twal_fun} to check a function to
generate a twin alignment.
}
\author{
Richèl J.C. Bilderbeek
}
