% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_sim_twin_tree_function.R
\name{check_sim_twin_tree_function}
\alias{check_sim_twin_tree_function}
\title{Check if the \code{sim_twin_tree_function} is valid}
\usage{
check_sim_twin_tree_function(sim_twin_tree_function)
}
\arguments{
\item{sim_twin_tree_function}{function to simulate a twin tree with.
This function must have one argument called \code{phylogeny}
of type \link[ape]{phylo} and have a return type of type \link[ape]{phylo}
as well.

Some standard functions:\cr
\itemize{
  \item Use \link{create_sim_yule_twin_tree_function} to use a
    Yule (aka Pure Birth) process
  \item Use \link{create_copy_twin_tree_from_true_function} to for a function
    that copies the true tree
  \item Use \link{create_sim_bd_twin_tree_function} to use a
    Birth-Death process
}}
}
\description{
Will \link{stop} if not
}
