% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_tree_and_model.R
\name{check_tree_and_model}
\alias{check_tree_and_model}
\title{Check if the \code{tree_and_model} is valid}
\usage{
check_tree_and_model(tree_and_model)
}
\arguments{
\item{tree_and_model}{one combination of a tree and model,
as created by \link{get_tree_and_model_values}}
}
\value{
nothing
}
\description{
Check if the \code{tree_and_model} is valid
}
\author{
Richèl J.C. Bilderbeek
}
