% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_tree_and_model_errors.R
\name{check_tree_and_model_errors}
\alias{check_tree_and_model_errors}
\title{Plot the error 'BEAST2' makes from a known phylogeny
from tidy data}
\usage{
check_tree_and_model_errors(tree_and_model_errors)
}
\value{
a \code{ggplot2} plot
}
\description{
Plot the error 'BEAST2' makes from a known phylogeny
from tidy data
}
\author{
Richèl J.C. Bilderbeek
}
