% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_tree_and_models.R
\name{check_tree_and_models}
\alias{check_tree_and_models}
\title{Check if the \code{tree_and_model} is valid}
\usage{
check_tree_and_models(tree_and_models)
}
\arguments{
\item{tree_and_models}{one or more combination of a tree and model,
as created by \link{get_tree_and_model_values}}
}
\value{
a \code{ggplot2} plot
}
\description{
Check if the \code{tree_and_model} is valid
}
\author{
Richèl J.C. Bilderbeek
}
