% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_tree_type.R
\name{check_tree_type}
\alias{check_tree_type}
\title{Checks if tree type is valid
Will \link{stop} if not.}
\usage{
check_tree_type(tree_type)
}
\arguments{
\item{tree_type}{type of tree, can be \code{true} for the true
phylogeny, and \code{twin} for its twin tree}
}
\value{
nothing. Will \link{stop} if not
}
\description{
Checks if tree type is valid
Will \link{stop} if not.
}
\examples{
library(testthat)

expect_silent(check_tree_type(get_tree_types()[1]))
expect_silent(check_tree_type(get_tree_types()[2]))
expect_error(check_tree_type("nonsense"))
expect_error(check_tree_type(NA))
expect_error(check_tree_type(NULL))
}
\author{
Giovanni Laudanno, Richèl J.C. Bilderbeek
}
