% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_tree_types.R
\name{check_tree_types}
\alias{check_tree_types}
\title{Checks if the tree types are valid
Will \link{stop} if not.}
\usage{
check_tree_types(tree_types)
}
\arguments{
\item{tree_types}{types of tree, a vector of \code{true} for a true
phylogeny, and \code{twin} for a twin tree}
}
\value{
nothing. Will \link{stop} if not
}
\description{
Checks if the tree types are valid
Will \link{stop} if not.
}
\examples{
library(testthat)

expect_silent(check_tree_types(get_tree_types()))
expect_error(check_tree_types("nonsense"))
expect_error(check_tree_types(NA))
expect_error(check_tree_types(NULL))
}
\author{
Richèl J.C. Bilderbeek
}
