% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_twin_phylogeny.R
\name{check_twin_phylogeny}
\alias{check_twin_phylogeny}
\title{Check if a twin phylogeny is a valid phylogeny}
\usage{
check_twin_phylogeny(twin_phylogeny)
}
\arguments{
\item{twin_phylogeny}{a phylogeny of class \link[ape]{phylo}}
}
\description{
Check if a twin phylogeny is a valid phylogeny
}
\seealso{
Use \link[beautier]{check_phylogeny}
for checking phylogenies in general
}
\author{
Richèl J.C. Bilderbeek
}
