% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_twinning_params_list_elements.R
\name{check_twinning_params_list_elements}
\alias{check_twinning_params_list_elements}
\title{Check that the \code{twinning_params} has all the list elements}
\usage{
check_twinning_params_list_elements(twinning_params)
}
\description{
Check that the \code{twinning_params} has all the list elements
}
