% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_twinning_params_names.R
\name{check_twinning_params_names}
\alias{check_twinning_params_names}
\title{Check that the \code{twinning_params} has all the list elements' names}
\usage{
check_twinning_params_names(twinning_params)
}
\arguments{
\item{twinning_params}{can be \code{NA} if no twinning is desired,
or can be the twinning parameters,
as can be created by \link{create_twinning_params}}
}
\value{
nothing
}
\description{
Check that the \code{twinning_params} has all the list elements' names
}
