% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_tree_and_model.R
\name{collapse_tree_and_model}
\alias{collapse_tree_and_model}
\title{Internal function}
\usage{
collapse_tree_and_model(tree_and_model)
}
\arguments{
\item{tree_and_model}{one combination of a tree and model,
as created by \link{get_tree_and_model_values}}
}
\value{
a releveled \code{tree_and_model}
}
\description{
Internal function to relevel the \code{tree_and_model}, so that
\link{pir_plot} has the legend labels in the right order
}
\author{
Richèl J.C. Bilderbeek
}
