% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_models.R
\name{collect_models}
\alias{collect_models}
\title{Create a list of all model combinations needed}
\usage{
collect_models(model_select_params)
}
\value{
a list of all model combinations. Each element contains
  one site model, one clock model and one tree prior
}
\description{
Create a list of all model combinations needed
}
\author{
Richel J.C. Bilderbeek
}
