% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_models.R
\name{combine_models}
\alias{combine_models}
\title{Combine a combination of site models, clock models and tree priors
into a collection of inference models}
\usage{
combine_models(
  site_models = beautier::create_site_models(),
  clock_models = beautier::create_clock_models(),
  tree_priors = beautier::create_tree_priors()
)
}
\description{
Combine a combination of site models, clock models and tree priors
into a collection of inference models
}
