% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_pir_out_to_long.R
\name{convert_pir_out_to_long}
\alias{convert_pir_out_to_long}
\title{Convert a \code{pir_out} to its long form}
\usage{
convert_pir_out_to_long(pir_out, verbose = FALSE)
}
\arguments{
\item{pir_out}{the output of \link{pir_run}}

\item{verbose}{if TRUE, show more output}
}
\value{
the \code{pir_out} in long form
}
\description{
Convert a \code{pir_out} to its long form
}
\examples{
pir_out <- create_test_pir_run_output(
  add_twin = TRUE,
  add_best = TRUE
)
convert_pir_out_to_long(pir_out)
}
\author{
Richèl J.C. Bilderbeek, Giovanni Laudanno
}
