% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_tree2brts}
\alias{convert_tree2brts}
\title{Convert a tree into branching times}
\usage{
convert_tree2brts(tree, precision = 8)
}
\arguments{
\item{tree}{an ultrametric phylogenetic tree of class \code{phylo}}

\item{precision}{define the precision of the approximation.}
}
\value{
the branching times
}
\description{
Convert a tree into branching times. Differently from the ape's
 function, it will keep the multiple events. Since the units are million
 years, a precision of 8 means that the approximation goes up to the 8-th
 digits. With such approximation we consider events happening within an
 interval of 4 days (1 million years / 10^8 = 1 year / 100) as simultaneous.
}
\author{
Giovanni Laudanno
}
