% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_n_mutation.R
\name{count_n_mutation}
\alias{count_n_mutation}
\title{Count mutations}
\usage{
count_n_mutation(alignment, root_sequence)
}
\arguments{
\item{alignment}{a DNA alignment}

\item{root_sequence}{the DNA sequence at the root of the phylogeny.
By default, this will consist out of an equal amount of each letter}
}
\value{
the number of mutations
}
\description{
Count mutations
}
\examples{
library(testthat)

# it works
alignment <- ape::as.DNAbin(
  x = list(species_1 = strsplit("aaaa", split = "")[[1]])
)
expect_equal(count_n_mutation(alignment, "acgt"), 3)

}
\author{
Giovanni Laudanno
}
