% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_alignment.R
\name{create_alignment_newskool}
\alias{create_alignment_newskool}
\title{Simulate an alignment from the true phylogeny}
\usage{
create_alignment_newskool(phylogeny,
  alignment_params = create_alignment_params(), verbose = FALSE)
}
\arguments{
\item{phylogeny}{a phylogeny of class \link[ape]{phylo}}

\item{alignment_params}{parameters to simulate an alignment,
as can be created by \link{create_alignment_params}}

\item{verbose}{if TRUE, show more output}
}
\value{
an alignment of type \code{DNAbin}
}
\description{
Simulate an alignment from the true phylogeny
}
\examples{
library(testthat)

# Create the phylogeny to simulate the alignment on
n_taxa <- 5
phylogeny <- ape::rcoal(n_taxa)

# Use default settings to create the alignment
alignment_params <- pirouette::create_alignment_params()

# Simulate the alignment
alignment <- create_alignment(
   phylogeny = phylogeny,
   alignment_params = alignment_params,
 )

expect_equal(class(alignment), "DNAbin")
expect_equal(nrow(alignment), n_taxa)
expect_equal(ncol(alignment), nchar(alignment_params$root_sequence))
}
\seealso{
Use \link{create_alignment_file} to save the simulated alignment
  directly to a file
}
\author{
Richèl J.C. Bilderbeek
}
