% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_alignment_params.R
\name{create_alignment_params}
\alias{create_alignment_params}
\title{Create the parameters for the alignment simulation}
\usage{
create_alignment_params(root_sequence, mutation_rate, rng_seed = 0)
}
\arguments{
\item{root_sequence}{the DNA sequence at the root of the phylogeny.
By default, this will consist out of only adenine}

\item{mutation_rate}{the mutation rate per base pair per time unit}

\item{rng_seed}{the random number generator seed as used in the
simulation of an alignment}
}
\value{
a list of alignment parameters
}
\description{
Create the parameters for the alignment simulation
}
\examples{
n_base_pairs <- 4
alignment_params <- create_alignment_params(
   root_sequence = create_blocked_dna(length = n_base_pairs),
   mutation_rate = 0.1
 )
 testit::assert("root_sequence" \%in\% names(alignment_params))
 testit::assert("mutation_rate" \%in\% names(alignment_params))
 testit::assert("rng_seed" \%in\% names(alignment_params))
}
\author{
Richel J.C. Bilderbeek
}
