% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_alignment_with_std_nsm.R
\name{create_alignment_with_std_nsm}
\alias{create_alignment_with_std_nsm}
\title{Create an alignment with a standard site model}
\usage{
create_alignment_with_std_nsm(phylogeny, alignment_params)
}
\arguments{
\item{phylogeny}{a phylogeny of class \link[ape]{phylo}}

\item{alignment_params}{parameters to simulate an alignment,
as can be created by \link{create_alignment_params}}
}
\value{
an alignment of type \code{DNAbin}
}
\description{
Create an alignment with a standard site model
}
