% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_model_select_param.R
\name{create_best_model_select_param}
\alias{create_best_model_select_param}
\title{Create model selection parameters
for the model with the model evidence (also: marginal likelihood)
from a set of any combination of site models, clock models and
tree priors.}
\usage{

  create_best_model_select_param(site_models = beautier::create_site_models(),
  clock_models = beautier::create_clock_models(),
  tree_priors = beautier::create_tree_priors())
}
\arguments{
\item{site_models}{a list of one or more site models,
as created by \link[beautier]{create_site_model}}

\item{clock_models}{a list of one or more clock models,
as created by \link[beautier]{create_clock_model}}

\item{tree_priors}{a list of one or more tree priors,
as created by \link[beautier]{create_tree_prior}}
}
\description{
Create model selection parameters
for the model with the model evidence (also: marginal likelihood)
from a set of any combination of site models, clock models and
tree priors.
}
\examples{
  # Pick the model with most evidence to be used in inference
  model_select_params <- create_best_model_select_param()
  # In such a case, multiple site models, clock models
  #  and tree priors are tested
  testthat::expect_true(length(model_select_params$site_models) > 1)
  testthat::expect_true(length(model_select_params$clock_models) > 1)
  testthat::expect_true(length(model_select_params$tree_priors) > 1)
}
\author{
Richel J.C. Bilderbeek
}
