% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_clock_model_from_name.R
\name{create_clock_model_from_name}
\alias{create_clock_model_from_name}
\title{Create a clock model from name}
\usage{
create_clock_model_from_name(clock_model_name)
}
\arguments{
\item{clock_model_name}{name of a clock model}
}
\description{
Create a clock model from name
}
\author{
Richel J.C. Bilderbeek
}
