% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_copy_twin_tree_from_true_function.R
\name{create_copy_twin_tree_from_true_function}
\alias{create_copy_twin_tree_from_true_function}
\title{Create a function that can simulate the twin tree from the true tree,
by just copying the true tree}
\usage{
create_copy_twin_tree_from_true_function()
}
\value{
a function
}
\description{
Create a function that can simulate the twin tree from the true tree,
by just copying the true tree
}
