% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_copy_twin_tree_from_true_fun.R
\name{create_copy_twtr_from_true_fun}
\alias{create_copy_twtr_from_true_fun}
\title{Create a function that can simulate the twin tree from the true tree,
by just copying the true tree}
\usage{
create_copy_twtr_from_true_fun()
}
\value{
a function
}
\description{
Create a function that can simulate the twin tree from the true tree,
by just copying the true tree
}
\examples{
f <- create_copy_twtr_from_true_fun()
phylo_in <- ape::read.tree(text = "((A:1, B:1):1, C:2);")
f(phylo_in)
}
\seealso{
Use \link{create_sim_yule_twin_tree_fun} to get a function to
produce a Yule tree.
Use \link{get_sim_bd_twin_tree_fun} to get a function to
produce a Birth-Death tree.
}
\author{
Richèl J.C. Bilderbeek
}
