% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_experiment.R
\name{create_experiment}
\alias{create_experiment}
\title{Create a \link{pirouette} experiment.}
\usage{
create_experiment(model_type = "generative", run_if = "always",
  do_measure_evidence = FALSE,
  inference_model = beautier::create_inference_model(),
  beast2_options = beastier::create_beast2_options())
}
\arguments{
\item{inference_model}{an inference model, which is a combination
of site model, clock model, tree prior and BEAST2 input and
input filenames.}
}
\value{
a \link{pirouette} experiment.
}
\description{
Create a \link{pirouette} experiment.
}
\examples{
 experiment <- create_experiment()
 testit::assert("rng_seed" \%in\% names(experiment))
 testit::assert("twin_tree_filename" \%in\% names(experiment))
}
\author{
Richel J.C. Bilderbeek
}
