% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_fig_2.R
\name{create_fig_2}
\alias{create_fig_2}
\title{Create figure 2 for pirouette article}
\usage{
create_fig_2()
}
\value{
a ggplot2 plot
}
\description{
Create figure 2 for pirouette article
}
\author{
Richel J.C. Bilderbeek, Giovanni Laudanno
}
