% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_fig_3.R
\name{create_fig_3}
\alias{create_fig_3}
\title{Create figure 3 for pirouette article}
\usage{
create_fig_3()
}
\value{
a ggplot2 plot
}
\description{
Create figure 3 for pirouette article
}
\author{
Richel J.C. Bilderbeek, Giovanni Laudanno
}
