% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_fig_4.R
\name{create_fig_4}
\alias{create_fig_4}
\title{Create figure 4 for pirouette article}
\usage{
create_fig_4()
}
\value{
a ggplot2 plot
}
\description{
Create figure 4 for pirouette article
}
\author{
Richel J.C. Bilderbeek, Giovanni Laudanno
}
