% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_fig_4_file.R
\name{create_fig_4_file}
\alias{create_fig_4_file}
\title{Create figure 4 file}
\usage{
create_fig_4_file(project_folder_name = getwd())
}
\arguments{
\item{project_folder_name}{project folder name}
}
\value{
the name of the file the plot is saved to
}
\description{
Create figure 4 file
}
\author{
Richel J.C. Bilderbeek, Giovanni Laudanno
}
