% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_fig_bd.R
\name{create_fig_bd}
\alias{create_fig_bd}
\title{Create figure bd for pirouette article}
\usage{
create_fig_bd()
}
\value{
a ggplot2 plot
}
\description{
Create figure bd for pirouette article
}
\author{
Richel J.C. Bilderbeek, Giovanni Laudanno
}
