% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_model_select_param.R
\name{create_gen_model_select_param}
\alias{create_gen_model_select_param}
\title{Select the generative model to be used in inference}
\usage{
create_gen_model_select_param(alignment_params,
  tree_prior = beautier::create_bd_tree_prior())
}
\arguments{
\item{alignment_params}{parameters to simulate an alignment,
as can be created by \link{create_alignment_params}}

\item{tree_prior}{a tree prior,
as created by \link[beautier]{create_tree_prior}}
}
\description{
Create model selection parameters
in which the generative model of the alignmnent is (rightfully)
assumed to underly its creation.
The tree prior underlying the phylogeny is assumed to be a Birth-Death prior.
}
\examples{
  # Pick the generative model
  alignment_params <- create_alignment_params(
    root_sequence = "acgt", mutation_rate = 0.01
  )
  model_select_param <- create_gen_model_select_param(alignment_params)
  # In such a case, the site model and clock models of the alignment is
  # stored in the model selection parameters
  testthat::expect_equal(
    alignment_params$site_model,
    model_select_param$site_models[[1]]
  )
  testthat::expect_equal(
    alignment_params$clock_model,
    model_select_param$clock_models[[1]]
  )
  # By default, a birth-death model is assumed to underly the phylogeny
  testthat::expect_equal(
    model_select_param$tree_priors[[1]],
    beautier::create_bd_tree_prior()
  )
}
\seealso{
\link{create_model_select_param} contains an overview
  of all model selection possible
}
\author{
Richel J.C. Bilderbeek
}
