% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_model_selection_params.R
\name{create_gen_model_select_params}
\alias{create_gen_model_select_params}
\title{Create model selection parameters
in which the generative model of the alignmnent is (rightfully)
assumed to underly its creation.
The tree prior underlying the phylogeny is assumed to be a Birth-Death prior.}
\usage{
create_gen_model_select_params(alignment_params,
  tree_prior = beautier::create_bd_tree_prior())
}
\arguments{
\item{alignment_params}{parameters to simulate an alignment,
as can be created by \link{create_alignment_params}}

\item{tree_prior}{a tree prior,
as created by \link[beautier]{create_tree_prior}}
}
\description{
Create model selection parameters
in which the generative model of the alignmnent is (rightfully)
assumed to underly its creation.
The tree prior underlying the phylogeny is assumed to be a Birth-Death prior.
}
\author{
Richel J.C. Bilderbeek
}
