% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_inference_conditions.R
\name{create_inference_conditions}
\alias{create_inference_conditions}
\title{Create the inference conditions}
\usage{
create_inference_conditions(model_type = "generative",
  run_if = "always", do_measure_evidence = FALSE)
}
\arguments{
\item{model_type}{type of inference model supplied for an experiment.
Possible values:
\itemize{
  \item \code{generative}: the inference model is (or is assumed to be)
    the inference model underlying the phylogeny
  \item \code{candidate}: the inference model is a candidate model,
    that competes with other models for having the most
    evidence (aka highest marginal likelihood)
}}

\item{run_if}{the condition for an experiment's inference model to be run.
Possible values:
\itemize{
  \item \code{always}: always
  \item \code{best_candidate}: if the inference model is the
    candidate model with the most evidence (aka highest marginal
    likelihood)
}}

\item{do_measure_evidence}{boolean to indicate if the
evidence (aka marginal likelihood) of an experiment must be
measured}
}
\value{
the inference conditions
}
\description{
Create the parameters to determine how to choose
a model for the inference
}
\examples{
  library(testthat)

  model_type <- "candidate"
  run_if <- "best_candidate"
  do_measure_evidence <- TRUE

  inference_conditions <- create_inference_conditions(
    model_type = model_type,
    run_if = run_if,
    do_measure_evidence = do_measure_evidence
  )

  expect_true("model_type" \%in\% names(inference_conditions))
  expect_true("run_if" \%in\% names(inference_conditions))
  expect_true("do_measure_evidence" \%in\% names(inference_conditions))
}
\author{
Giovanni Laudanno
}
