% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_inference_model.R
\name{create_inference_model}
\alias{create_inference_model}
\title{Create an inference model.}
\usage{
create_inference_model(site_model, clock_model, tree_prior,
  input_filename = tempfile(fileext = ".xml"),
  log_filename = tempfile(fileext = ".log"),
  trees_filename = tempfile(fileext = ".trees"),
  state_filename = tempfile(fileext = ".xml.state"))
}
\arguments{
\item{site_model}{a nucleotide substitution model,
as created by \link[beautier]{create_site_model}}

\item{clock_model}{a clock model,
as created by \link[beautier]{create_clock_model}}

\item{tree_prior}{a tree prior,
as created by \link[beautier]{create_tree_prior}}

\item{trees_filename}{name of the BEAST2 posterior phylogenies file}
}
\value{
an inference model
}
\description{
An inference model is a combination of a site model, clock
model and tree prior. An inference model is also the part
of the BEAST2 setups that differ; the part that is shared
is created by \link{create_inference_param})
}
\author{
Richel J.C. Bilderbeek
}
