% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_inference_params.R
\name{create_inference_params}
\alias{create_inference_params}
\title{Creates a set of parameters used in the \link[babette]{babette} inference}
\usage{
create_inference_params(mcmc, rng_seed = 0)
}
\arguments{
\item{mcmc}{MCMC options, as created by \link[beautier]{create_mcmc}}
}
\description{
Creates a set of parameters used in the \link[babette]{babette} inference
}
\author{
Richel J.C. Bilderbeek
}
