% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_long_pir_out_from_folder.R
\name{create_long_pir_out_from_folder}
\alias{create_long_pir_out_from_folder}
\title{Plot the error 'BEAST2' makes from a known phylogeny
from tidy data}
\usage{
create_long_pir_out_from_folder(folder_name)
}
\arguments{
\item{folder_name}{name of the main folder}
}
\value{
a \code{ggplot2} plot
}
\description{
Plot the error 'BEAST2' makes from a known phylogeny
from tidy data
}
\author{
Richèl J.C. Bilderbeek
}
