% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_long_pir_out_from_folders.R
\name{create_long_pir_out_from_folders}
\alias{create_long_pir_out_from_folders}
\title{Plot the error 'BEAST2' makes from a known phylogeny
from tidy data}
\usage{
create_long_pir_out_from_folders(folder_names)
}
\arguments{
\item{folder_names}{one or more folder names}
}
\value{
a \code{ggplot2} plot
}
\description{
Plot the error 'BEAST2' makes from a known phylogeny
from tidy data
}
\author{
Richèl J.C. Bilderbeek
}
