% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_model_select_param.R
\name{create_model_select_param}
\alias{create_model_select_param}
\title{Create model selection parameters.}
\usage{
create_model_select_param(type,
  site_models = beautier::create_site_models(),
  clock_models = beautier::create_clock_models(),
  tree_priors = beautier::create_tree_priors(), epsilon = 1e-12,
  verbose = FALSE, marg_lik_filename = tempfile(fileext = ".csv"))
}
\arguments{
\item{type}{one or more ways to select the models used in inference:
\itemize{
  \item \code{"generative"}: pick the generative model
  \item \code{most_evidence} picks the model with most evidence
}
See \link{get_model_selections} for a list.}

\item{site_models}{a list of one or more site models,
as created by \link[beautier]{create_site_model}}

\item{clock_models}{a list of one or more clock models,
as created by \link[beautier]{create_clock_model}}

\item{tree_priors}{a list of one or more tree priors,
as created by \link[beautier]{create_tree_prior}}

\item{epsilon}{measure of relative accuracy when estimating a model's
evidence (also known as marginal likelihood).
Smaller values result in more precise estimations, that take
longer to compute}

\item{verbose}{if TRUE, show more output}

\item{marg_lik_filename}{name of the file the marginal
likelihoods (also known as 'evidences') are saved to}
}
\description{
Determine the strategy to select between multiple models to
be used in inference:
}
\details{
\itemize{
  \item \link{create_gen_model_select_param}: do not select between
    inference models: use the generative model
  \item \link{create_best_model_select_param}: select the
    inference model with the most evidence (aka marginal likelihood),
    and use that model for inference
}
The simplest combination uses \code{model_selections = "generative"},
in which the generative model of the alignmnent is (rightfully)
assumed to underly its creation. The tree prior underlying the phylogeny
is assumed to be a Birth-Death prior.
For a combination such as \code{model_selections = "most_evidence"},
}
\examples{
  # Pick the generative model
  alignment_params <- create_alignment_params(
    root_sequence = "acgt", mutation_rate = 0.01
  )
  model_select_param <- create_gen_model_select_param(alignment_params)
  # In such a case, the site model and clock models of the alignment is
  # stored in the model selection parameters
  testthat::expect_equal(
    alignment_params$site_model,
    model_select_param$site_models[[1]]
  )
  testthat::expect_equal(
    alignment_params$clock_model,
    model_select_param$clock_models[[1]]
  )
  # By default, a birth-death model is assumed to underly the phylogeny
  testthat::expect_equal(
    model_select_param$tree_priors[[1]],
    beautier::create_bd_tree_prior()
  )

  # Pick the model with most evidence to be used in inference
  model_select_param <- create_best_model_select_param()
  # In such a case, multiple site models, clock models
  #  and tree priors are tested
  testthat::expect_true(length(model_select_param$site_models) > 1)
  testthat::expect_true(length(model_select_param$clock_models) > 1)
  testthat::expect_true(length(model_select_param$tree_priors) > 1)
}
\author{
Richel J.C. Bilderbeek
}
