% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_model_selection_params.R
\name{create_model_select_params}
\alias{create_model_select_params}
\title{Create model selection parameters.
The simplest combination uses \code{model_selections = "generative"},
in which the generative model of the alignmnent is (rightfully)
assumed to underly its creation. The tree prior underlying the phylogeny
is assumed to be a Birth-Death prior.
For a combination such as \code{model_selections = "most_evidence"},}
\usage{
create_model_select_params(model_selections = "generative",
  site_models = beautier::create_site_models(),
  clock_models = beautier::create_clock_models(),
  tree_priors = beautier::create_tree_priors())
}
\arguments{
\item{model_selections}{one or more ways to select the models used in
inference, for example, \code{generative} picks the generative
model, where \code{most_evidence} picks the model with most
evidence. See \link{get_model_selections} for a list of}
}
\description{
Create model selection parameters.
The simplest combination uses \code{model_selections = "generative"},
in which the generative model of the alignmnent is (rightfully)
assumed to underly its creation. The tree prior underlying the phylogeny
is assumed to be a Birth-Death prior.
For a combination such as \code{model_selections = "most_evidence"},
}
\author{
Richel J.C. Bilderbeek
}
