% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_model_selection_params.R
\name{create_most_evidence_model_select_params}
\alias{create_most_evidence_model_select_params}
\title{Create model selection parameters
for the model with the model evidence (also: marginal likelihood)
from a set of any combination of site models, clock models and
tree priors.}
\usage{

  create_most_evidence_model_select_params(site_models = beautier::create_site_models(),
  clock_models = beautier::create_clock_models(),
  tree_priors = beautier::create_tree_priors())
}
\description{
Create model selection parameters
for the model with the model evidence (also: marginal likelihood)
from a set of any combination of site models, clock models and
tree priors.
}
\author{
Richel J.C. Bilderbeek
}
