% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_nltts.R
\name{create_nltts}
\alias{create_nltts}
\title{Create the NLTT statistics distribution files from a posterior file.
Assumes for a posterior file named '1x.trees'
  that there is a 'true' tree file called '1x.tree'}
\usage{
create_nltts(phylogeny, posterior)
}
\arguments{
\item{phylogeny}{a phylogeny of class \link[ape]{phylo}}
}
\value{
nltt statistics
}
\description{
Create the NLTT statistics distribution files from a posterior file.
Assumes for a posterior file named '1x.trees'
  that there is a 'true' tree file called '1x.tree'
}
\author{
Richel J.C. Bilderbeek
}
