% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_inference_model.R
\name{create_old_skool_inference_model}
\alias{create_old_skool_inference_model}
\title{Create an inference model.}
\usage{
create_old_skool_inference_model(site_model, clock_model, tree_prior,
  beast2_input_filename = tempfile(fileext = ".xml"),
  beast2_output_log_filename = tempfile(fileext = ".log"),
  beast2_output_trees_filename = tempfile(fileext = ".trees"),
  beast2_output_state_filename = tempfile(fileext = ".xml.state"))
}
\arguments{
\item{site_model}{a nucleotide substitution model,
as created by \link[beautier]{create_site_model}}

\item{clock_model}{a clock model,
as created by \link[beautier]{create_clock_model}}

\item{tree_prior}{a tree prior,
as created by \link[beautier]{create_tree_prior}}

\item{beast2_input_filename}{path of the BEAST2 configuration file.
By default, this file is put in a temporary folder with a random filename,
as the user needs not read it: it is used as input of BEAST2.
Specifying a \code{beast2_input_filename} allows
to store that file in a more permanently stored location.}

\item{beast2_output_log_filename}{name of the log file created by BEAST2,
containing the parameter estimates in time.
By default, this file is put a temporary folder with a random filename,
as the user needs not read it.
Specifying a beast2_output_log_filename allows to store that file
in a more permanently stored location.}

\item{beast2_output_trees_filename}{name of a trees files
created by BEAST2.
By default, this file is put a temporary folder with a random filename,
as the user needs not read it: its content is parsed and
compared to a true phylogeny to obtain the inference errors.
Specifying \code{beast2_output_trees_filename} allows to store
this file in a more permanently stored location.}

\item{beast2_output_state_filename}{name of the final state file
created by BEAST2, containing the operator acceptances.
By default, this file is put a temporary folder with a random filename,
as the user needs not read it.
Specifying a beast2_output_state_filename allows to store
that file in a more permanently stored location.}
}
\value{
an inference model
}
\description{
An inference model is a combination of a site model, clock
model and tree prior. An inference model is also the part
of the BEAST2 setups that differ; the part that is shared
is created by \link{create_inference_params})
}
\author{
Richel J.C. Bilderbeek
}
