% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pff_twinning_params.R
\name{create_pff_twinning_params}
\alias{create_pff_twinning_params}
\title{Create a PFF twinning parameter set}
\usage{
create_pff_twinning_params(
  rng_seed_twin_tree = pirouette::create_twinning_params()$rng_seed_twin_tree,
  sim_twin_tree_fun = pirouette::create_twinning_params()$sim_twin_tree_fun,
  rng_seed_twin_alignment = pirouette::create_twinning_params()$rng_seed_twin_alignment,
  sim_twal_fun = pirouette::create_twinning_params()$sim_twal_fun,
  twin_tree_filename = peregrine::get_pff_tempfile(pattern = "tree_twin_", fileext =
    ".newick"),
  twin_alignment_filename = peregrine::get_pff_tempfile(pattern = "alignment_twin_",
    fileext = ".fasta"),
  twin_evidence_filename = NA
)
}
\arguments{
\item{rng_seed_twin_tree}{the random number generator seed as used in the
simulation of a twin tree}

\item{sim_twin_tree_fun}{function to simulate a twin tree with.
This function must have one argument called \code{phylogeny}
of type \link[ape]{phylo} and have a return type of type \link[ape]{phylo}
as well.

Some standard functions:\cr
\itemize{
  \item Use \link{create_sim_yule_twin_tree_fun} to use a
    Yule (aka Pure Birth) process
  \item Use \link{create_copy_twtr_from_true_fun} to for a function
    that copies the true tree
  \item Use \link{get_sim_bd_twin_tree_fun} to use a
    Birth-Death process
}}

\item{rng_seed_twin_alignment}{the random number generator seed
as used in the simulation of a twin alignment}

\item{sim_twal_fun}{function to simulate a
twin alignment with.
This function must have two arguments called \code{twin_phylogeny} (which
will hold the twin phylogeny) and \code{true_alignment} (which will
hold the alignment simulated from the true phylogeny). The
return type must be \link[ape]{DNAbin}.

Use \link{check_sim_twal_fun} to verify if the function
has the right signature and output.

Some standard functions:\cr
\itemize{
  \item Use \link{get_copy_tral_fun}
    to get a function
    (\link{copy_true_alignment})
    that copies a true to alignment to create a twin alignment
  \item Use \link{get_sim_twal_with_std_nsm_fun}
    to get a function
    (\link{sim_twal_with_std_nsm})
    that simulates a twin alignment using a standard site model
  \item Use \link{get_sim_twal_same_n_muts_fun}
    to get a function
    (\link{sim_twal_with_same_n_mutation})
    that simulates -using a standard model- a twin alignment with as much
    mutations compared to the root sequence as the true alignment has
}}

\item{twin_tree_filename}{name of the (\code{.newick}) file the twin
tree will be saved to}

\item{twin_alignment_filename}{name of the FASTA file the twin
alignment will be saved to}

\item{twin_evidence_filename}{filename to store the estimated
evidences (aka marginal likelihoods) of the twin tree}
}
\description{
Create a PFF twinning parameter set
}
