% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pir_out_from_folder.R
\name{create_pir_out_from_folder}
\alias{create_pir_out_from_folder}
\title{Create a \code{pir_out} from a folder}
\usage{
create_pir_out_from_folder(folder_name)
}
\arguments{
\item{folder_name}{name of the main folder}
}
\value{
a \code{pir_out}, as can be checked by \link{check_pir_out}
}
\description{
Create a \code{pir_out} from a folder
}
\author{
Richèl J.C. Bilderbeek
}
