% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pir_outs_from_folders.R
\name{create_pir_outs_from_folders}
\alias{create_pir_outs_from_folders}
\title{Create a \code{pir_out} from a folder}
\usage{
create_pir_outs_from_folders(folder_names)
}
\value{
a \code{pir_out}, as can be checked by \link{check_pir_out}
}
\description{
Create a \code{pir_out} from a folder
}
\author{
Richèl J.C. Bilderbeek
}
