% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pir_params_twin.R
\name{create_pir_params_twin}
\alias{create_pir_params_twin}
\title{Create the twin parameters for \code{pirouette}}
\usage{
create_pir_params_twin(pir_params, pir_out)
}
\arguments{
\item{pir_params}{the parameters of \link[pirouette]{pirouette}.
They are created by \link{create_pir_params}.}

\item{pir_out}{the output of \link{pir_run_tree} applied on the original tree}
}
\value{
a list with all \code{pirouette} twin parameters
}
\description{
Create the twin parameters for \code{pirouette}
}
\author{
Giovanni Laudanno
}
