% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_result.R
\name{create_result}
\alias{create_result}
\title{Create a result}
\usage{
create_result(log_evidence, weight, errors)
}
\arguments{
\item{log_evidence}{the natural logarithm of the evidence (aka marginal
likelihood). Can be NA if this is not measured}

\item{weight}{the weight of the model, compared to other (candidate)
models. This weight will be between 0.0 (there is no evidence for
this model) to 1.0 (all evidence indicates this is the best model).
A weight of NA denotes that the weight is not measured}

\item{errors}{a numeric vector of (positive) Bayesian inference errors.
Use NA if these are not measured (yet)}
}
\value{
a result
}
\description{
Create a result
}
\examples{
  log_evidence <- -123.456
  weight <- 0.1234
  errors <- c(3, 1, 4, 1, 5, 9, 3)

  result <- create_result(
    log_evidence = log_evidence,
    weight = weight,
    errors = errors
  )

  expect_equal(log_evidence, result$log_evidence)
  expect_equal(weight, result$weight)
  expect_equal(errors, result$errors)

  testthat::expect_silent(
    check_result(create_test_result())
  )
}
\seealso{
Use \link{check_result} to check a result
}
\author{
Richel J.C. Bilderbeek
}
