% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_run_experiment.R
\name{create_run_experiment}
\alias{create_run_experiment}
\title{Create a valid \link{pirouette} run experiment.}
\usage{
create_run_experiment(experiment)
}
\arguments{
\item{experiment}{a \link{pirouette} experiment,
as can be created by \link{create_experiment}}
}
\value{
a \link{pirouette} run experiment
}
\description{
The arguments are checked by \link{check_run_experiment}.
}
\examples{
  library(testthat)

  run_experiment <- create_test_run_experiment()

  expect_silent(check_run_experiment(run_experiment))
}
\author{
Richel J.C. Bilderbeek
}
