% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sim_bd_twin_tree_function.R
\name{create_sim_bd_twin_tree_function}
\alias{create_sim_bd_twin_tree_function}
\title{Create a function that can simulate the twin tree from the true tree,
using the BD speciation model}
\usage{
create_sim_bd_twin_tree_function(seed = 0, method = "random_tree",
  n_replicates = 10000)
}
\arguments{
\item{seed}{a random number generator seed}

\item{method}{determines how to create the twin tree
\itemize{
    \item 'random_tree' just produces a random tree;
    \item 'max_clade_cred' simulates \code{n_replicates} trees and
      uses \link[phangorn]{maxCladeCred} to create a consensus tree;
    \item 'max_likelihood' simulates \code{n_replicates} trees
     and selects the most likely;
  }}

\item{n_replicates}{number of replicas to evaluate in order to create the
twin tree}
}
\value{
a function
}
\description{
Create a function that can simulate the twin tree from the true tree,
using the BD speciation model
}
\examples{
library(testthat)

f <- create_sim_bd_twin_tree_function()
phylo_in <- ape::read.tree(text = "((A:1, B:1):1, C:2);")
phylo_out <- f(phylo_in)
expect_true(
  !all(ape::branching.times(phylo_in) == ape::branching.times(phylo_out))
)
}
\seealso{
Use \link{create_sim_yule_twin_tree_function} to get a function to
produce a Yule tree.
Use \link{create_copy_twin_tree_from_true_function} to get a
function to simply copy the tree
}
\author{
Richèl J.C. Bilderbeek
}
