% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sim_twin_tree_params.R
\name{create_sim_twin_tree_params}
\alias{create_sim_twin_tree_params}
\title{Create the parameters how to simulate a twin tree}
\usage{
create_sim_twin_tree_params()
}
\description{
Create the parameters how to simulate a twin tree
}
