% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sim_yule_twin_tree_function.R
\name{create_sim_yule_twin_tree_function}
\alias{create_sim_yule_twin_tree_function}
\title{Create a function that can simulate the twin tree from the true tree,
using the Yule speciation model}
\usage{
create_sim_yule_twin_tree_function()
}
\value{
a function
}
\description{
Create a function that can simulate the twin tree from the true tree,
using the Yule speciation model
}
\examples{
f <- create_sim_yule_twin_tree_function()
phylo_in <- ape::read.tree(text = "((A:1, B:1):1, C:2);")
phylo_out <- f(phylo_in)
expect_true(
  !all(ape::branching.times(phylo_in) == ape::branching.times(phylo_out))
)
}
