% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_site_model_from_name.R
\name{create_site_model_from_name}
\alias{create_site_model_from_name}
\title{Create a site model from name}
\usage{
create_site_model_from_name(site_model_name)
}
\arguments{
\item{site_model_name}{name of a site model}
}
\description{
Create a site model from name
}
\author{
Richel J.C. Bilderbeek
}
