% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_standard_mutation_rate.R
\name{create_standard_mutation_rate}
\alias{create_standard_mutation_rate}
\title{Create the parameters for the mutation rate}
\usage{
create_standard_mutation_rate(phylogeny)
}
\arguments{
\item{phylogeny}{a phylogeny of class \link[ape]{phylo}}
}
\value{
the mutation rate
}
\description{
Create the parameters for the mutation rate
}
\author{
Giovanni Laudanno, Richel J.C. Bilderbeek
}
