% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_stunning_tree.R
\name{create_stunning_tree}
\alias{create_stunning_tree}
\title{Create figure bd for pirouette article}
\usage{
create_stunning_tree(t_0 = 10, n_taxa = 5, n_0 = 2)
}
\arguments{
\item{t_0}{starting time of a tree}

\item{n_taxa}{number of tree tips}

\item{n_0}{number of starting species}
}
\value{
a list containing a tree in phylo format, the same tree in
newick format and the plot of the tree
}
\description{
Create figure bd for pirouette article
}
\author{
Giovanni Laudanno
}
