% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_alignment_params.R
\name{create_test_alignment_params}
\alias{create_test_alignment_params}
\title{Create testing parameters for the alignment simulation.}
\usage{
create_test_alignment_params()
}
\value{
a list of alignment parameters
}
\description{
Create testing parameters for the alignment simulation.
}
\author{
Richel J.C. Bilderbeek
}
