% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_experiment.R
\name{create_test_gen_experiment}
\alias{create_test_gen_experiment}
\title{Create a valid testing \link{pirouette} generative experiment.}
\usage{
create_test_gen_experiment()
}
\value{
a \link{pirouette} experiment.
}
\description{
Create a valid testing \link{pirouette} generative experiment.
}
\author{
Richel J.C. Bilderbeek
}
