% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_test_phylogeny.R
\name{create_test_phylogeny}
\alias{create_test_phylogeny}
\title{Create a testing phylogeny with 3 taxa and a crown age of 3}
\usage{
create_test_phylogeny()
}
\value{
a \code{phylo} from the \code{ape} package
}
\description{
Create a testing phylogeny with 3 taxa and a crown age of 3
}
\author{
Richèl J.C. Bilderbeek
}
