% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_test_pir_run_output.R
\name{create_test_pir_run_output2}
\alias{create_test_pir_run_output2}
\title{Create the same output of \link{pir_run}
to be used for testing, but with more data}
\usage{
create_test_pir_run_output2(add_twin = FALSE, add_best = FALSE, n_errors = 200)
}
\arguments{
\item{add_twin}{add rows for twin tree}

\item{add_best}{add rows for best inference model}

\item{n_errors}{number of errors in the pir_out}
}
\value{
a data frame with errors, with as many rows as model selection
parameter sets. The output can be checked using \link{check_pir_out}.
}
\description{
Create the same output of \link{pir_run}
to be used for testing, but with more data
}
\examples{
  pir_plot(
    create_test_pir_run_output2(
      add_twin = TRUE,
      add_best = TRUE,
      n_errors = 1e2
    )
  )
}
\author{
Giovanni Laudanno, Richèl J.C. Bilderbeek
}
