% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_run_experiment.R
\name{create_test_run_experiment}
\alias{create_test_run_experiment}
\title{Create a valid testing \link{pirouette} run experiment}
\usage{
create_test_run_experiment()
}
\value{
a \link{pirouette} run experiment.
}
\description{
Create a valid testing \link{pirouette} run experiment
}
\examples{
  library(testthat)

  expect_silent(check_run_experiment(create_test_run_experiment()))
}
\author{
Richel J.C. Bilderbeek
}
