% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tree_and_model_errors_from_folder.R
\name{create_tree_and_model_errors_from_folder}
\alias{create_tree_and_model_errors_from_folder}
\title{Internal function to create a \code{tree_and_model_errors}
from the files in a folder}
\usage{
create_tree_and_model_errors_from_folder(folder_name)
}
\arguments{
\item{folder_name}{name of the main folder}
}
\value{
a \code{tree_and_model_errors},
  as can be checked by \link{check_tree_and_model_errors}
}
\description{
Internal function to create a \code{tree_and_model_errors}
from the files in a folder
}
\author{
Richèl J.C. Bilderbeek
}
