% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tree_and_model_errors_from_folders.R
\name{create_tree_and_model_errors_from_folders}
\alias{create_tree_and_model_errors_from_folders}
\title{Internal function to create a \code{tree_and_model_errors}
from the files in one or more folders}
\usage{
create_tree_and_model_errors_from_folders(folder_names)
}
\arguments{
\item{folder_names}{one or more folder names}
}
\value{
a \code{tree_and_model_errors},
  as can be checked by \link{check_tree_and_model_errors}
}
\description{
Internal function to create a \code{tree_and_model_errors}
from the files in one or more folders
}
\author{
Richèl J.C. Bilderbeek
}
