% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tree_prior_from_name.R
\name{create_tree_prior_from_name}
\alias{create_tree_prior_from_name}
\title{Create a tree prior from name}
\usage{
create_tree_prior_from_name(tree_prior_name)
}
\description{
Create a tree prior from name
}
\author{
Richel J.C. Bilderbeek
}
