% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_twin_alignment.R
\name{create_twin_alignment}
\alias{create_twin_alignment}
\title{Creates a twin alignment.}
\usage{
create_twin_alignment(twin_phylogeny, true_alignment, alignment_params,
  twinning_params)
}
\arguments{
\item{twin_phylogeny}{a phylogeny of class \link[ape]{phylo}}

\item{true_alignment}{a DNA alignment, of class \link[ape]{DNAbin}}

\item{alignment_params}{parameters to simulate an alignment,
as can be created by \link{create_alignment_params}}

\item{twinning_params}{can be \code{NA} if no twinning is desired,
or can be the twinning parameters,
as can be created by \link{create_twinning_params}}
}
\value{
an alignment of class DNAbin that has as much
  mutations accumulated from crown to the tips as the
  original, 'true' alignment
}
\description{
A twin alignment is an alignment that has as much
mutations accumulated from crown to the tips
}
\author{
Richèl J.C. Bilderbeek, Giovanni Laudanno
}
