% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_alignment.R
\name{create_twin_alignment_with_standard_site_model_raw}
\alias{create_twin_alignment_with_standard_site_model_raw}
\title{Create an alignment with a standard site model using a raw interface}
\usage{
create_twin_alignment_with_standard_site_model_raw(twin_phylogeny,
  true_alignment, root_sequence, mutation_rate, site_model)
}
\arguments{
\item{twin_phylogeny}{a phylogeny of class \link[ape]{phylo}}

\item{true_alignment}{a DNA alignment, of class \link[ape]{DNAbin}}

\item{root_sequence}{the DNA sequence at the root of the phylogeny.
By default, this will consist out of an equal amount of each letter
Use \link{check_root_sequence} to check if a root sequence is valid.}

\item{mutation_rate}{the mutation rate per base pair per time unit.
Use \link{check_mutation_rate} to check if a mutation rate is valid.}

\item{site_model}{a nucleotide substitution model,
which can be:
\itemize{
  \item{
    A standard nucloetide substitution model,
    as created by \link[beautier]{create_site_model}
  }
  \item{
    \code{linked_node_sub}: a linked node-substitution model
  }
  \item{
    \code{unlinked_node_sub}: an unlinked node-substitution model
  }
}}
}
\value{
an alignment of type \code{DNAbin}
}
\description{
Create an alignment with a standard site model using a raw interface
}
