% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_params_doc.R
\name{default_params_doc}
\alias{default_params_doc}
\title{This function does nothing. It is intended to inherit is parameters'
documentation.}
\usage{
default_params_doc(alignment, bd_mutation_rate, bd_tree, bd_tree_filename,
  brts, chain_length, clock_model, crown_age, fasta_filename, filename,
  folder_name, init_speciation_rate, init_extinction_rate, lambda,
  mbd_l_matrix, mbd_mutation_rate, mbd_tree, mu, mutation_rate, nu,
  nu_events, parameters, parameter_filename, parameters_filename, phylo,
  posterior_trees, precision, project_folder_name, q, sample_interval,
  seed, sequence_length, sim_pars, sim_phylo, site_model, sub_chain_length,
  tree, tree_filename, trees_filename, verbose)
}
\arguments{
\item{alignment}{a DNA alignment}

\item{bd_mutation_rate}{the mutation rate when creating an alignment
from a BD tree}

\item{bd_tree}{a phylogent of class \code{phylo}, created by a Birth Death
process}

\item{bd_tree_filename}{name of the file that stores a BD twin tree}

\item{brts}{set of branching times}

\item{chain_length}{something}

\item{clock_model}{Name of the clock model that has
to be used for the inference. Valid names are 'strict' and 'rln'.}

\item{crown_age}{The crown age of the tree.}

\item{fasta_filename}{name of a FASTA file}

\item{filename}{the file's name, without the path}

\item{folder_name}{name of the main folder}

\item{init_speciation_rate}{a speciation rate}

\item{init_extinction_rate}{an extinction rate}

\item{lambda}{per-lineage speciation rate}

\item{mbd_l_matrix}{the L matrix of an MBD tree}

\item{mbd_mutation_rate}{the mutation rate when creating an alignment
from a MBD tree}

\item{mbd_tree}{an MBD tree}

\item{mu}{per-species extinction rate}

\item{mutation_rate}{something}

\item{nu}{the rate at which a multiple-birth specation is triggered}

\item{nu_events}{the number of nu-triggered events that have to be
present in the simulated tree}

\item{parameters}{the razzo parameters}

\item{parameter_filename}{full path to a 'parameters.csv' file}

\item{parameters_filename}{full path to a 'parameters.csv' file}

\item{phylo}{a phylogeny}

\item{posterior_trees}{phylogenetic trees in a BEAST2 posterior,
of class \code{multiphylo}}

\item{precision}{define the precision of the approximation.}

\item{project_folder_name}{project folder name,
will be the full path to \code{razzo_project}}

\item{q}{something}

\item{sample_interval}{the interval at which the MCMC algorithm
makes a measurement}

\item{seed}{a random number generator seed}

\item{sequence_length}{the length of each DNA sequence in an alignment}

\item{sim_pars}{something}

\item{sim_phylo}{something}

\item{site_model}{Name of the site model that
has to be used for the inference. Valid names are 'jc69' and 'gtr'.}

\item{sub_chain_length}{length of the sub-chain used by the Nested Sampling
algorithm to estimate the marginal likelihood}

\item{tree}{an ultrametric phylogenetic tree of class \code{phylo}}

\item{tree_filename}{name of the phylogeny file}

\item{trees_filename}{name of the BEAST2 posterior phylogenies file}

\item{verbose}{give more output}
}
\description{
This function does nothing. It is intended to inherit is parameters'
documentation.
}
\note{
This is an internal function, so it should be marked with
  \code{@noRd}. This is not done, as this will disallow all
  functions to find the documentation parameters
}
\author{
Documentation by Giovanni Laudanno,
  use of this function by Richel J.C. Bilderbeek
}
