% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pirouette.R
\name{enpointe}
\alias{enpointe}
\title{Random rotation tree}
\usage{
enpointe(x, y, weights = NULL, maxrows = 500, ...)
}
\arguments{
\item{x}{a sparse matrix of x variables}

\item{y}{the target variable for classification or regression}

\item{maxrows}{the maximum number of rows to use for the tree fit}

\item{...}{passed to rpart}
}
\value{
an object of class pirouette
}
\description{
This is the workhorse function of the pirouette package.  It
splats the dateset and fits an rpart model to it.
}

