% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_evidences.R
\name{est_evidences}
\alias{est_evidences}
\title{Estimate the evidences}
\usage{
est_evidences(fasta_filename, experiments, evidence_epsilon = 1e-12,
  evidence_filename = tempfile(pattern = "evidence_", fileext = ".csv"),
  verbose = FALSE)
}
\arguments{
\item{fasta_filename}{name of a FASTA file}

\item{experiments}{a list of one or more \link{pirouette} experiments,
as can be created by \link{create_experiment}. If more than one experiment
is provided and a "generative" experiment is part of them, the "generative"
one has to be the first in the list.}

\item{evidence_epsilon}{relative error in estimating the
evidence (aka marginal likelihood).}

\item{evidence_filename}{filename to store the estimated
evidences (aka marginal likelihoods)}

\item{verbose}{if TRUE, show more output}
}
\value{
a data frame with evidences. Returns NULL if there
  are no experiments that have their evidence measured.
}
\description{
Estimate the evidences
}
\examples{
if (is_on_travis() &&
  is_beast2_installed() && is_beast2_ns_pkg_installed()
) {
  fasta_filename <- system.file(
    "extdata", "alignment.fas", package = "pirouette"
  )

  # Create a single one candidate experiment
  experiments <- list(create_test_cand_experiment())

  # Be sloppy amd fast in estimating the evidence
  evidence_epsilon <- 100.0

  evidences <- est_evidences(
    fasta_filename = fasta_filename,
    experiments = experiments,
    evidence_epsilon = evidence_epsilon
  )

  library(testthat)
  expect_true("site_model_name" \%in\% names(evidences))
  expect_true("clock_model_name" \%in\% names(evidences))
  expect_true("tree_prior_name" \%in\% names(evidences))
  expect_true("marg_log_lik" \%in\% names(evidences))
  expect_true("marg_log_lik_sd" \%in\% names(evidences))
  expect_true("weight" \%in\% names(evidences))

  # As the only experiment, its weight is 1.0
  expect_equal(1.0, evidences$weight[1])
}
}
\author{
Richèl J.C. Bilderbeek
}
