% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_evidences.R
\name{est_evidences_new_skool}
\alias{est_evidences_new_skool}
\title{Estimate the evidences}
\usage{
est_evidences_new_skool(fasta_filename,
  experiments = list(create_experiment()), evidence_epsilon = 1e-12,
  evidence_filename = tempfile(fileext = ".csv"))
}
\arguments{
\item{fasta_filename}{name of a FASTA file}

\item{experiments}{a list of one or more \link{pirouette} experiments,
as can be created by \link{create_experiment}}
}
\value{
a data frame with evidences
}
\description{
Estimate the evidences
}
\author{
Richel J.C. Bilderbeek
}
