% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_evidences.R
\name{est_evidences_old_skool}
\alias{est_evidences_old_skool}
\title{Estimate the evidences old skool}
\usage{
est_evidences_old_skool(fasta_filename, model_select_params)
}
\arguments{
\item{fasta_filename}{name of a FASTA file}

\item{model_select_params}{a list of one or more parameter sets
to select an inference model
(a combination of site model, clock model and tree prior),
as can be created by \link{create_model_select_param}}
}
\value{
a data frame with evidences
}
\description{
Estimate the evidences old skool
}
\author{
Richel J.C. Bilderbeek
}
