% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_alignment_sequences.R
\name{get_alignment_sequences}
\alias{get_alignment_sequences}
\title{Get the sequences from an alignment}
\usage{
get_alignment_sequences(alignment, verbose = FALSE)
}
\arguments{
\item{alignment}{a DNA alignment, of class \link[ape]{DNAbin}}

\item{verbose}{if TRUE, show more output}
}
\value{
a numeric vector with the sequences
}
\description{
Get the sequences from an alignment
}
