% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_beast2_options_filenames.R
\name{get_beast2_options_filenames}
\alias{get_beast2_options_filenames}
\title{Extract the filenames from a \code{beast2_options}}
\usage{
get_beast2_options_filenames(beast2_options)
}
\arguments{
\item{beast2_options}{BEAST2 options,
as can be created by \link[beastier]{create_beast2_options}}
}
\description{
Extract the filenames from a \code{beast2_options}
}
\author{
Richèl J.C. Bilderbeek
}
