% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_copy_tral_fun.R
\name{get_copy_tral_fun}
\alias{get_copy_tral_fun}
\title{Get a function to 'simulate' a twin alignment by simply
copying the true alignment.}
\usage{
get_copy_tral_fun()
}
\value{
the function
\link{copy_true_alignment}
}
\description{
Get a function to 'simulate' a twin alignment by simply
copying the true alignment.
}
\examples{

f <- get_copy_tral_fun()
# This adapter function must be a sim_twin_alignment function
expect_silent(
  check_sim_twal_fun(f)
)
}
\seealso{
See \link{check_sim_twal_fun} to the the other
functions to simulate a twin alignment.
Use \link{sim_twin_alignment} to use this function to
create a twin alignment.
}
