% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_copy_true_alignment_function.R
\name{get_copy_true_alignment_function}
\alias{get_copy_true_alignment_function}
\title{Get a function to 'simulate' a twin alignment by simply
copying the true alignment.}
\usage{
get_copy_true_alignment_function()
}
\value{
the function
\link{copy_true_alignment}
}
\description{
Get a function to 'simulate' a twin alignment by simply
copying the true alignment.
}
\examples{
library(testthat)

f <- get_copy_true_alignment_function()
# This adapter function must be a sim_twin_alignment function
expect_silent(
  check_sim_twin_alignment_function(f)
)
}
\seealso{
See \link{check_sim_twin_alignment_function} to the the other
functions to simulate a twin alignment.
Use \link{sim_twin_alignment} to use this function to
create a twin alignment.
}
