% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_default_sim_true_alignment_function.R
\name{get_default_sim_true_alignment_function}
\alias{get_default_sim_true_alignment_function}
\title{Get a default function to simulate a 'true' alignment from a 'true' phylogeny}
\usage{
get_default_sim_true_alignment_function()
}
\description{
Get a default function to simulate a 'true' alignment from a 'true' phylogeny
}
