% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_default_sim_twin_alignment_function.R
\name{get_default_sim_twin_alignment_function}
\alias{get_default_sim_twin_alignment_function}
\title{Get a default function to simulate a twin alignment from a
twin phylogeny and a true alignment}
\usage{
get_default_sim_twin_alignment_function()
}
\description{
Get a default function to simulate a twin alignment from a
twin phylogeny and a true alignment
}
