% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_experiment_filenames.R
\name{get_experiment_filenames}
\alias{get_experiment_filenames}
\title{Extract the filenames from an experiment}
\usage{
get_experiment_filenames(experiment)
}
\arguments{
\item{experiment}{a \link{pirouette} experiment,
as can be created by \link{create_experiment}}
}
\description{
Extract the filenames from an experiment
}
\examples{
if (beautier::is_on_ci()) {
  get_experiment_filenames(
    experiment = create_test_experiment()
  )
}
}
\author{
Richèl J.C. Bilderbeek
}
