% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_experiments_filenames.R
\name{get_experiments_filenames}
\alias{get_experiments_filenames}
\title{Extract the filesnames in the experiments}
\usage{
get_experiments_filenames(experiments)
}
\arguments{
\item{experiments}{a list of one or more \link{pirouette} experiments,
as can be created by \link{create_experiment}. If more than one experiment
is provided and a "generative" experiment is part of them, the "generative"
one has to be the first in the list.}
}
\description{
Extract the filesnames in the experiments
}
\examples{
get_experiments_filenames(
  experiments = list(create_test_experiment())
)
}
\author{
Richèl J.C. Bilderbeek
}
