% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_model_type_names}
\alias{get_model_type_names}
\title{Get the names of the model types}
\usage{
get_model_type_names()
}
\value{
the model types
}
\description{
Get the names of the model types
}
\examples{

"candidate" \%in\% get_model_type_names())
"generative" \%in\% get_model_type_names())
"Jensen Ackles" \%in\% get_model_type_names())
}
\author{
Giovanni Laudanno, Richèl J.C. Bilderbeek
}
