% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_model_types}
\alias{get_model_types}
\title{Get the names of the model types}
\usage{
get_model_types()
}
\value{
the model types
}
\description{
Get the names of the model types
}
\examples{
library(testthat)

expect_true("candidate" \%in\% get_model_types())
expect_true("generative" \%in\% get_model_types())
expect_false("Jensen Ackles" \%in\% get_model_types())
}
\author{
Giovanni Laudanno, Richel J.C. Bilderbeek
}
