% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nltt_error_function.R
\name{get_nltt_error_function}
\alias{get_nltt_error_function}
\title{Get an error function that uses the nLTT statistic}
\usage{
get_nltt_error_function()
}
\description{
Get an error function that uses the nLTT statistic
}
\author{
Richel J.C. Bilderbeek
}
