% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pir_params_filenames.R
\name{get_pir_params_filenames}
\alias{get_pir_params_filenames}
\title{Extract the filenames from a \code{pir_params}}
\usage{
get_pir_params_filenames(pir_params)
}
\arguments{
\item{pir_params}{the parameters of \link[pirouette]{pirouette}.
They are created by \link{create_pir_params}.}
}
\value{
a character vector of filenames
}
\description{
Extract the filenames from a \code{pir_params}
}
\examples{
if (beautier::is_on_ci()) {
  pir_params <- create_pir_params(
    alignment_params = create_test_alignment_params(),
    experiments = list(create_test_experiment())
  )
  get_pir_params_filenames(
    pir_params = pir_params
  )
}
}
\author{
Richèl J.C. Bilderbeek
}
