% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pir_params_filenames.R
\name{get_pir_params_filenames_oldskool}
\alias{get_pir_params_filenames_oldskool}
\title{Extract the filenames from a \code{pir_params}
using the \code{oldskool} method.}
\usage{
get_pir_params_filenames_oldskool(pir_params)
}
\arguments{
\item{pir_params}{the parameters of \link[pirouette]{pirouette}.
They are created by \link{create_pir_params}.}
}
\description{
Extract the filenames from a \code{pir_params}
using the \code{oldskool} method.
}
\author{
Richèl J.C. Bilderbeek
}
