% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pir_plot_fill_colors.R
\name{get_pir_plot_fill_colors}
\alias{get_pir_plot_fill_colors}
\title{Internal function to get the fill colors for \link{pir_plot}}
\usage{
get_pir_plot_fill_colors()
}
\value{
a \code{ggplot2} plot
}
\description{
Internal function to get the fill colors for \link{pir_plot}
}
\author{
Giovanni Laudanno, Richèl J.C. Bilderbeek
}
