% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pir_plot_theme.R
\name{get_pir_plot_theme}
\alias{get_pir_plot_theme}
\title{Get the \link{pir_plot} theme}
\usage{
get_pir_plot_theme()
}
\value{
the \link{pir_plot} theme
}
\description{
Get the \link{pir_plot} theme
}
\author{
Giovanni Laudanno, Richèl J.C. Bilderbeek
}
