% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pir_plot_tree_and_model_labels.R
\name{get_pir_plot_tree_and_model_labels}
\alias{get_pir_plot_tree_and_model_labels}
\title{Internal function to obtain the \link{pir_plot} legend labels}
\usage{
get_pir_plot_tree_and_model_labels(pir_out)
}
\arguments{
\item{pir_out}{the output of \link{pir_run}}
}
\value{
the \link{pir_plot} legend labels
}
\description{
Internal function to obtain the \link{pir_plot} legend labels
}
\author{
Giovanni Laudanno, Richèl J.C. Bilderbeek
}
