% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_remove_dir_fun.R
\name{get_remove_dir_fun}
\alias{get_remove_dir_fun}
\title{Get a function that, from a filename, returns the part
without the directory.}
\usage{
get_remove_dir_fun()
}
\description{
Or: get a function that returns the local version of a filename.
Also, the function will return \link{NA} if the filename is \link{NA}
}
\seealso{
see \link{check_rename_fun} for an overview of file renaming functions
}
\author{
Richèl J.C. Bilderbeek
}
