% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_replace_dir_fun.R
\name{get_replace_dir_fun}
\alias{get_replace_dir_fun}
\title{Get a function to replace the directory of a filename}
\usage{
get_replace_dir_fun(new_dir_name = "")
}
\arguments{
\item{new_dir_name}{the new directory name}
}
\description{
Get a function to replace the directory of a filename
}
