% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/get_sim_true_alignment_with_standard_site_model_function.R
\name{get_sim_true_alignment_with_standard_site_model_function}
\alias{get_sim_true_alignment_with_standard_site_model_function}
\title{Get a function to simulate the true alignment with,
that uses a standard site model.}
\usage{
get_sim_true_alignment_with_standard_site_model_function(root_sequence,
  mutation_rate, site_model = beautier::create_jc69_site_model())
}
\arguments{
\item{root_sequence}{the DNA sequence at the root of the phylogeny.
By default, this will consist out of an equal amount of each letter
Use \link{check_root_sequence} to check if a root sequence is valid.}

\item{mutation_rate}{the mutation rate per base pair per time unit.
Use \link{check_mutation_rate} to check if a mutation rate is valid.}

\item{site_model}{a nucleotide substitution model,
which can be:
\itemize{
  \item{
    A standard nucloetide substitution model,
    as created by \link[beautier]{create_site_model}
  }
  \item{
    \code{linked_node_sub}: a linked node-substitution model
  }
  \item{
    \code{unlinked_node_sub}: an unlinked node-substitution model
  }
}}
}
\description{
Get a function to simulate the true alignment with,
that uses a standard site model.
}
\examples{
library(testthat)

f <- get_sim_true_alignment_with_standard_site_model_function(
  root_sequence = "acgt",
  mutation_rate = 0.1
)
check_sim_true_alignment_function(f)

alignment_params <- create_test_alignment_params(
  sim_true_alignment_function = f
)
true_alignment <- pirouette::sim_true_alignment(
  true_phylogeny = ape::read.tree(text = "((A:1, B:1):1, C:2);"),
  alignment_params = alignment_params
)
}
