% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/get_sim_twin_alignment_with_same_n_mutation_function.R
\name{get_sim_twin_alignment_with_same_n_mutation_function}
\alias{get_sim_twin_alignment_with_same_n_mutation_function}
\title{Get a function to simulate a twin alignment from a
twin phylogeny and a true alignment, of which the twin
alignment}
\usage{
get_sim_twin_alignment_with_same_n_mutation_function()
}
\description{
Get a function to simulate a twin alignment from a
twin phylogeny and a true alignment, of which the twin
alignment
}
