% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_temp_errors_filename.R
\name{get_temp_errors_filename}
\alias{get_temp_errors_filename}
\title{Get the name for a temporary file to store
inference errors.}
\usage{
get_temp_errors_filename()
}
\value{
one string
}
\description{
Get the name for a temporary file to store
inference errors.
}
