% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_temp_evidence_filename.R
\name{get_temp_evidence_filename}
\alias{get_temp_evidence_filename}
\title{Get the name for a temporary file to store the
evidences (aka marginal likelihoods) as
a comma-separated file}
\usage{
get_temp_evidence_filename()
}
\value{
one string
}
\description{
Get the name for a temporary file to store the
evidences (aka marginal likelihoods) as
a comma-separated file
}
