% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_temp_fasta_filename.R
\name{get_temp_fasta_filename}
\alias{get_temp_fasta_filename}
\title{Get the name for a temporary file to store an
alignment in the FASTA format}
\usage{
get_temp_fasta_filename()
}
\value{
one string
}
\description{
Get the name for a temporary file to store an
alignment in the FASTA format
}
