% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_temp_tree_filename.R
\name{get_temp_tree_filename}
\alias{get_temp_tree_filename}
\title{Get the name for a temporary file to store a tree
in Newick format}
\usage{
get_temp_tree_filename()
}
\description{
Get the name for a temporary file to store a tree
in Newick format
}
