% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_test_alignment.R
\name{get_test_alignment}
\alias{get_test_alignment}
\title{Get an alignment for testing.}
\usage{
get_test_alignment(n_taxa = 3, sequence_length = 4)
}
\arguments{
\item{n_taxa}{number of tree tips}

\item{sequence_length}{the length of each DNA sequence in an alignment}
}
\description{
Get an alignment for testing.
}
\examples{
library(testthat)

n_taxa <- 7
sequence_length <- 11
alignment <- get_test_alignment(
  n_taxa = n_taxa,
  sequence_length = sequence_length
)
expect_silent(check_alignment(alignment))
expect_equal(n_taxa, get_alignment_n_taxa(alignment))
expect_equal(
  sequence_length,
  get_alignment_sequence_length(alignment)
)
}
