% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tree_and_model_values.R
\name{get_tree_and_model_values}
\alias{get_tree_and_model_values}
\title{Get the valid \code{tree_and_model} values}
\usage{
get_tree_and_model_values()
}
\value{
the four valid \code{tree_and_model} values
}
\description{
Get the valid \code{tree_and_model} values
}
\author{
Richèl J.C. Bilderbeek
}
