% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_tree_types}
\alias{get_tree_types}
\title{Get the names of the tree types}
\usage{
get_tree_types()
}
\value{
the tree types
}
\description{
Get the names of the tree types
}
\examples{

"true" \%in\% get_tree_types())
"twin" \%in\% get_tree_types())
"Jensen Ackles" \%in\% get_tree_types())
}
\author{
Giovanni Laudanno, Richèl J.C. Bilderbeek
}
