% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_twin_methods}
\alias{get_twin_methods}
\title{Twin methods}
\usage{
get_twin_methods()
}
\value{
the twin methods
}
\description{
Twin methods
}
\examples{

"random_tree" \%in\% get_twin_methods())
"max_clade_cred" \%in\% get_twin_methods())
"max_likelihood" \%in\% get_twin_methods())
"nonsense" \%in\% get_twin_methods())
}
\author{
Giovanni Laudanno, Richèl J.C. Bilderbeek
}
