% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_twin_models}
\alias{get_twin_models}
\title{Twin models}
\usage{
get_twin_models()
}
\value{
the twin models
}
\description{
Twin models
}
\examples{

"yule" \%in\% get_twin_models())
"birth_death" \%in\% get_twin_models())
"copy_true" \%in\% get_twin_models())
"nonsense" \%in\% get_twin_models())
}
\author{
Giovanni Laudanno, Richèl J.C. Bilderbeek
}
