% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_candidate_experiments.R
\name{has_candidate_experiments}
\alias{has_candidate_experiments}
\title{Detect if there is at least one candidate model amongst the
set of experiments.}
\usage{
has_candidate_experiments(pir_params)
}
\arguments{
\item{pir_params}{the parameters of \link[pirouette]{pirouette}.
They are created by \link{create_pir_params}.}
}
\description{
Detect if there is at least one candidate model amongst the
set of experiments.
}
