% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_pir_params.R
\name{init_pir_params}
\alias{init_pir_params}
\title{Initialize the \code{pir_params}.}
\usage{
init_pir_params(pir_params)
}
\arguments{
\item{pir_params}{the parameters of \link[pirouette]{pirouette}.
They are created by \link{create_pir_params}.}
}
\description{
A normal user should never need to call this function.
}
\seealso{
Use \link{check_init_pir_params} to check
  if a \code{pir_params} is initialized
}
