% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_pir_deps.R
\name{install_pir_deps}
\alias{install_pir_deps}
\title{Install the \code{pirouette} dependencies, including BEAST2}
\usage{
install_pir_deps()
}
\description{
Installs newer code of the \code{pirouette} dependencies
that is not on CRAN yet, but still on GitHub only.
This function installs the GitHub's master/stable branch
of these dependencies.
}
\details{
This function also installs BEAST2 if absent and installs the
\code{NS} BEAST2 package.
}
\author{
Richèl J.C. Bilderbeek
}
