% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_best_candidate.R
\name{is_best_candidate}
\alias{is_best_candidate}
\title{Is the experiment the one with the most evidence?}
\usage{
is_best_candidate(experiment, marg_liks)
}
\arguments{
\item{experiment}{a \link{pirouette} experiment,
as can be created by \link{create_experiment}}

\item{marg_liks}{a data frame with marginal likelihoods/evidences.
A test data frame can be created by \link{create_test_marg_liks}}
}
\description{
Is the experiment the one with the most evidence?
}
