% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_dna_seq.R
\name{is_dna_seq}
\alias{is_dna_seq}
\title{Determine if the string is a lowercase DNA sequence
of at least one base pair}
\usage{
is_dna_seq(s)
}
\arguments{
\item{s}{the string to be checked}
}
\value{
TRUE if the string is a lowercase DNA sequence
  of at least one base pair
}
\description{
Determine if the string is a lowercase DNA sequence
of at least one base pair
}
\examples{

# TRUE: valid and lowercase characters
is_dna_seq("acgt"))

# FALSE: Must be lowercase
is_dna_seq("AGCT"))

# FALSE: Must be only valid characters
is_dna_seq("xxxx"))

# FALSE: Must have at least one nucleotide
is_dna_seq(""))
}
\author{
Richèl J.C. Bilderbeek
}
