% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_one_int.R
\name{is_one_int}
\alias{is_one_int}
\title{Determines if the argument is one int}
\usage{
is_one_int(x)
}
\arguments{
\item{x}{the object to be determined of if it is one integer}
}
\description{
Determines if the argument is one int
}
\author{
Richèl J.C. Bilderbeek
}
