% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_pir_params.R
\name{is_pir_params}
\alias{is_pir_params}
\title{Determine if the \code{pir_params} is valid.}
\usage{
is_pir_params(pir_params, verbose = FALSE)
}
\arguments{
\item{pir_params}{the parameters of \link[pirouette]{pirouette}.
They are created by \link{create_pir_params}.}

\item{verbose}{if TRUE, show more output}
}
\description{
Determine if the \code{pir_params} is valid.
}
