% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo_to_errors.R
\name{phylo_to_errors}
\alias{phylo_to_errors}
\title{Measure the error BEAST2 makes from a true phylogeny.}
\usage{
phylo_to_errors(phylogeny, alignment_params,
  experiment = create_experiment(),
  error_measure_params = create_error_measure_params())
}
\arguments{
\item{phylogeny}{a phylogeny of class \link[ape]{phylo}}

\item{alignment_params}{parameters to simulate an alignment,
as can be created by \link{create_alignment_params}}

\item{experiment}{a \link{pirouette} experiment,
as can be created by \link{create_experiment}}

\item{error_measure_params}{parameter set to specify how the
error between the given phylogeny and the Bayesian
posterior is determined.
Use \link{create_error_measure_params} to create such
a parameter set}
}
\description{
The supplied phylogeny (\code{phylogeny}) is the true/known phylogeny.
From the phylogeny, already an alignment is simulated and saved
as a FASTA file with name \code{alignment_params$fasta_filename}.
}
\details{
The posterior phylogenies are compared to the true/known phylogeny
using the nLTT statistics. These nLTT statistics, all with values
betweem (including) zero and (including) one, are returned.
}
\author{
Richèl J.C. Bilderbeek
}
