% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo_to_nltts.R
\name{phylo_to_nltts}
\alias{phylo_to_nltts}
\title{Create the NLTT statistics distribution files from a given phylogeny.
Additional information about mcmc and number of base pairs are needed}
\usage{
phylo_to_nltts(phylogeny, mcmc, n_base_pairs, seed)
}
\arguments{
\item{phylogeny}{a phylogeny of class \link[ape]{phylo}}

\item{mcmc}{MCMC options, as created by \link[beautier]{create_mcmc}}

\item{seed}{a random number generator seed}
}
\value{
nltt statistics
}
\description{
Create the NLTT statistics distribution files from a given phylogeny.
Additional information about mcmc and number of base pairs are needed
}
\author{
Giovanni Laudanno, Richel J.C. Bilderbeek
}
