% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo_to_posterior.R
\name{phylo_to_posterior}
\alias{phylo_to_posterior}
\title{Creates a posterior from the phylogeny}
\usage{
phylo_to_posterior(phylogeny, root_sequence = create_mono_nuc_dna(length
  = sequence_length), mutation_rate, mcmc,
  site_model = beautier::create_jc69_site_model(),
  clock_model = beautier::create_strict_clock_model(),
  tree_prior = beautier::create_bd_tree_prior(), crown_age = NA,
  mrca_distr = NA, alignment_rng_seed = 0, beast2_rng_seed = 1,
  verbose = FALSE, beast2_path = beastier::get_default_beast2_path())
}
\arguments{
\item{phylogeny}{a phylogeny of class \link[ape]{phylo}}

\item{root_sequence}{the DNA sequence at the root of the phylogeny.
By default, this will consist out of only adenine}

\item{mutation_rate}{the mutation rate per base pair per time unit}

\item{mcmc}{MCMC options, as created by \link[beautier]{create_mcmc}}

\item{site_model}{a nucleotide substitution model,
as created by \link[beautier]{create_site_model}}

\item{clock_model}{a clock model,
as created by \link[beautier]{create_clock_model}}

\item{tree_prior}{a tree prior,
as created by \link[beautier]{create_tree_prior}}

\item{crown_age}{the fixed crown age of the posterior. Set to NA
to let it be estimated}

\item{mrca_distr}{if MRCA prior used on all taxa.
Set to NA to not use an MRCA prior}

\item{alignment_rng_seed}{The random number generator seed used
to generate an alignment}

\item{beast2_rng_seed}{The random number generator seed used by BEAST2}

\item{verbose}{if TRUE, show more output}

\item{beast2_path}{Path to the
BEAST2 jar file (\code{beast.jar})
or BEAST2 binary file '(\code{beast})'.
Use \link[beastier]{get_default_beast2_jar_path} for the default
BEAST2 jar file path.
Use \link[beastier]{get_default_beast2_bin_path} for the default
BEAST2 binary file path.}
}
\value{
a posterior of phylogenies
}
\description{
Creates a posterior from the phylogeny
}
\author{
Richel J.C. Bilderbeek
}
