% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pir_plot.R
\name{pir_plot}
\alias{pir_plot}
\title{Plot the error BEAST2 make from the known phylogeny}
\usage{
pir_plot(pir_out)
}
\arguments{
\item{pir_out}{the output created by \code{\link{pir_run}}}
}
\value{
a ggplot2 plot
}
\description{
Plot the error BEAST2 make from the known phylogeny
}
\author{
Richel J.C. Bilderbeek
}
