% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pir_plot_from_file.R
\name{pir_plot_from_file}
\alias{pir_plot_from_file}
\title{Plot the error BEAST2 make from the known phylogeny}
\usage{
pir_plot_from_file(pir_out_filename)
}
\arguments{
\item{pir_out_filename}{name of the file with the saved
output as created by \code{\link{pir_run}}}
}
\value{
a ggplot2 plot
}
\description{
Plot the error BEAST2 make from the known phylogeny
}
\seealso{
Use \link{pir_plot} to directly plot the return value
  of \link{pir_run}
}
\author{
Richèl J.C. Bilderbeek
}
