% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pir_plot_from_long.R
\name{pir_plot_from_long}
\alias{pir_plot_from_long}
\title{Plot the error 'BEAST2' makes from a known phylogeny
from tidy data}
\usage{
pir_plot_from_long(
  tree_and_model_errors,
  tree_and_model_labels = get_tree_and_model_descriptions()
)
}
\arguments{
\item{tree_and_model_errors}{a tibble of a \code{tree_and_model}
and errors, which passes \link{check_tree_and_model_errors}}

\item{tree_and_model_labels}{tabular data that maps
a \code{tree_and_model} (e.g. \code{generative_true}) to
a description (e.g. "Generative, true"),
as created by \link{get_tree_and_model_descriptions}}
}
\value{
a \code{ggplot2} plot
}
\description{
Plot the error 'BEAST2' makes from a known phylogeny
from tidy data
}
\author{
Richèl J.C. Bilderbeek, Giovanni Laudanno
}
