% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pir_plots.R
\name{pir_plots}
\alias{pir_plots}
\title{Plot the output of \link{pir_runs}.}
\usage{
pir_plots(pir_outs)
}
\arguments{
\item{pir_outs}{the output of \link{pir_runs}}
}
\value{
a \link{ggplot2} plot
}
\description{
Plot the output of \link{pir_runs}.
}
\author{
Richèl J.C. Bilderbeek
}
