% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pir_rename.R
\name{pir_rename}
\alias{pir_rename}
\title{Rename the filenames in a \code{pir_params}
using a rename function.}
\usage{
pir_rename(pir_params, rename_fun)
}
\description{
Rename the filenames in a \code{pir_params}
using a rename function.
}
\author{
Richèl J.C. Bilderbeek
}
