% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pir_rename_to_std.R
\name{pir_rename_to_std}
\alias{pir_rename_to_std}
\title{Rename the \code{pir_params} filenames to follow a standard naming scheme.}
\usage{
pir_rename_to_std(pir_params, folder_name)
}
\arguments{
\item{pir_params}{the parameters of \link[pirouette]{pirouette}.
They are created by \link{create_pir_params}.}

\item{folder_name}{name of the main folder}
}
\description{
By default, \code{pir_params} uses temporary filenames for all files.
For \link{pir_run}, when only a computer reads those filenames,
this is fine. This function conformizes the filenames to
a human-friendly form.
}
\details{
The standard naming scheme is this:
\itemize{
  \item \code{pir_params$alignment_params$fasta_filename}
    becomes \code{[folder_name]/alignment.fas}
}

For the (zero or one) experiment at index \code{i}
that is generative:

\itemize{
  \item \code{pir_params$experiments[[i]]$beast2_options$input_filename}
    becomes \code{[folder_name]/gen.xml}
  \item
    \code{pir_params$experiments[[i]]$beast2_options$output_state_filename}
    becomes \code{[folder_name]/gen.xml.state}
  \item
    \code{pir_params$experiments[[i]]$beast2_options$input_filename}
    becomes \code{[folder_name]/errors_gen.csv},
  \item
    \code{pir_params$experiments[[i]]$inference_model$mcmc$tracelog$filename}
    becomes \code{[folder_name]/gen.log}
  \item
    \code{pir_params$experiments[[i]]$inference_model$mcmc$screenlog$filename}
    becomes \code{[folder_name]/gen.csv}
  \item
    \code{pir_params$experiments[[i]]$inference_model$mcmc$treelog$filename}
    becomes \code{[folder_name]/gen.trees}
  \item
    \code{pir_params$experiments[[i]]$est_evidence_mcmc$tracelog$filename}
    becomes \code{[folder_name]/gen_evidence.log}
  \item
    \code{pir_params$experiments[[i]]$est_evidence_mcmc$screenlog$filename}
    becomes \code{[folder_name]/gen_evidence.csv}
  \item
    \code{pir_params$experiments[[i]]$est_evidence_mcmc$treelog$filename}
    becomes \code{[folder_name]/gen_evidence.trees}
}
}
\seealso{
use \link{get_pir_params_filenames} to obtain all the filenames
}
