% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pir_run.R
\name{pir_run_tree}
\alias{pir_run_tree}
\title{Measure the error BEAST2 makes from a phylogeny}
\usage{
pir_run_tree(phylogeny, tree_type = "true",
  pir_params = create_test_pir_params())
}
\arguments{
\item{phylogeny}{a phylogeny of class \link[ape]{phylo}}

\item{tree_type}{type of tree, can be \code{true} for the true
phylogeny, and \code{twin} for its twin tree}

\item{pir_params}{the parameters of \link[pirouette]{pirouette}.
They are created by \link{create_pir_params}.}
}
\value{
a data frame with errors, with as many rows as model selection
  parameter sets
}
\description{
The phylogeny can be the true tree or its twin.
}
\author{
Richèl J.C. Bilderbeek, Giovanni Laudanno
}
