% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pir_run_true_tree.R
\name{pir_run_true_tree}
\alias{pir_run_true_tree}
\title{Measure the error BEAST2 makes from a phylogeny}
\usage{
pir_run_true_tree(true_phylogeny, pir_params)
}
\arguments{
\item{true_phylogeny}{a phylogeny of class \link[ape]{phylo}}

\item{pir_params}{the parameters of \link[pirouette]{pirouette}.
They are created by \link{create_pir_params}.}
}
\value{
a data frame with errors, with as many rows as model selection
  parameter sets
}
\description{
The phylogeny can be the true tree or its twin.
}
\examples{
library(testthat)

pir_params <- create_test_pir_params()

# The user should never need to initialize the pir_params
# but a develop calling this function needs to
pir_params <- init_pir_params(pir_params)


df <- pir_run_true_tree(
  true_phylogeny = ape::rcoal(4),
  pir_params = pir_params
)

expect_equal(1, nrow(df))
expected_col_names <- c(
  "tree", "inference_model", "inference_model_weight", "site_model",
  "clock_model", "tree_prior", "error_1", "error_2", "error_3"
)
expect_true(all(expected_col_names \%in\% names(df)))
}
\author{
Richèl J.C. Bilderbeek, Giovanni Laudanno
}
