% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pir_runs.R
\name{pir_runs}
\alias{pir_runs}
\title{Do multiple \link{pirouette} runs}
\usage{
pir_runs(phylogenies, pir_paramses)
}
\arguments{
\item{phylogenies}{a list of phylogenies,
each phylogeny being of class \link[ape]{phylo}}

\item{pir_paramses}{a list of \link[pirouette]{pirouette} parameters,
each element created by \link{create_pir_params}.}
}
\description{
Do multiple \link{pirouette} runs
}
\examples{
if (is_on_travis() && is_beast2_installed()) {

  pir_paramses <- list()
  pir_paramses[[1]] <- pirouette::create_test_pir_params()
  pir_paramses[[2]] <- pirouette::create_test_pir_params()

  phylogenies <- list()
  phylogenies[[1]] <- ape::read.tree(text = "((A:2, B:2):1, C:3);")
  phylogenies[[2]] <- ape::read.tree(text = "((A:1, B:1):2, C:3);")

  pir_outs <- pir_runs(
    phylogenies = phylogenies,
    pir_paramses = pir_paramses
  )
  for (pir_out in pir_outs) {
    expect_silent(check_pir_out(pir_out))
  }
}
}
\seealso{
Use \link{pir_run} for a single \link{pirouette} run.
}
