% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pir_table.R
\name{pir_table}
\alias{pir_table}
\title{Displays the results of \link{pir_run} as a table}
\usage{
pir_table(run_experiments)
}
\arguments{
\item{run_experiments}{a list of one or more \link{pirouette} run experiments}
}
\value{
the results of \link{pir_run} as a table
}
\description{
Displays the results of \link{pir_run} as a table
}
\seealso{
Use \link{pir_plot} to display the output of \link{pir_run} as a
  figure.
}
\author{
Richel J.C. Bilderbeek
}
