% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pirouette.R
\name{pirouette}
\alias{pirouette}
\title{pirouette}
\usage{
pirouette(x, y, prob = 1/sqrt(ncol(x)), ctrl = pirouetteControl(),
  gbm_control = list(n.trees = 1, interaction.depth = 2, shrinkage = 0.1,
  verbose = FALSE), ...)
}
\arguments{
\item{x}{a sparse matrix of x variables}

\item{y}{the target variable for classification or regression}

\item{prob}{the probability each entry of the rotation matrix is non-zero.  1/3 is one heuristic to use (for 3x faster processing), 1/sqrt(ncol(x)) is another, for sqrt(x)-fold faster processing.  Passed through enpointe to splat.}

\item{ctrl}{a list of control parameters for the algorithm}

\item{gbm_control}{a list of control parameters for the gbm}

\item{...}{passed through enpointe to gbm}
}
\value{
an object of class pirouette
}
\description{
Random rotation forests for sparse data
}
\references{
\itemize{
 \item{\url{https://stat.ethz.ch/pipermail/r-sig-hpc/2013-January/001575.html}}
 \item{\url{http://web.stanford.edu/~hastie/Papers/Ping/KDD06_rp.pdf}}
 \item{\url{http://stats.lse.ac.uk/fryzlewicz/rre/rre.pdf}}
}
}

