% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pirouette.R
\name{pirouetteControl}
\alias{pirouetteControl}
\title{pirouetteControl}
\usage{
pirouetteControl(mtry = 2, nobs = 500)
}
\arguments{
\item{mtry}{dimensionality of the random projection}

\item{nobs}{maximum number of observations used per tree}
}
\value{
a list
}
\description{
Define a random sparse projection
}

