% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_alignment_from_file.R
\name{plot_alignment_from_file}
\alias{plot_alignment_from_file}
\title{the plot title}
\usage{
plot_alignment_from_file(fasta_filename, title)
}
\arguments{
\item{fasta_filename}{name of a FASTA file}
}
\description{
Plot an alignment stored as a file
}
