% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_phylo.R
\name{plot_phylo}
\alias{plot_phylo}
\title{Plot a phylogeny}
\usage{
plot_phylo(phylogeny)
}
\arguments{
\item{phylogeny}{a phylogeny of class \link[ape]{phylo}}
}
\value{
a fancy tree plot
}
\description{
Plot a phylogeny
}
\examples{
phylogeny <- ape::read.tree(text = "((A:1, B:1):1, C:2);")
pirouette::plot_phylo(phylogeny)
}
\author{
Giovanni
}
