% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pirouette.R
\name{predict.enpointe}
\alias{predict.enpointe}
\title{Make predictions from a enpointe object}
\usage{
\method{predict}{enpointe}(object, newx, ...)
}
\arguments{
\item{object}{an enpointe object}

\item{newx}{A sparse matrix.}

\item{...}{passed to predict.rpart}
}
\value{
a vector
}
\description{
Predict from a single rotation tree.
}
\examples{
m <- enpointe(matrix(runif(1000), nrow=100), runif(100))
predict(m, matrix(runif(10), ncol=10))
}

