% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pirouette.R
\name{predict.pirouette}
\alias{predict.pirouette}
\title{Make predictions from a pirouette object}
\usage{
\method{predict}{pirouette}(object, newx, allowParallel = FALSE, ...)
}
\arguments{
\item{object}{an pirouette object}

\item{newx}{A sparse matrix.}

\item{allowParallel}{Allow predictions in parallel}

\item{...}{passed to predict.enpointe}
}
\value{
a vector
}
\description{
Predict from a pirouette forest
}
\examples{
nrow <- 10000
ncol <- 100
m <- pirouette(matrix(runif(nrow * ncol), ncol=ncol), runif(nrow), distribution='gaussian')
predict(m, matrix(runif(ncol), ncol=ncol))
}

