% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pirouette.R
\name{predict.splat}
\alias{predict.splat}
\title{Make predictions from a splatted object}
\usage{
\method{predict}{splat}(object, newx, ...)
}
\arguments{
\item{object}{a splat object}

\item{newx}{A sparse matrix.}

\item{...}{ignored}
}
\value{
a sparse Matrix
}
\description{
Splat a new matrix into the same rotation as an old one.
}
\examples{
x <- splat(matrix(runif(100), ncol=10), 2)
predict(x, matrix(runif(10), ncol=10))
}

