% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_errors_csv.R
\name{read_errors_csv}
\alias{read_errors_csv}
\title{Read the errors from a \code{.csv} file}
\usage{
read_errors_csv(errors_filename)
}
\arguments{
\item{errors_filename}{baseline name for errors filenames}
}
\value{
a numeric vector of error values
}
\description{
Read the errors from a \code{.csv} file
}
\author{
Richèl J.C. Bilderbeek
}
