% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relevel_inference_model.R
\name{relevel_inference_model}
\alias{relevel_inference_model}
\title{Internal function}
\usage{
relevel_inference_model(inference_model)
}
\arguments{
\item{inference_model}{one or more inference model types, either
\code{generative} or \code{candidate}}
}
\value{
a releveled \code{inference_model}
}
\description{
Internal function to relevel the \code{inference_model}, so that
\link{pir_plot} has the legend labels in the right order
}
\author{
Richèl J.C. Bilderbeek
}
