% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renum_rng_seeds.R
\name{renum_rng_seeds}
\alias{renum_rng_seeds}
\title{Renumber the RNG seeds}
\usage{
renum_rng_seeds(pir_paramses, rng_seeds)
}
\arguments{
\item{pir_paramses}{a list of \link[pirouette]{pirouette} parameters,
each element created by \link{create_pir_params}.}

\item{rng_seeds}{a vector of random number generator seeds}
}
\value{
a `pir_paramses` with renumbered RNG seeds
}
\description{
Renumber the RNG seeds
}
