% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicate_pir_params.R
\name{replicate_pir_params}
\alias{replicate_pir_params}
\title{Replicate pir_params assigning new names to each file}
\usage{
replicate_pir_params(pir_params, n_replicates)
}
\arguments{
\item{pir_params}{the parameters of \link[pirouette]{pirouette}.
They are created by \link{create_pir_params}.}

\item{n_replicates}{number of replicas to evaluate in order to create the
twin tree}
}
\value{
a list of replicated pir_params.
}
\description{
Replicate pir_params assigning new names to each file
}
\author{
Giovanni Laudanno
}
