% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_pir_param_files.R
\name{rm_pir_param_files}
\alias{rm_pir_param_files}
\title{Deletes all files}
\usage{
rm_pir_param_files(pir_params)
}
\arguments{
\item{pir_params}{the parameters of \link[pirouette]{pirouette}.
They are created by \link{create_pir_params}.}
}
\value{
Nothing
}
\description{
Deletes all files
}
\examples{
library(testthat)

pir_params <- create_test_pir_params(
  experiments = list(create_test_gen_experiment())
)

filenames <- c(
  pir_params$alignment_params$fasta_filename,
  pir_params$experiments[[1]]$beast2_options$input_filename,
  pir_params$experiments[[1]]$beast2_options$output_log_filename,
  pir_params$experiments[[1]]$beast2_options$output_trees_filenames,
  pir_params$experiments[[1]]$beast2_options$output_state_filename
)

if (is_on_travis() && is_beast2_installed()) {
  # Running all one experiments
  errors <- pir_run(
    phylogeny = ape::read.tree(text = "((A:2, B:2):1, C:3);"),
    pir_params = pir_params
  )
  # Files are created
  testit::assert(all(file.exists(filenames)))

  # Removing the files
  rm_pir_param_files(pir_params)

  # All files should be gone
  expect_true(all(!file.exists(filenames)))
}
}
\author{
Richèl J.C. Bilderbeek
}
