% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run}
\alias{run}
\title{Creates a posterior from the phylogeny}
\usage{
run(phylogeny)
}
\arguments{
\item{phylogeny}{a phylogeny}
}
\value{
a posterior of phylogenies
}
\description{
Creates a posterior from the phylogeny
}
\author{
Richel J.C. Bilderbeek
}
