% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_candidate_evidences.R
\name{select_candidate_evidences}
\alias{select_candidate_evidences}
\title{Select the evidences for candidate experiments}
\usage{
select_candidate_evidences(experiments = list(create_test_experiment),
  marg_liks = create_test_marg_liks())
}
\description{
Select the evidences for candidate experiments
}
\author{
Richel J.C. Bilderbeek
}
