% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_experiments.R
\name{select_experiments}
\alias{select_experiments}
\title{Select the experiments to do a Bayesian inference with.}
\usage{
select_experiments(experiments = list(create_test_experiment()),
  marg_liks = NULL, verbose = FALSE)
}
\arguments{
\item{experiments}{a list of one or more \link{pirouette} experiments,
as can be created by \link{create_experiment}}

\item{marg_liks}{a data frame with marginal likelihoods/evidences.
A test data frame can be created by \link{create_test_marg_liks}}

\item{verbose}{if TRUE, show more output}
}
\value{
a list of inference models
}
\description{
Select the experiments to do a Bayesian inference with.
}
\author{
Richel J.C. Bilderbeek
}
