% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_inference_models.R
\name{select_inference_models}
\alias{select_inference_models}
\title{Select inference model}
\usage{
select_inference_models(alignment_params, model_select_params,
  marg_liks = NULL)
}
\arguments{
\item{alignment_params}{parameters to simulate an alignment,
as can be created by \link{create_alignment_params}}

\item{model_select_params}{a list of one or more parameter sets
to select an inference model
(a combination of site model, clock model and tree prior),
as can be created by \link{create_model_select_param}}
}
\description{
Select inference model
}
