% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shorten_pir_params.R
\name{shorten_pir_params}
\alias{shorten_pir_params}
\title{Shorten the \code{pir_params}}
\usage{
shorten_pir_params(pir_params)
}
\arguments{
\item{pir_params}{the parameters of \link[pirouette]{pirouette}.
They are created by \link{create_pir_params}.}
}
\description{
Shorten the \code{pir_params}
}
