% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shorten_pir_paramses.R
\name{shorten_pir_paramses}
\alias{shorten_pir_paramses}
\title{Shorten the list of \code{pir_params}}
\usage{
shorten_pir_paramses(pir_paramses)
}
\arguments{
\item{pir_paramses}{a list of \link[pirouette]{pirouette} parameters,
each element created by \link{create_pir_params}.}
}
\value{
a shortened \link{list} of \code{pir_params}
}
\description{
Shorten the list of \code{pir_params}
}
