% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_alignment.R
\name{sim_alignment}
\alias{sim_alignment}
\title{Converts a phylogeny to a random DNA alignment}
\usage{
sim_alignment(phylogeny, alignment_params)
}
\arguments{
\item{phylogeny}{a phylogeny of class \link[ape]{phylo}}

\item{alignment_params}{parameters to simulate an alignment,
as can be created by \link{create_alignment_params}}
}
\value{
an alignment of type \code{DNAbin}
}
\description{
Converts a phylogeny to a random DNA alignment
}
\examples{
n_taxa <- 5
n_base_pairs <- 4
alignment <- sim_alignment(
   phylogeny = ape::rcoal(n_taxa),
   alignment_params = create_alignment_params(
     root_sequence = create_blocked_dna(length = n_base_pairs),
     mutation_rate = 0.1
   )
 )
 testit::assert(class(alignment) == "DNAbin")
 testit::assert(nrow(alignment) == n_taxa)
 testit::assert(ncol(alignment) == n_base_pairs)
}
\author{
Richel Bilderbeek
}
