% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_alignment_file.R
\name{sim_alignment_file}
\alias{sim_alignment_file}
\title{Simulates a DNA alignment and saves it to a FASTA file.}
\usage{
sim_alignment_file(phylogeny, alignment_params)
}
\arguments{
\item{phylogeny}{a phylogeny of class \link[ape]{phylo}}

\item{alignment_params}{parameters to simulate an alignment,
as can be created by \link{create_alignment_params}}
}
\value{
nothing
}
\description{
The simulation is performed by \link{sim_alignment}.
}
\examples{
library(testthat)

n_taxa <- 5
n_base_pairs <- 4
fasta_filename <- tempfile()

# File does not exist yet
expect_false(file.exists(fasta_filename))

alignment <- sim_alignment_file(
  phylogeny = ape::rcoal(n_taxa),
  alignment_params = create_alignment_params(
    root_sequence = create_blocked_dna(length = n_base_pairs),
    mutation_rate = 0.1,
    fasta_filename = fasta_filename
  )
)
# File does exist now
expect_true(file.exists(fasta_filename))
}
\author{
Richèl Bilderbeek
}
