% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_alignment.R
\name{sim_alignment_raw}
\alias{sim_alignment_raw}
\title{Converts a phylogeny to a random DNA alignment}
\usage{
sim_alignment_raw(phylogeny, root_sequence, rng_seed, mutation_rate,
  site_model, n_mutations = NA)
}
\arguments{
\item{phylogeny}{a phylogeny of class \link[ape]{phylo}}

\item{root_sequence}{the DNA sequence at the root of the phylogeny.
By default, this will consist out of an equal amount of each letter}

\item{rng_seed}{a random number generator seed}

\item{mutation_rate}{the mutation rate per base pair per time unit}

\item{site_model}{a nucleotide substitution model,
as created by \link[beautier]{create_site_model}}

\item{n_mutations}{the number of different base pairs between
root sequence and the resulting alignment}
}
\value{
an alignment of type \code{DNAbin}
}
\description{
Converts a phylogeny to a random DNA alignment
}
\examples{
library(testthat)

# Create the ancestor's DNA sequence
n_base_pairs <- 4
root_sequence <- create_blocked_dna(length = n_base_pairs)
mutation_rate <- 0.1
n_taxa <- 5
rng_seed <- 314
phylogeny <- ape::rcoal(n_taxa)
site_model <- create_jc69_site_model()

# Simulate the alignment
alignment <- sim_alignment_raw(
   phylogeny = phylogeny,
   root_sequence = root_sequence,
   rng_seed = rng_seed,
   mutation_rate = mutation_rate,
   site_model = site_model
 )

expect_equal(class(alignment), "DNAbin")
expect_equal(nrow(alignment), n_taxa)
expect_equal(ncol(alignment), n_base_pairs)
}
\seealso{
Use \link{sim_alignment_file} to save the simulated alignment
  directly to a file
}
\author{
Richèl J.C. Bilderbeek, Giovanni Laudanno
}
