% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_alignment_twin.R
\name{sim_alignment_twin}
\alias{sim_alignment_twin}
\title{Converts a twin phylogeny to a random DNA alignment}
\usage{
sim_alignment_twin(twin_phylogeny, root_sequence, rng_seed_twin_alignment,
  mutation_rate, site_model, n_mutations, verbose = FALSE)
}
\arguments{
\item{twin_phylogeny}{a phylogeny of class \link[ape]{phylo}}

\item{root_sequence}{the DNA sequence at the root of the phylogeny.
By default, this will consist out of an equal amount of each letter}

\item{rng_seed_twin_alignment}{the random number generator seed
as used in the simulation of a twin alignment}

\item{mutation_rate}{the mutation rate per base pair per time unit}

\item{site_model}{a nucleotide substitution model,
as created by \link[beautier]{create_site_model}}

\item{n_mutations}{costrained number of mutations}

\item{verbose}{if TRUE, show more output}
}
\value{
an alignment of type \code{DNAbin}
}
\description{
Converts a twin phylogeny to a random DNA alignment
}
\seealso{
Use \link{sim_alignment_twin_file}
  to save the simulated alignment directly to a file
}
\author{
Richèl J.C. Bilderbeek
}
