% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_alignment_twin_file.R
\name{sim_alignment_twin_file}
\alias{sim_alignment_twin_file}
\title{Simulates a twin DNA alignment and saves it to a FASTA file.}
\usage{
sim_alignment_twin_file(twin_phylogeny, alignment_params, twinning_params)
}
\arguments{
\item{twin_phylogeny}{a phylogeny of class \link[ape]{phylo}}

\item{alignment_params}{parameters to simulate an alignment,
as can be created by \link{create_alignment_params}}

\item{twinning_params}{can be \code{NA} if no twinning is desired,
or can be the twinning parameters,
as can be created by \link{create_twinning_params}}
}
\value{
nothing
}
\description{
The simulation is performed by \link{sim_alignment_twin}.
}
\seealso{
Use \link{sim_alignment_twin} to only simulate the twin alignment,
  without saving the alignment to file
}
\author{
Richèl J.C. Bilderbeek
}
