% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/sim_true_alignment_with_linked_node_sub_site_model.R
\name{sim_true_alignment_with_linked_node_sub_site_model}
\alias{sim_true_alignment_with_linked_node_sub_site_model}
\title{Create an alignment with the \code{linked_node_sub} site model}
\usage{
sim_true_alignment_with_linked_node_sub_site_model(true_phylogeny,
  root_sequence, subst_matrix = NULL, branch_mutation_rate = 1,
  node_mutation_rate = 1, base_frequencies = NULL, node_time = 0.001)
}
\arguments{
\item{true_phylogeny}{the true phylogeny; the actual evolutionary
history of the species, of class \link[ape]{phylo}}

\item{root_sequence}{the DNA sequence at the root of the phylogeny.
By default, this will consist out of an equal amount of each letter
Use \link{check_root_sequence} to check if a root sequence is valid.}

\item{subst_matrix}{nucleotide substitution matrix}

\item{branch_mutation_rate}{mutation rate along the branch.
See, among others, \link[nodeSub]{sim_dual_independent} for more details}

\item{node_mutation_rate}{mutation rate on the node.
See, among others, \link[nodeSub]{sim_dual_independent} for more details}

\item{base_frequencies}{the four base frequencies (a, c, g, t) to be
specified to create the rate matrix (i.e. Q matrix)
used to simulate alignments}

\item{node_time}{amount of time spent at the nodes.
See, among others, \link[nodeSub]{sim_dual_independent} for more details}
}
\value{
an alignment of type \code{DNAbin}
}
\description{
Create an alignment with the \code{linked_node_sub} site model
}
\seealso{
Use \link{sim_true_alignment_with_unlinked_node_sub_site_model}
to simulate the true alignment with an unlinked node substitution model.
Use \link{sim_twin_alignment_with_linked_node_sub_site_model}
to simulate the twin alignment with an linked node substitution model.
Use \link{get_sim_true_alignment_with_linked_node_sub_site_model_function}
to get a partially evaluated unary function.
}
