% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_true_alignment_with_standard_site_model.R
\name{sim_true_alignment_with_standard_site_model}
\alias{sim_true_alignment_with_standard_site_model}
\title{Simulate a twin alignment using a standard site model}
\usage{
sim_true_alignment_with_standard_site_model(true_phylogeny,
  root_sequence = "acgt", mutation_rate = 0.1,
  site_model = beautier::create_jc69_site_model())
}
\arguments{
\item{true_phylogeny}{a phylogeny of class \link[ape]{phylo}}

\item{root_sequence}{the DNA sequence at the root of the phylogeny.
By default, this will consist out of an equal amount of each letter
Use \link{check_root_sequence} to check if a root sequence is valid.}

\item{mutation_rate}{the mutation rate per base pair per time unit.
Use \link{check_mutation_rate} to check if a mutation rate is valid.}

\item{site_model}{a nucleotide substitution model,
which can be:
\itemize{
  \item{
    A standard nucloetide substitution model,
    as created by \link[beautier]{create_site_model}
  }
  \item{
    \code{linked_node_sub}: a linked node-substitution model
  }
  \item{
    \code{unlinked_node_sub}: an unlinked node-substitution model
  }
}}
}
\description{
This is an adapter function
(see \url{https://en.wikipedia.org/wiki/Adapter_pattern}),
with the purpose of passing \link{check_sim_twin_alignment_function},
by being a function with the function arguments \code{twin_phylogeny}
and \code{true_alignment}.
}
\examples{
library(testthat)

# This adapter function must be a sim_true_alignment function
expect_silent(
  check_sim_true_alignment_function(
    sim_true_alignment_with_standard_site_model
  )
)

# Simulate the true DNA alignment
alignment <- sim_true_alignment_with_standard_site_model(
  true_phylogeny = ape::read.tree(text = "((A:1, B:1):2, C:3);"),
  root_sequence = "aaaa",
  mutation_rate = 0.1
)
expect_silent(check_alignment(alignment))
}
