% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/sim_true_alignment_with_unlinked_node_sub_site_model.R
\name{sim_true_alignment_with_unlinked_node_sub_site_model}
\alias{sim_true_alignment_with_unlinked_node_sub_site_model}
\title{Create an alignment with the \code{linked_node_sub} site model}
\usage{
sim_true_alignment_with_unlinked_node_sub_site_model(true_phylogeny,
  root_sequence, site_model = "unlinked_node_sub")
}
\value{
an alignment of type \code{DNAbin}
}
\description{
Create an alignment with the \code{linked_node_sub} site model
}
\seealso{
Use \link{sim_true_alignment_with_linked_node_sub_site_model}
to simulate the true alignment with a linked node substitution model.
Use \link{sim_twin_alignment_with_unlinked_node_sub_site_model}
to simulate the twin alignment with an unlinked node substitution model.
}
