% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/sim_twin_alignment_with_linked_node_sub_site_model.R
\name{sim_twin_alignment_with_linked_node_sub_site_model}
\alias{sim_twin_alignment_with_linked_node_sub_site_model}
\title{Adapter function to simulte the twin alignment
using the \code{linked_node_sub} site model}
\usage{
sim_twin_alignment_with_linked_node_sub_site_model(twin_phylogeny,
  true_alignment = "irrelevant", root_sequence, subst_matrix = NULL,
  branch_mutation_rate = 1, node_mutation_rate = 1,
  base_frequencies = NULL, node_time = 0.001)
}
\value{
an alignment of type \code{DNAbin}
}
\description{
Adapter function to simulte the twin alignment
using the \code{linked_node_sub} site model
}
\seealso{
Use \link{sim_twin_alignment_with_unlinked_node_sub_site_model}
to simulate using an unlinked node substitution model.
Use \link{sim_true_alignment_with_linked_node_sub_site_model}
to simulate a true alignment.
}
