% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{sparse_text_matrix}
\alias{sparse_text_matrix}
\title{Simulated text data}
\format{A dgCMatrix with 99,339 rows and 14,052 variables}
\source{
\url{http://stackoverflow.com/questions/31570437/really-fast-word-ngram-vectorization-in-r}
}
\usage{
sparse_text_matrix
}
\description{
A dataset containing simulated wordcounts for documents
}
\keyword{datasets}

