% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pirouette.R
\name{splat}
\alias{splat}
\title{splat}
\usage{
splat(x, mtry = 2, prob = 1/sqrt(ncol(x)), retx = TRUE)
}
\arguments{
\item{x}{a sparse matrix}

\item{mtry}{new dimensionality}

\item{prob}{the probability each entry of the rotation matrix is non-zero.  1/3 is one heuristic to use (for 3x faster processing), 1/sqrt(ncol(x)) is another, for sqrt(x)-fold faster processing.}

\item{retx}{If true, return the splatted matrix as well as the rotation.}
}
\value{
a splatted Matrix
}
\description{
Splat a matrix into a lower dimension.
}
\examples{
splat(matrix(runif(100), ncol=10), 2)
}
\references{
\url{http://web.stanford.edu/~hastie/Papers/Ping/KDD06_rp.pdf}
}

