% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_evidence_filename.R
\name{to_evidence_filename}
\alias{to_evidence_filename}
\title{Converts a filename to an evidence filename}
\usage{
to_evidence_filename(filename)
}
\arguments{
\item{filename}{the file's name, without the path}
}
\value{
evidence tree filename
}
\description{
Converts a filename to an evidence filename
}
\examples{
library(testthat)

filename <- "beast2_output.xml.state"
created <- to_evidence_filename(filename)
expected <- "beast2_output_evidence.xml.state"
expect_equal(expected, created)
}
\author{
Richèl J.C. Bilderbeek
}
