% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_fasta_text.R
\name{to_fasta_text}
\alias{to_fasta_text}
\title{Show a DNA alignment as text in FASTA format}
\usage{
to_fasta_text(alignment, alignment_file = tempfile(pattern =
  "alignment_", fileext = ".fasta"))
}
\description{
Show a DNA alignment as text in FASTA format
}
