% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_twin_filename.R
\name{to_twin_filename}
\alias{to_twin_filename}
\title{Converts true tree filenames to twin tree filenames}
\usage{
to_twin_filename(filename)
}
\arguments{
\item{filename}{the file's name, without the path}
}
\value{
twin tree filename
}
\description{
Converts true tree filenames to twin tree filenames
}
\author{
Giovanni Laudanno
}
