% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twin_to_bd_tree.R
\name{twin_to_bd_tree_obsolete}
\alias{twin_to_bd_tree_obsolete}
\title{Create an BD twin tree from a phylogeny
and save it as a file}
\usage{
twin_to_bd_tree_obsolete(phylogeny, twinning_params)
}
\arguments{
\item{phylogeny}{a phylogeny of class \link[ape]{phylo}}

\item{twinning_params}{can be \code{NA} if no twinning is desired,
or can be the twinning parameters,
as can be created by \link{create_twinning_params}}
}
\value{
a twin BD tree of class \link[ape]{phylo},
  obtained from the corresponding phylogeny.
}
\description{
Create an BD twin tree from a phylogeny
and save it as a file
}
\examples{
library(testthat)

phylogeny <- ape::read.tree(text = "((A:2, B:2):1, C:3);")
twinning_params <- create_twinning_params()
bd_tree <- twin_to_bd_tree_obsolete(
  phylogeny = phylogeny,
  twinning_params = twinning_params
)

expect_equal(class(bd_tree), "phylo")

# Branching times will differ, except the crown
expect_false(
  all(
    ape::branching.times(phylogeny) ==
    ape::branching.times(bd_tree)
  )
)

# Crown age stays the same
expect_equal(
  max(ape::branching.times(bd_tree)),
  max(ape::branching.times(phylogeny))
)
}
\author{
Richèl J.C. Bilderbeek, Giovanni Laudanno
}
