% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twin_to_cbs_tree.R
\name{twin_to_cbs_tree}
\alias{twin_to_cbs_tree}
\title{Convert a phylogeny to a twin Coalescent Bayesian
Skyline tree}
\usage{
twin_to_cbs_tree(phylogeny = ape::read.tree(text =
  "((A:1, B:1):1, C:2);"), winning_params = create_twinning_params())
}
\arguments{
\item{phylogeny}{a phylogeny of class \link[ape]{phylo}}
}
\value{
a CBS tree
}
\description{
Convert a phylogeny to a twin Coalescent Bayesian
Skyline tree
}
\author{
Richel J.C Bilderbeek
}
