% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twin_to_yule_tree.R
\name{twin_to_yule_tree}
\alias{twin_to_yule_tree}
\title{Create a twin tree from a phylogeny using a Yule process}
\usage{
twin_to_yule_tree(phylogeny, twinning_params)
}
\arguments{
\item{phylogeny}{a phylogeny of class \link[ape]{phylo}}

\item{twinning_params}{can be \code{NA} if no twinning is desired,
or can be the twinning parameters,
as can be created by \link{create_twinning_params}}
}
\value{
a twin Yule tree of class \link[ape]{phylo},
}
\description{
Create a twin tree from a phylogeny using a Yule process
}
\examples{
library(testthat)

phylogeny <- ape::read.tree(text = "((A:2, B:2):1, C:3);")
twinning_params <- create_twinning_params()
yule_tree <- twin_to_yule_tree(
  phylogeny = phylogeny,
  twinning_params = twinning_params
)

expect_equal(class(yule_tree), "phylo")

# Branching times will differ, except the crown
expect_false(
  all(
    ape::branching.times(phylogeny) ==
    ape::branching.times(yule_tree)
  )
)

# Crown age stays the same
expect_equal(
  max(ape::branching.times(yule_tree)),
  max(ape::branching.times(phylogeny))
)
}
\author{
Richèl J.C. Bilderbeek, Giovanni Laudanno
}
