% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment_params_to_posterior_trees.R
\name{alignment_params_to_posterior_trees}
\alias{alignment_params_to_posterior_trees}
\title{Creates a posterior of phylogenies from a known alignment.}
\usage{
alignment_params_to_posterior_trees(alignment_params, experiment,
  verbose = FALSE)
}
\arguments{
\item{alignment_params}{parameters to simulate an alignment,
as can be created by \link{create_alignment_params}}

\item{experiment}{a \link{pirouette} experiment,
as can be created by \link{create_experiment}}

\item{verbose}{if TRUE, show more output}
}
\value{
a list of phylogenies in the posterior,
  as a \link[ape]{multiphylo}
}
\description{
Ignores the jointly-estimated posterior estimates
}
\examples{
library(testthat)

if (is_on_travis() && is_beast2_installed()) {

  alignment_params <- pirouette::create_test_alignment_params()
  pirouette::create_true_alignment_file(
    phylogeny = ape::read.tree(text = "((A:1, B:1):1, C:2);"),
    alignment_params = alignment_params
  )

  testthat::expect_true(file.exists(alignment_params$fasta_filename))

  trees <- pirouette::alignment_params_to_posterior_trees(
    alignment_params = alignment_params,
    experiment = create_test_experiment()
  )
  testthat::expect_equal("multiPhylo", class(trees))
}
}
\author{
Richèl J.C. Bilderbeek
}
