% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment_params_to_posterior_trees.R
\name{alignment_params_to_posterior_trees}
\alias{alignment_params_to_posterior_trees}
\title{Creates a posterior of phylogenies from a known alignment.}
\usage{
alignment_params_to_posterior_trees(alignment_params,
  experiment = create_experiment())
}
\arguments{
\item{alignment_params}{parameters to simulate an alignment,
as can be created by \link{create_alignment_params}}

\item{experiment}{a \link{pirouette} experiment,
as can be created by \link{create_experiment}}
}
\value{
a list of:
\itemize{
  \item \code{trees}: the phylogenies in the posterior,
    as a \link[ape]{multiphylo}
  \item \code{estimates}: the BEAST2 estimates, as a \link{data.frame}
}
}
\description{
Creates a posterior of phylogenies from a known alignment.
}
\author{
Richel J.C. Bilderbeek
}
