% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bd_phylo_2_l_table}
\alias{bd_phylo_2_l_table}
\title{Convert bd phylo to L table}
\usage{
bd_phylo_2_l_table(phylo)
}
\arguments{
\item{phylo}{a phylogeny of class \link[ape]{phylo}}
}
\value{
the L table
}
\description{
Convert bd phylo to L table. Don't use for mbd.
}
\author{
Xu Liang, Giovanni Laudanno, Richel J.C. Bilderbeek
}
