% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_rate_matrix.R
\name{calc_base_freq}
\alias{calc_base_freq}
\title{Calculate base frequencies}
\usage{
calc_base_freq(root_sequence)
}
\arguments{
\item{root_sequence}{the DNA sequence at the root of the phylogeny.
By default, this will consist out of an equal amount of each letter}
}
\value{
a frequencies vector
}
\description{
Calculate base frequencies
}
\author{
Giovanni Laudanno
}
