% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_alignment.R
\name{check_alignment}
\alias{check_alignment}
\title{Check if the alignment is of the right type}
\usage{
check_alignment(alignment)
}
\arguments{
\item{alignment}{a DNA alignment, of class \link[ape]{DNAbin}}
}
\description{
Will \link{stop} if not.
}
