% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_alignment.R
\name{check_alignment}
\alias{check_alignment}
\title{Check if the alignment is of the right type}
\usage{
check_alignment(alignment)
}
\arguments{
\item{alignment}{a DNA alignment, of class \link[ape]{DNAbin}}
}
\value{
nothing
}
\description{
Will \link{stop} if not.
}
\examples{
# Check cleanup by other functions
beastier::check_empty_beaustier_folders()

check_alignment(ape::as.DNAbin("ACGT"))
}
\author{
Richèl J.C. Bilderbeek
}
