% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_alignment_params.R
\name{check_alignment_params}
\alias{check_alignment_params}
\title{Checks if the argument is a valid alignment parameters structure.
Will \link{stop} if not}
\usage{
check_alignment_params(alignment_params)
}
\arguments{
\item{alignment_params}{parameters to simulate an alignment,
as can be created by \link{create_alignment_params}}
}
\value{
nothing. Will \link{stop} if nit
}
\description{
Checks if the argument is a valid alignment parameters structure.
Will \link{stop} if not
}
\author{
Richel J.C. Bilderbeek
}
