% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_experiments_candidates_have_same_mcmcs.R
\name{check_experiments_candidates_have_same_mcmcs}
\alias{check_experiments_candidates_have_same_mcmcs}
\title{Check that the candidate experiments have a same MCMC}
\usage{
check_experiments_candidates_have_same_mcmcs(experiments)
}
\description{
If all the candidates share a same MCMC, they share the same
tracelog and treelog. In this way, one can predict where the parameter
estimates (the trace) and posterior trees are written to, as only the
best candidate will run.
}
\author{
Richèl J.C. Bilderbeek
}
