% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_is_ns_beast2_pkg_installed.R
\name{check_is_ns_beast2_pkg_installed}
\alias{check_is_ns_beast2_pkg_installed}
\title{Checks if the \code{NS} BEAST2 package is installed.}
\usage{
check_is_ns_beast2_pkg_installed()
}
\description{
Will \link{stop} if not.
}
\examples{
  if (mauricer::is_beast2_pkg_installed("NS")) {
    testthat::expect_silent(
      check_is_ns_beast2_pkg_installed()
    )
  } else {
    testthat::expect_error(
      check_is_ns_beast2_pkg_installed()
    )
  }
}
\author{
Richel J.C. Bilderbeek
}
