% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_is_ns_beast2_pkg_installed.R
\name{check_is_ns_beast2_pkg_installed}
\alias{check_is_ns_beast2_pkg_installed}
\title{Checks if the \code{NS} BEAST2 package is installed.}
\usage{
check_is_ns_beast2_pkg_installed()
}
\description{
Will \link{stop} if not.
}
\examples{
  library(babette)
  library(testthat)

  # We need BEAST2 installed
  if (is_beast2_installed()) {

    if (is_beast2_pkg_installed("NS")) {
      # No error If NS is installed
      expect_silent(check_is_ns_beast2_pkg_installed())
    } else {
      # Error If NS is installed
      expect_error(check_is_ns_beast2_pkg_installed())
    }
  }
}
\author{
Richèl J.C. Bilderbeek
}
