% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pir_out.R
\name{check_pir_out}
\alias{check_pir_out}
\title{Checks if the argument is a valid \link{pirouette} parameter set.}
\usage{
check_pir_out(pir_out)
}
\arguments{
\item{pir_out}{the output of \link{pir_run_tree} applied on the original tree}
}
\value{
nothing. Will \link{stop} if not
}
\description{
Will \link{stop} if not.
A valid \link{pirouette} parameter set
can be created by \link{create_pir_params}.
}
\examples{
  library(testthat)

  expect_silent(check_pir_out(create_test_pir_run_output()))
  expect_error(check_pir_out("nonsense"))
  expect_error(check_pir_out(NULL))
  expect_error(check_pir_out(NA))
}
\author{
Richèl J.C. Bilderbeek
}
