% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pir_params.R
\name{check_pir_params}
\alias{check_pir_params}
\title{Checks if the argument is a valid \link{pirouette} parameter set.}
\usage{
check_pir_params(pir_params, max_evidence_epsilon = 0.0001)
}
\arguments{
\item{pir_params}{the parameters of \link[pirouette]{pirouette}.
They are created by \link{create_pir_params}.}

\item{max_evidence_epsilon}{set the maximum acceptable threshold for the
parameter \code{evidence_epsilon}}
}
\value{
nothing. Will \link{stop} if not
}
\description{
Will \link{stop} if not.
A valid \link{pirouette} parameter set
can be created by \link{create_pir_params}.
}
\examples{
 phylogeny <- ape::read.tree(text = "(((A:1, B:1):1, C:2):1, D:3);")
 pir_params <- create_pir_params(
 alignment_params = create_alignment_params(
     root_sequence = create_mono_nuc_dna(length = 4),
     mutation_rate = create_standard_mutation_rate(phylogeny)
   )
 )
 testthat::expect_silent(check_pir_params(pir_params))
}
\author{
Giovanni Laudanno, Richel J.C. Bilderbeek
}
