% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_result.R
\name{check_result}
\alias{check_result}
\title{Checks if the argument is a valid \link{pirouette} result.}
\usage{
check_result(result)
}
\value{
nothing. Will \link{stop} if not
}
\description{
Will \link{stop} if not.
A valid \link{pirouette} result
can be created by \link{create_result}.
}
\examples{
  testthat::expect_silent(
    check_result(create_test_result())
  )
  testthat::expect_error(check_result("nonsense"))
  testthat::expect_error(check_result(NA))
  testthat::expect_error(check_result(NULL))
}
\seealso{
Use \link{check_results} to check if an object
  is a list of results
}
\author{
Richel J.C. Bilderbeek
}
