% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_run_experiment.R
\name{check_run_experiment}
\alias{check_run_experiment}
\title{Check if the experiment is a run experiment}
\usage{
check_run_experiment(run_experiment)
}
\arguments{
\item{run_experiment}{one \link{pirouette} run experiment.
A run experiment is an extension of a normal experiment (that
has not been run), as it has added:
\itemize{
  \item log_evidence the natural logarithm of the evidence (aka marginal
    likelihood). Can be NA if this is not measured
  \item weight the weight of the model, compared to other (candidate)
    models. This weight will be between 0.0 (there is no evidence for
    this model) to 1.0 (all evidence indicates this is the best model).
    A weight of NA denotes that the weight is not measured
  \item errors a numeric vector of (positive) Bayesian inference errors.
    Will be NA if these are not measured.
}}
}
\value{
nothing. Will \link{stop} if not
}
\description{
Check if the experiment is a run experiment
}
\examples{
  testthat::expect_silent(
    check_run_experiment(create_test_run_experiment())
  )
}
\author{
Richel J.C. Bilderbeek
}
