% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_run_experiment.R
\name{check_run_experiment}
\alias{check_run_experiment}
\title{Check if the experiment is a run experiment}
\usage{
check_run_experiment(run_experiment)
}
\value{
nothing. Will \link{stop} if not
}
\description{
Check if the experiment is a run experiment
}
\examples{
  testthat::expect_silent(
    check_run_experiment(create_test_run_experiment())
  )
}
\author{
Richel J.C. Bilderbeek
}
