% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_sim_true_alignment_function.R
\name{check_sim_true_alignment_function}
\alias{check_sim_true_alignment_function}
\title{Check if the function to create a true alignment with (from the
true phylogeny) is valid.}
\usage{
check_sim_true_alignment_function(sim_true_alignment_function)
}
\arguments{
\item{sim_true_alignment_function}{function to simulate a
true alignment with.
This function must have two arguments,
called \code{true_phylogeny} (which will hold the true phylogeny)
and \code{root_sequence} (which holds the DNA root sequence).
The return type must be \link[ape]{DNAbin}.

Use \link{check_sim_true_alignment_function} to verify if the function
has the right signature and output.

Some standard functions:\cr
\itemize{
  \item Use \link{get_sim_true_alignment_with_standard_site_model_function}
  to get a function (\link{sim_true_alignment_with_std_site_model})
  the use a standard site model.
  \item Use
  \link{get_sim_true_alignment_with_linked_node_sub_site_model_function}
  to get a function (\link{sim_true_alignment_with_linked_node_sub_site_model})
  the use a linked node substitution site model.
  \item Use
  \link{get_sim_true_alignment_with_unlinked_node_sub_site_model_function}
  to get a function (\link{sim_true_alignment_with_unlinked_node_sub_site_model})
  the use an unlinked node substitution site model.
}}
}
\description{
Will \link{stop} if not
}
\seealso{
Use \link{check_sim_twin_alignment_function} to check a function to
generate a twin alignment
}
