% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_sim_twin_alignment_function.R
\name{check_sim_twin_alignment_function}
\alias{check_sim_twin_alignment_function}
\title{Check if the function to create a twin alignment with (from a
true phylogeny and a true alignment) is valid.}
\usage{
check_sim_twin_alignment_function(sim_twin_alignment_function)
}
\arguments{
\item{sim_twin_alignment_function}{function to simulate a
twin alignment with.
This function must have two arguments called \code{twin_phylogeny} (which
will hold the twin phylogeny) and \code{true_alignment} (which will
hold the alignment simulated from the true phylogeny). The
return type must be \link[ape]{DNAbin}.

Use \link{check_sim_twin_alignment_function} to verify if the function
has the right signature and output.

Some standard functions:\cr
\itemize{
  \item Use \link{get_default_sim_twin_alignment_function} to use a
    default function
  \item Use \link{get_sim_twin_alignment_with_same_n_mutation_function} to
    use a function that ensure that the twin alignment has as much
    mutations compared to the root sequence as the true alignment has

}}
}
\description{
Will \link{stop} if not
}
\seealso{
Use \link{check_sim_true_alignment_function} to check a function to
generate a true alignment.
Use \link{check_sim_twin_tree_function} to check a function to
generate a twin phylogeny.
}
