% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_twinning_params.R
\name{check_twinning_params}
\alias{check_twinning_params}
\title{Checks if the argument is a valid twinning parameters structure.}
\usage{
check_twinning_params(twinning_params)
}
\arguments{
\item{twinning_params}{can be \code{NA} if no twinning is desired,
or can be the twinning parameters,
as can be created by \link{create_twinning_params}}
}
\value{
nothing. Will \link{stop} if nit
}
\description{
Will \link{stop} if not.
A valid twinning parameters structure can be created
by \link{create_twinning_params}.
}
\examples{
if (beautier::is_on_ci()) {
  check_twinning_params(create_twinning_params())
}
}
\author{
Richèl J.C. Bilderbeek
}
