% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_brts_and_topology.R
\name{combine_brts_and_topology}
\alias{combine_brts_and_topology}
\title{Substitute branching times keeping the topology}
\usage{
combine_brts_and_topology(brts, tree)
}
\arguments{
\item{brts}{set of branching times}

\item{tree}{an ultrametric phylogenetic tree of class \link[ape]{phylo}}
}
\value{
a tree
}
\description{
Given a tree topology and a set of branching times,
it combines them together. They have to be compatible.
}
\author{
Giovanni Laudanno, David Bapst
}
