% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_pir_out_to_long.R
\name{convert_pir_out_to_long}
\alias{convert_pir_out_to_long}
\title{Convert a \code{pir_out} to its long form}
\usage{
convert_pir_out_to_long(pir_out, verbose = FALSE)
}
\arguments{
\item{pir_out}{the output of \link{pir_run}}

\item{verbose}{if TRUE, show more output}
}
\value{
the \code{pir_out} in long form
}
\description{
Convert a \code{pir_out} to its long form
}
\examples{
library(testthat)

pir_out <- create_test_pir_run_output()
t <- convert_pir_out_to_long(pir_out)
expect_equal(2, ncol(t))
expect_true("error_value" \%in\% names(t))
expect_true("tree_and_model" \%in\% names(t))
expect_silent(pir_plot_from_long(t))
}
\author{
Richèl J.C. Bilderbeek, Giovanni Laudanno
}
