% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_n_mutations.R
\name{count_n_mutations}
\alias{count_n_mutations}
\title{Count mutations}
\usage{
count_n_mutations(alignment, root_sequence)
}
\arguments{
\item{alignment}{a DNA alignment}

\item{root_sequence}{the DNA sequence at the root of the phylogeny.
By default, this will consist out of an equal amount of each letter}
}
\value{
the number of mutations
}
\description{
Count mutations
}
\examples{
library(testthat)

# Create an AAAA alignent
alignment <- ape::as.DNAbin(
  x = list(species_1 = strsplit("aaaa", split = "")[[1]])
)

# Count the number of mutations from AAAA
expect_equal(count_n_mutations(alignment, "aaaa"), 0)
expect_equal(count_n_mutations(alignment, "acaa"), 1)
expect_equal(count_n_mutations(alignment, "acga"), 2)
expect_equal(count_n_mutations(alignment, "acgt"), 3)
expect_equal(count_n_mutations(alignment, "ccgt"), 4)

}
\author{
Giovanni Laudanno, Richèl J.C. Bilderbeek
}
