% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_alignment_params.R
\name{create_alignment_params}
\alias{create_alignment_params}
\title{Create the parameters for the alignment simulation}
\usage{
create_alignment_params(root_sequence, mutation_rate)
}
\arguments{
\item{root_sequence}{the DNA sequence at the root of the phylogeny.
By default, this will consist out of only adenine}

\item{mutation_rate}{the mutation rate per base pair per time unit}
}
\value{
a list of alignment parameters
}
\description{
Create the parameters for the alignment simulation
}
\author{
Richel J.C. Bilderbeek
}
