% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_all_experiments.R
\name{create_all_experiments}
\alias{create_all_experiments}
\title{Create all \link{pirouette} experiments.}
\usage{
create_all_experiments()
}
\value{
all \link{pirouette} experiments.
}
\description{
Create all \link{pirouette} experiments.
}
\author{
Richel J.C. Bilderbeek, Giovanni Laudanno
}
