% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_bd_tree.R
\name{create_bd_tree}
\alias{create_bd_tree}
\title{Create an BD twin tree from an MBD tree
and save it as a file}
\usage{
create_bd_tree(phylogeny, seed)
}
\arguments{
\item{phylogeny}{a phylogeny of class \link[ape]{phylo}}

\item{seed}{a random number generator seed}
}
\value{
a twin BD tree of class \code{phylo},
  obtained from the corresponding MBD tree.
}
\description{
Create an BD twin tree from an MBD tree
and save it as a file
}
\author{
Richel J.C. Bilderbeek, Giovanni Laudanno
}
