% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_blocked_dna.R
\name{create_blocked_dna}
\alias{create_blocked_dna}
\title{Create a 'blocked' DNA sequence, which is a sequence
with four equal-sized nucleotide sections}
\usage{
create_blocked_dna(length)
}
\arguments{
\item{length}{number of nucleotides. Must be a multitude of four.}
}
\description{
Create a 'blocked' DNA sequence, which is a sequence
with four equal-sized nucleotide sections
}
\examples{
  testthat::expect_equal(create_blocked_dna(n = 4), "acgt")
  testthat::expect_equal(create_blocked_dna(n = 8), "aaccggtt")
}
\author{
Richel J.C. Bilderbeek
}
