% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_copy_twin_tree_from_true_function.R
\name{create_copy_twin_tree_from_true_function}
\alias{create_copy_twin_tree_from_true_function}
\title{Create a function that can simulate the twin tree from the true tree,
by just copying the true tree}
\usage{
create_copy_twin_tree_from_true_function()
}
\value{
a function
}
\description{
Create a function that can simulate the twin tree from the true tree,
by just copying the true tree
}
\examples{
library(testthat)

f <- create_copy_twin_tree_from_true_function()
phylo_in <- ape::read.tree(text = "((A:1, B:1):1, C:2);")
phylo_out <- f(phylo_in)
expect_equal(
  ape::branching.times(phylo_in),
  ape::branching.times(phylo_out)
)
}
\seealso{
Use \link{create_sim_yule_twin_tree_function} to get a function to
produce a Yule tree.
Use \link{create_sim_bd_twin_tree_function} to get a function to
produce a Birth-Death tree.
}
\author{
Richèl J.C. Bilderbeek
}
