% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_generative_trees.R
\name{create_dd_tree}
\alias{create_dd_tree}
\title{Create an exemplary diversity-dependent (DD) birth-death tree.}
\usage{
create_dd_tree(
  n_taxa = 6,
  crown_age = 10,
  extinction_rate = 0.1,
  best_of_n_trees = 100
)
}
\arguments{
\item{n_taxa}{number of tree tips}

\item{crown_age}{the fixed crown age of the posterior. Set to NA
to let it be estimated}

\item{extinction_rate}{per-species extinction rate}

\item{best_of_n_trees}{simulate this number of DD trees with
the desired number of taxa,}
}
\description{
Or: create a DD tree with a strong DD effect.
}
\details{
This algorithm does so, by simulating \code{best_of_n_trees}
trees, then pick the first tree that has a gamma statistic
within the lowest 5 percent of all gammas. Trees with such
a low gamma, have the strongest DD effect, as these
deviate strongest from the expected exponential growth
that regular birth-death (BD) trees have.
}
\examples{
library(testthat)

n_taxa <- 31
crown_age <- 41

phylogeny <- create_dd_tree(
  n_taxa = n_taxa,
  crown_age = crown_age
)

expect_equal(n_taxa, ape::Ntip(phylogeny))
expect_equal(crown_age, beautier::get_crown_age(phylogeny))
}
\author{
Giovanni Laudanno, Richèl J.C. Bilderbeek
}
