% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_error_measure_params.R
\name{create_error_measure_params}
\alias{create_error_measure_params}
\title{Create the parameters to specify hwo the error
between teh given phylogeny and the Bayesian posterior trees is measured.}
\usage{
create_error_measure_params(burn_in_fraction = 0.1,
  error_function = get_nltt_error_function())
}
\arguments{
\item{rng_seed}{the random number generator seed as used in the
simulation of a twin tree}
}
\value{
a error_measure parameter set
}
\description{
Create the parameters to specify hwo the error
between teh given phylogeny and the Bayesian posterior trees is measured.
}
\examples{
 error_measure_params <- create_error_measure_params()
 testit::assert("rng_seed" \%in\% names(error_measure_params))
 testit::assert("twin_tree_filename" \%in\% names(error_measure_params))
}
\author{
Richel J.C. Bilderbeek
}
