% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_experiment.R
\name{create_experiment}
\alias{create_experiment}
\title{Create a valid \link{pirouette} experiment.}
\usage{
create_experiment(model_type = "generative", run_if = "always",
  do_measure_evidence = FALSE,
  inference_model = beautier::create_inference_model(),
  beast2_options = beastier::create_beast2_options())
}
\arguments{
\item{model_type}{type of inference model supplied for an experiment.
Possible values:
\itemize{
  \item \code{generative}: the inference model is (or is assumed to be)
    the inference model underlying the phylogeny
  \item \code{candidate}: the inference model is a candidate model,
    that competes with other models for having the most
    evidence (aka highest marginal likelihood)
}}

\item{run_if}{the condition for an experiment's inference model to be run.
Possible values:
\itemize{
  \item \code{always}: always
  \item \code{best_candidate}: if the inference model is the
    candidate model with the most evidence (aka highest marginal
    likelihood)
}}

\item{do_measure_evidence}{boolean to indicate if the
evidence (aka marginal likelihood) of an experiment must be
measured}

\item{inference_model}{an inference model, which is a combination
of site model, clock model, tree prior and BEAST2 input and
input filenames.}

\item{beast2_options}{BEAST2 options,
as can be created by \link[beastier]{create_beast2_options}}
}
\value{
a \link{pirouette} experiment.
}
\description{
The arguments are checked by \link{check_experiment}.
}
\examples{
 experiment <- create_experiment()
 testthat::expect_silent(check_experiment(create_experiment()))
 testit::assert("model_type" \%in\% names(experiment))
 testit::assert("run_if" \%in\% names(experiment))
 testit::assert("do_measure_evidence" \%in\% names(experiment))
 testit::assert("inference_model" \%in\% names(experiment))
 testit::assert("beast2_options" \%in\% names(experiment))
}
\author{
Richel J.C. Bilderbeek
}
