% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_experiment.R
\name{create_experiment}
\alias{create_experiment}
\title{Create a valid \link{pirouette} experiment.}
\usage{
create_experiment(model_type = "generative", run_if = "always",
  do_measure_evidence = FALSE,
  inference_model = beautier::create_inference_model(mcmc =
  create_mcmc(store_every = 1000)),
  beast2_options = beastier::create_beast2_options(input_filename =
  tempfile(pattern = "beast2_", fileext = ".xml"), output_log_filename =
  tempfile(pattern = "beast2_", fileext = ".log"), output_trees_filenames =
  tempfile(pattern = "beast2_", fileext = "trees"), output_state_filename =
  tempfile(pattern = "beast2_", fileext = ".state.xml")),
  est_evidence_mcmc = beautier::create_nested_sampling_mcmc(),
  beast2_bin_path = beastier::get_default_beast2_bin_path(),
  errors_filename = tempfile(pattern = "errors_", fileext = ".csv"))
}
\arguments{
\item{model_type}{type of inference model supplied for an experiment.
Possible values:
\itemize{
  \item \code{generative}: the inference model is (or is assumed to be)
    the inference model underlying the phylogeny
  \item \code{candidate}: the inference model is a candidate model,
    that competes with other models for having the most
    evidence (aka highest marginal likelihood)
}}

\item{run_if}{the condition for an experiment's inference model to be run.
Possible values:
\itemize{
  \item \code{always}: always
  \item \code{best_candidate}: if the inference model is the
    candidate model with the most evidence (aka highest marginal
    likelihood)
}}

\item{do_measure_evidence}{boolean to indicate if the
evidence (aka marginal likelihood) of an experiment must be
measured}

\item{inference_model}{an inference model, which is a combination
of site model, clock model, tree prior and BEAST2 input and
input filenames.}

\item{beast2_options}{BEAST2 options,
as can be created by \link[beastier]{create_beast2_options}}

\item{est_evidence_mcmc}{MCMC used in the estimation of
the evidence (aka marginal likelihood).
The MCMC must be a Nested Sampling MCMC,
as can be created by \link[beautier]{create_nested_sampling_mcmc}.}

\item{beast2_bin_path}{path to BEAST2 binary file. The use of the
binary BEAST2 file is required for estimation of the evidence (aka
marginal likelihood). The default BEAST2 binary path can be
obtained using \link[beastier]{get_default_beast2_bin_path}}

\item{errors_filename}{baseline name for errors filenames}
}
\value{
a \link{pirouette} experiment.
}
\description{
The arguments are checked by \link{check_experiment}.
}
\examples{
 experiment <- create_experiment()
 testthat::expect_silent(check_experiment(create_experiment()))
 testit::assert("model_type" \%in\% names(experiment))
 testit::assert("run_if" \%in\% names(experiment))
 testit::assert("do_measure_evidence" \%in\% names(experiment))
 testit::assert("inference_model" \%in\% names(experiment))
 testit::assert("beast2_options" \%in\% names(experiment))
 testit::assert("est_evidence_mcmc" \%in\% names(experiment))
 testit::assert("beast2_bin_path" \%in\% names(experiment))
}
\author{
Richel J.C. Bilderbeek, Giovanni Laudanno
}
