% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_experiment.R
\name{create_experiment}
\alias{create_experiment}
\title{Create a \link{pirouette} experiment.}
\usage{
create_experiment(rng_seed = 0, twin_model = "bd",
  twin_tree_filename = tempfile(fileext = ".newick"),
  twin_alignment_filename = tempfile(fileext = ".fasta"))
}
\arguments{
\item{twin_model}{the model you want to use to generate the twin tree}

\item{twin_tree_filename}{name of the (\code{.newick}) file the twin
tree will be saved to}

\item{twin_alignment_filename}{name of the FASTA file the twin
alignment will be saved to}
}
\value{
a \link{pirouette} experiment.
}
\description{
Create a \link{pirouette} experiment.
}
\examples{
 experiment <- create_experiment()
 testit::assert("rng_seed" \%in\% names(experiment))
 testit::assert("twin_tree_filename" \%in\% names(experiment))
}
\author{
Richel J.C. Bilderbeek
}
