% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_experiment.R
\name{create_experiment}
\alias{create_experiment}
\title{Create a valid \link{pirouette} experiment.}
\usage{
create_experiment(
  inference_conditions = create_inference_conditions(),
  inference_model = beautier::create_inference_model(mcmc =
    beautier::create_mcmc(store_every = 1000)),
  beast2_options = beastier::create_beast2_options(input_filename =
    beastier::create_temp_input_filename(), output_state_filename =
    beastier::create_temp_state_filename()),
  est_evidence_mcmc = beautier::create_ns_mcmc(epsilon = 1e-12),
  beast2_bin_path = beastier::get_default_beast2_bin_path(),
  errors_filename = pirouette::get_temp_errors_filename()
)
}
\arguments{
\item{inference_conditions}{conditions under which the inference model
is used in the inference}

\item{inference_model}{an inference model, which is a combination
of site model, clock model, tree prior and BEAST2 input and
input filenames.}

\item{beast2_options}{BEAST2 options,
as can be created by \link[beastier]{create_beast2_options}}

\item{est_evidence_mcmc}{MCMC used in the estimation of
the evidence (aka marginal likelihood).
The MCMC must be a Nested Sampling MCMC,
as can be created by \link[beautier]{create_ns_mcmc}.}

\item{beast2_bin_path}{path to BEAST2 binary file. The use of the
binary BEAST2 file is required for estimation of the evidence (aka
marginal likelihood). The default BEAST2 binary path can be
obtained using \link[beastier]{get_default_beast2_bin_path}}

\item{errors_filename}{baseline name for errors filenames,
as created by \link{get_temp_errors_filename}}
}
\value{
a \link{pirouette} experiment.
}
\description{
The arguments are checked by \link{check_experiment}.
}
\examples{
if (beautier::is_on_ci()) {
  create_experiment()
}
}
\author{
Richèl J.C. Bilderbeek, Giovanni Laudanno
}
