% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_fig_bd.R
\name{create_fig_bd}
\alias{create_fig_bd}
\title{Create figure bd for pirouette article}
\usage{
create_fig_bd(phylogeny,
  twinning_params = create_twinning_params(twin_model = "bd", method =
  "random_tree"))
}
\arguments{
\item{phylogeny}{a phylogeny of class \link[ape]{phylo}}

\item{twinning_params}{can be \code{NA} if no twinning is desired,
or can be the twinning parameters,
as can be created by \link{create_twinning_params}}
}
\value{
a ggplot2 plot
}
\description{
Create figure bd for pirouette article
}
\author{
Richel J.C. Bilderbeek, Giovanni Laudanno
}
