% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ideal_tree.R
\name{create_ideal_tree}
\alias{create_ideal_tree}
\title{Create an "ideal" tree}
\usage{
create_ideal_tree(n_taxa, crown_age, n_0 = 2,
  ideal_method = "stunning")
}
\arguments{
\item{n_taxa}{number of tree tips}

\item{crown_age}{the fixed crown age of the posterior. Set to NA
to let it be estimated}

\item{n_0}{number of starting species}

\item{ideal_method}{method to generate the "ideal" tree}
}
\description{
Create an "ideal" tree
}
\author{
Richel J.C. Bilderbeek, Giovanni Laudanno
}
