% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_inference_params.R
\name{create_inference_params}
\alias{create_inference_params}
\title{Creates the parameters all BEAST2 runs share.}
\usage{
create_inference_params(mrca_prior = NA,
  mcmc = beautier::create_mcmc(), rng_seed = NA,
  beast2_path = beastier::get_default_beast2_path(), verbose = FALSE)
}
\arguments{
\item{mrca_prior}{an MRCA prior,
as created by \link[beautier]{create_mrca_prior}}

\item{mcmc}{MCMC options, as created by \link[beautier]{create_mcmc}}

\item{rng_seed}{the random number generator seed
used in the Bayesian inference.
The Bayesian inference is handled by the \link[babette]{babette}
R package, that calls the phylogenetic tool \code{BEAST2}.}

\item{beast2_path}{Path to the
BEAST2 jar file (\code{beast.jar})
or BEAST2 binary file '(\code{beast})'.
Use \link[beastier]{get_default_beast2_jar_path} for the default
BEAST2 jar file path.
Use \link[beastier]{get_default_beast2_bin_path} for the default
BEAST2 binary file path.}

\item{verbose}{if TRUE, show more output}
}
\description{
These are for example, the MRCA prior, the MCMC, the RNG seed
and the path to BEAST2. The parts of a BEAST2 run that
is unique, which is a combination of a site model, clock model
and tree prior, is created by \link{create_inference_model})
}
\author{
Richel J.C. Bilderbeek
}
