% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pir_params.R
\name{create_pir_params}
\alias{create_pir_params}
\title{Create the parameters for pirouette}
\usage{
create_pir_params(alignment_params, twinning_params = NA,
  model_select_params = as.list(seq(1, 314)),
  inference_params = create_inference_params(),
  experiments = list(create_experiment()),
  error_measure_params = create_error_measure_params(),
  evidence_epsilon = 1e-12, evidence_filename = tempfile(fileext =
  ".csv"))
}
\arguments{
\item{alignment_params}{parameters to simulate an alignment,
as can be created by \link{create_alignment_params}}

\item{twinning_params}{can be \code{NA} if no twinning is desired,
or can be the twinning parameters,
as can be created by \link{create_twinning_params}}

\item{model_select_params}{a list of one or more parameter sets
to select an inference model
(a combination of site model, clock model and tree prior),
as can be created by \link{create_model_select_param}}

\item{inference_params}{one set of parameters for the Bayesian inference,
as can be created by \link{create_inference_params}.
The Bayesian inference is handled by the \link[babette]{babette}
R package, that calls the phylogenetic tool \code{BEAST2}.}

\item{experiments}{a list of one or more \link{pirouette} experiments,
as can be created by \link{create_experiment}}

\item{error_measure_params}{parameter set to specify how the
error between the given phylogeny and the Bayesian
posterior is determined.
Use \link{create_error_measure_params} to create such
a parameter set}

\item{evidence_epsilon}{relative error in estimating the
evidence (aka marginal likelihood).
TODO: put in \code{misc_params},}

\item{evidence_filename}{filename to store the estimated
evidences (aka marginal likelihoods)}
}
\value{
a list with all pirouette parameters
}
\description{
Create the parameters for pirouette
}
\author{
Giovanni Laudanno
}
