% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pir_run_test_output.R
\name{create_pir_run_test_output}
\alias{create_pir_run_test_output}
\title{Create the same output of \link{pir_run}
to be used for testing}
\usage{
create_pir_run_test_output()
}
\description{
Create the same output of \link{pir_run}
to be used for testing
}
\author{
Richel J.C. Bilderbeek
}
