% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_standard_mutation_rate.R
\name{create_standard_mutation_rate}
\alias{create_standard_mutation_rate}
\title{Create the parameters for the mutation rate}
\usage{
create_standard_mutation_rate(crown_age)
}
\arguments{
\item{crown_age}{the fixed crown age of the posterior. Set to NA
to let it be estimated}
}
\value{
the mutation rate
}
\description{
Create the parameters for the mutation rate
}
\author{
Giovanni Laudanno, Richel J.C. Bilderbeek
}
