% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_alignment_params.R
\name{create_test_alignment_params}
\alias{create_test_alignment_params}
\title{Create testing parameters for the alignment simulation.}
\usage{

  create_test_alignment_params(sim_true_alignment_function = pirouette::sim_true_alignment_with_standard_site_model,
  root_sequence = "acgt",
  mutation_rate = pirouette::create_standard_mutation_rate,
  site_model = beautier::create_jc69_site_model(),
  clock_model = beautier::create_strict_clock_model(), rng_seed = 0,
  fasta_filename = tempfile(pattern = "alignment_", fileext = ".fasta"))
}
\arguments{
\item{sim_true_alignment_function}{function to simulate a
true alignment with.
This function must have two arguments,
called \code{true_phylogeny} (which will hold the true phylogeny)
and \code{root_sequence} (which holds the DNA root sequence).
The return type must be \link[ape]{DNAbin}.

Use \link{check_sim_true_alignment_function} to verify if the function
has the right signature and output.

Some standard functions:\cr
\itemize{
  \item Use \link{get_sim_true_alignment_with_standard_site_model_function}
  to get a function (\link{sim_true_alignment_with_standard_site_model})
  the use a standard site model.
  \item Use
  \link{get_sim_true_alignment_with_linked_node_sub_site_model_function}
  to get a function (\link{sim_true_alignment_with_linked_node_sub_site_model})
  the use a linked node substitution site model.
  \item Use
  \link{get_sim_true_alignment_with_unlinked_node_sub_site_model_function}
  to get a function (\link{sim_true_alignment_with_unlinked_node_sub_site_model})
  the use an unlinked node substitution site model.
}}

\item{root_sequence}{the DNA sequence at the root of the phylogeny.
By default, this will consist out of an equal amount of each letter
Use \link{check_root_sequence} to check if a root sequence is valid.}

\item{mutation_rate}{the mutation rate per base pair per time unit.
Use \link{check_mutation_rate} to check if a mutation rate is valid.}

\item{site_model}{a nucleotide substitution model,
which can be:
\itemize{
  \item{
    A standard nucloetide substitution model,
    as created by \link[beautier]{create_site_model}
  }
  \item{
    \code{linked_node_sub}: a linked node-substitution model
  }
  \item{
    \code{unlinked_node_sub}: an unlinked node-substitution model
  }
}}

\item{clock_model}{a clock model,
as created by \link[beautier]{create_clock_model}}

\item{rng_seed}{a random number generator seed}

\item{fasta_filename}{name of a FASTA file}
}
\value{
a list of alignment parameters
library(testthat)

alignment_params <- create_test_alignment_params()
expect_silent(check_alignment_params(alignment_params))
}
\description{
Create testing parameters for the alignment simulation.
}
\author{
Richèl J.C. Bilderbeek
}
