% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_test_cand_experiment.R
\name{create_test_cand_experiment}
\alias{create_test_cand_experiment}
\title{Create a valid testing \link{pirouette} candidate experiment.}
\usage{

  create_test_cand_experiment(inference_conditions = create_inf_conds(model_type
  = "candidate", run_if = "best_candidate", do_measure_evidence = TRUE),
  inference_model = beautier::create_inference_model(mcmc =
  beautier::create_mcmc(chain_length = 2000, store_every = 1000)),
  beast2_options = beastier::create_beast2_options(input_filename =
  beastier::create_temp_input_filename(), output_state_filename =
  beastier::create_temp_output_state_filename()),
  est_evidence_mcmc = beautier::create_nested_sampling_mcmc(chain_length
  = 2000, store_every = 1000, epsilon = 100))
}
\arguments{
\item{inference_conditions}{conditions under which the inference model
is used in the inference}

\item{inference_model}{an inference model, which is a combination
of site model, clock model, tree prior and BEAST2 input and
input filenames.}

\item{beast2_options}{BEAST2 options,
as can be created by \link[beastier]{create_beast2_options}}

\item{est_evidence_mcmc}{MCMC used in the estimation of
the evidence (aka marginal likelihood).
The MCMC must be a Nested Sampling MCMC,
as can be created by \link[beautier]{create_nested_sampling_mcmc}.}
}
\value{
a \link{pirouette} experiment.
}
\description{
Create a valid testing \link{pirouette} candidate experiment.
}
\examples{
library(testthat)

if (rappdirs::app_dir()$os != "win") {
  experiment <- create_test_cand_experiment()

  expect_true("inference_conditions" \%in\% names(experiment))
  expect_true("inference_model" \%in\% names(experiment))
  expect_true("beast2_options" \%in\% names(experiment))
  expect_true("est_evidence_mcmc" \%in\% names(experiment))
  expect_true("beast2_bin_path" \%in\% names(experiment))
  expect_silent(check_experiment(experiment))
}
}
\author{
Richèl J.C. Bilderbeek
}
