% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_experiment.R
\name{create_test_gen_experiment}
\alias{create_test_gen_experiment}
\title{Create a valid testing \link{pirouette} generative experiment.}
\usage{
create_test_gen_experiment()
}
\value{
a \link{pirouette} experiment.
}
\description{
Create a valid testing \link{pirouette} generative experiment.
}
\examples{
 library(testthat)

 experiment <- create_cand_test_gen_experiment()

 expect_true("inference_conditions" \%in\% names(experiment))
 expect_true("inference_model" \%in\% names(experiment))
 expect_true("beast2_options" \%in\% names(experiment))
 expect_true("est_evidence_mcmc" \%in\% names(experiment))
 expect_true("beast2_bin_path" \%in\% names(experiment))

 expect_silent(check_experiment(experiment))
}
\author{
Richèl J.C. Bilderbeek
}
