% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_test_marg_liks.R
\name{create_test_marg_liks}
\alias{create_test_marg_liks}
\title{Create a testing marginal likelihoods data frame.}
\usage{
create_test_marg_liks(site_models = beautier::create_site_models(),
  clock_models = beautier::create_clock_models(),
  tree_priors = beautier::create_tree_priors())
}
\arguments{
\item{site_models}{a list of one or more site models,
as created by \link[beautier]{create_site_model}}

\item{clock_models}{a list of one or more clock models,
as created by \link[beautier]{create_clock_model}}

\item{tree_priors}{a list of one or more tree priors,
as created by \link[beautier]{create_tree_prior}}
}
\value{
a data frame
}
\description{
Create a testing marginal likelihoods data frame.
}
\examples{
  df <- create_test_marg_liks()

  testthat::expect_true("site_model_name" \%in\% names(df))
  testthat::expect_true("clock_model_name" \%in\% names(df))
  testthat::expect_true("tree_prior_name" \%in\% names(df))
  testthat::expect_true("marg_log_lik" \%in\% names(df))
  testthat::expect_true("marg_log_lik_sd" \%in\% names(df))
  testthat::expect_true("weight" \%in\% names(df))

  # Log likelihoods are zero or less
  testthat::expect_true(all(df$marg_log_lik <= 0.0))

  # Sum of all weights is one hundred percent
  testthat::expect_equal(sum(df$weight), 1.0)
}
\author{
Richel J.C. Bilderbeek
}
