% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_test_marg_liks.R
\name{create_test_marg_liks}
\alias{create_test_marg_liks}
\title{Create a testing marginal likelihoods data frame.}
\usage{
create_test_marg_liks(site_models = beautier::create_site_models(),
  clock_models = beautier::create_clock_models(),
  tree_priors = beautier::create_tree_priors())
}
\description{
Create a testing marginal likelihoods data frame.
}
\author{
Richel J.C. Bilderbeek
}
