% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pir_params.R
\name{create_test_pir_params}
\alias{create_test_pir_params}
\title{Create a set of testing parameters for \code{pirouette}}
\usage{

  create_test_pir_params(alignment_params = create_test_alignment_params(),
  twinning_params = NA, experiments = list(create_test_experiment()),
  error_measure_params = create_error_measure_params(),
  evidence_filename = tempfile(pattern = "evidence_", fileext = ".csv"),
  verbose = FALSE)
}
\arguments{
\item{alignment_params}{parameters to simulate an alignment,
as can be created by \link{create_alignment_params}}

\item{twinning_params}{can be \code{NA} if no twinning is desired,
or can be the twinning parameters,
as can be created by \link{create_twinning_params}}

\item{experiments}{a list of one or more \link{pirouette} experiments,
as can be created by \link{create_experiment}}

\item{error_measure_params}{parameter set to specify how the
error between the given phylogeny and the Bayesian
posterior is determined.
Use \link{create_error_measure_params} to create such
a parameter set}

\item{evidence_filename}{filename to store the estimated
evidences (aka marginal likelihoods)}

\item{verbose}{if TRUE, show more output}
}
\value{
a list with all \code{pirouette} parameters
}
\description{
Create a set of testing parameters for \code{pirouette}
}
\examples{
  pir_params <- create_test_pir_params()
  check_pir_params(pir_params)
}
\author{
Richel J.C. Bilderbeek
}
