% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_test_pir_params_setup.R
\name{create_test_pir_params_setup}
\alias{create_test_pir_params_setup}
\title{Create a \code{pir_params} that follows a specific setup}
\usage{
create_test_pir_params_setup(has_candidate = FALSE, has_twinning = FALSE)
}
\arguments{
\item{has_candidate}{TRUE to have a candidate experiment}

\item{has_twinning}{TRUE to use twinning}
}
\description{
Create a \code{pir_params} that follows a specific setup
}
\examples{
library(testthat)

# Minimal use
expect_silent(
  check_pir_params(create_test_pir_params_setup())
)

# Generative experiment only, without twinning
pir_params <- create_test_pir_params_setup(
  has_candidate = FALSE,
  has_twinning = FALSE
)
expect_true(!has_twinning(pir_params))
expect_true(!has_candidate_experiments(pir_params))

# Generative and candidate experiment, without twinning
pir_params <- create_test_pir_params_setup(
  has_candidate = TRUE,
  has_twinning = FALSE
)
expect_true(!has_twinning(pir_params))
expect_true(has_candidate_experiments(pir_params))

# Generative experiment only, with twinning
pir_params <- create_test_pir_params_setup(
  has_candidate = FALSE,
  has_twinning = TRUE
)
expect_true(has_twinning(pir_params))
expect_true(!has_candidate_experiments(pir_params))

# Generative and candidate experiment, with twinning
pir_params <- create_test_pir_params_setup(
  has_candidate = TRUE,
  has_twinning = TRUE
)
expect_true(has_twinning(pir_params))
expect_true(has_candidate_experiments(pir_params))
}
