% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_test_pir_run_output.R
\name{create_test_pir_run_output}
\alias{create_test_pir_run_output}
\title{Create the same output of \link{pir_run}
to be used for testing}
\usage{
create_test_pir_run_output(add_twin = FALSE, add_best = FALSE)
}
\arguments{
\item{add_twin}{add rows for twin tree}

\item{add_best}{add rows for best inference model}
}
\value{
a data frame with errors, with as many rows as model selection
parameter sets. The output can be checked using \link{check_pir_out}.
}
\description{
Create the same output of \link{pir_run}
to be used for testing
}
\examples{
# Check cleanup by other functions
beastier::check_empty_beaustier_folders()

check_pir_out(
  create_test_pir_run_output(
    add_twin = TRUE,
    add_best = TRUE
  )
)

pir_plot(
  create_test_pir_run_output(
    add_twin = TRUE,
    add_best = TRUE
  )
)

# Check cleanup by other functions
beastier::check_empty_beaustier_folders()
}
\author{
Richèl J.C. Bilderbeek
}
