% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_test_run_experiment.R
\name{create_test_run_experiment}
\alias{create_test_run_experiment}
\title{Create a run experiment for testing}
\usage{
create_test_run_experiment(true_result = create_test_result(),
  twin_result = create_test_result())
}
\arguments{
\item{true_result}{result obtained from using the true tree}

\item{twin_result}{result obtained from using the twin tree}
}
\value{
a run experiment
}
\description{
A run experiment is an extension of a normal experiment (that
has not been run), as it has added:
\itemize{
  \item true_result the results for the true tree
  \item twin_result the results for the twin tree
}
In both cases,
}
\examples{
  testthat::expect_silent(
    check_run_experiment(create_test_run_experiment())
  )
}
\author{
Richel J.C. Bilderbeek
}
