% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_twin_alignment_file.R
\name{create_twin_alignment_file}
\alias{create_twin_alignment_file}
\title{Simulates a twin DNA alignment and saves it to a FASTA file.}
\usage{
create_twin_alignment_file(twin_phylogeny, alignment_params,
  twinning_params, verbose = FALSE)
}
\arguments{
\item{twin_phylogeny}{a phylogeny of class \link[ape]{phylo}}

\item{alignment_params}{parameters to simulate an alignment,
as can be created by \link{create_alignment_params}}

\item{twinning_params}{can be \code{NA} if no twinning is desired,
or can be the twinning parameters,
as can be created by \link{create_twinning_params}}

\item{verbose}{if TRUE, show more output}
}
\value{
nothing
}
\description{
The simulation is performed by \link{create_twin_alignment}.
}
\seealso{
Use \link{create_twin_alignment} to only simulate the twin alignment,
  without saving the alignment to file
}
\author{
Richèl J.C. Bilderbeek
}
