% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_twin_tree.R
\name{create_twin_tree}
\alias{create_twin_tree}
\title{Create a twin tree}
\usage{
create_twin_tree(phylogeny, seed)
}
\arguments{
\item{phylogeny}{a phylogeny of class \link[ape]{phylo}}

\item{seed}{a random number generator seed}
}
\description{
Create a twin tree
}
\author{
Richel J.C. Bilderbeek
}
