% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_twin_tree.R
\name{create_twin_tree}
\alias{create_twin_tree}
\title{Create a twin tree}
\usage{
create_twin_tree(phylogeny, twin_model = "bd", method)
}
\arguments{
\item{phylogeny}{a phylogeny of class \link[ape]{phylo}}

\item{twin_model}{the model you want to use to generate the twin tree}
}
\description{
Create a twin tree
}
\author{
Richel J.C. Bilderbeek, Giovanni Laudanno
}
