% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_twin_tree.R
\name{create_twin_tree}
\alias{create_twin_tree}
\title{Create a twin tree}
\usage{
create_twin_tree(phylogeny, twinning_params = create_twinning_params())
}
\arguments{
\item{phylogeny}{a phylogeny of class \link[ape]{phylo}}

\item{twinning_params}{can be \code{NA} if no twinning is desired,
or can be the twinning parameters,
as can be created by \link{create_twinning_params}}
}
\description{
Create a twin tree
}
\author{
Richel J.C. Bilderbeek, Giovanni Laudanno
}
