% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_twinning_params.R
\name{create_twinning_params}
\alias{create_twinning_params}
\title{Create the parameters for the twinning simulation.
The site model and clock models will be used and
their combination will be called the generative model
of the twinning}
\usage{
create_twinning_params(rng_seed = 0,
  twin_tree_filename = tempfile(fileext = ".newick"))
}
\arguments{
\item{rng_seed}{the random number generator seed as used in the
simulation of a twin tree}
}
\value{
a twinning parameter set
}
\description{
Create the parameters for the twinning simulation.
The site model and clock models will be used and
their combination will be called the generative model
of the twinning
}
\examples{
 twinning_params <- create_twinning_params()
 testit::assert("rng_seed" \%in\% names(twinning_params))
 testit::assert("twin_tree_filename" \%in\% names(twinning_params))
}
\author{
Richel J.C. Bilderbeek
}
