% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_yule_tree.R
\name{create_yule_tree}
\alias{create_yule_tree}
\title{Create a twin tree from a phylogeny using a Yule process}
\usage{
create_yule_tree(phylogeny, seed)
}
\arguments{
\item{phylogeny}{a phylogeny of class \link[ape]{phylo}}

\item{seed}{a random number generator seed}
}
\value{
a twin Yule tree of class \code{phylo},
}
\description{
Create a twin tree from a phylogeny using a Yule process
}
\author{
Richel J.C. Bilderbeek, Giovanni Laudanno
}
