% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_params_doc.R
\name{default_params_doc}
\alias{default_params_doc}
\title{This function does nothing. It is intended to inherit is parameters'
documentation.}
\usage{
default_params_doc(alignment, alignment_params, alignment_rng_seed,
  bd_mutation_rate, bd_tree, bd_tree_filename, beast2_input_filename,
  beast2_output_log_filename, beast2_output_state_filename,
  beast2_output_trees_filenames, beast2_path, beast2_rng_seed, brts,
  chain_length, clock_model, clock_models, clock_model_name, crown_age,
  epsilon, fasta_filename, filename, folder_name, inference_param,
  init_speciation_rate, init_extinction_rate, lambda, marg_liks,
  mbd_l_matrix, mbd_mutation_rate, mbd_tree, mcmc, model_selection,
  model_select_param, model_select_params, mrca_prior, mu, mutation_rate,
  nu, nu_events, parameter_filename, parameters_filename, phylo, phylogeny,
  posterior_trees, precision, project_folder_name, root_sequence,
  sample_interval, seed, sequence_length, sim_pars, sim_phylo, site_model,
  site_models, site_model_name, sub_chain_length, tree, tree_filename,
  tree_prior, tree_priors, tree_prior_name, trees_filename,
  twinning_params, type, verbose)
}
\arguments{
\item{alignment}{a DNA alignment}

\item{alignment_params}{parameters to simulate an alignment,
as can be created by \link{create_alignment_params}}

\item{alignment_rng_seed}{The random number generator seed used
to generate an alignment}

\item{bd_mutation_rate}{the mutation rate when creating an alignment
from a BD tree}

\item{bd_tree}{a phylogent of class \link[ape]{phylo},
created by a Birth Death process}

\item{bd_tree_filename}{name of the file that stores a BD twin tree}

\item{beast2_input_filename}{path of the BEAST2 configuration file.
By default, this file is put in a temporary folder with a random filename,
as the user needs not read it: it is used as input of BEAST2.
Specifying a \code{beast2_input_filename} allows
to store that file in a more permanently stored location.}

\item{beast2_output_log_filename}{name of the log file created by BEAST2,
containing the parameter estimates in time.
By default, this file is put a temporary folder with a random filename,
as the user needs not read it.
Specifying a beast2_output_log_filename allows to store that file
in a more permanently stored location.}

\item{beast2_output_state_filename}{name of the final state file
created by BEAST2, containing the operator acceptances.
By default, this file is put a temporary folder with a random filename,
as the user needs not read it.
Specifying a beast2_output_state_filename allows to store
that file in a more permanently stored location.}

\item{beast2_output_trees_filenames}{name of the one or more trees files
created by BEAST2, one per alignment.
By default, these files are put a temporary folder with a random filename,
as the user needs not read it: its content is parsed and
compared to a true phylogeny to obtain the inference errors.
Specifying \code{beast2_output_trees_filenames} allows to store
these one or more files in a more permanently stored location.}

\item{beast2_path}{Path to the
BEAST2 jar file (\code{beast.jar})
or BEAST2 binary file '(\code{beast})'.
Use \link[beastier]{get_default_beast2_jar_path} for the default
BEAST2 jar file path.
Use \link[beastier]{get_default_beast2_bin_path} for the default
BEAST2 binary file path.}

\item{beast2_rng_seed}{The random number generator seed used by BEAST2}

\item{brts}{set of branching times}

\item{chain_length}{something}

\item{clock_model}{a clock model,
as created by \link[beautier]{create_clock_model}}

\item{clock_models}{a list of one or more clock models,
as created by \link[beautier]{create_clock_model}}

\item{clock_model_name}{name of a clock model}

\item{crown_age}{the fixed crown age of the posterior. Set to NA
to let it be estimated}

\item{epsilon}{measure of relative accuracy when estimating a model's
evidence (also known as marginal likelihood).
Smaller values result in more precise estimations, that take
longer to compute}

\item{fasta_filename}{name of a FASTA file}

\item{filename}{the file's name, without the path}

\item{folder_name}{name of the main folder}

\item{inference_param}{one set of parameters for the Bayesian inference,
as can be created by \link{create_inference_param}.
The Bayesian inference is handled by the \link[babette]{babette}
R package, that calls the phylogenetic tool \code{BEAST2}.}

\item{init_speciation_rate}{a speciation rate}

\item{init_extinction_rate}{an extinction rate}

\item{lambda}{per-lineage speciation rate}

\item{marg_liks}{a data frame with marginal likelihoods/evidences.
A test data frame can be created by \link{create_test_marg_liks}}

\item{mbd_l_matrix}{the L matrix of an MBD tree}

\item{mbd_mutation_rate}{the mutation rate when creating an alignment
from a MBD tree}

\item{mbd_tree}{an MBD tree}

\item{mcmc}{MCMC options, as created by \link[beautier]{create_mcmc}}

\item{model_selection}{one ways to select the models used in
inference, for example, \code{generative} picks the generative
model, where \code{most_evidence} picks the model with most
evidence. See \link{get_model_selections} for a list of}

\item{model_select_param}{one parameter set to select an inference model
(a combination of site model, clock model and tree prior),
as can be created by \link{create_model_select_param}}

\item{model_select_params}{a list of one or more parameter sets
to select an inference model
(a combination of site model, clock model and tree prior),
as can be created by \link{create_model_select_param}}

\item{mrca_prior}{an MRCA prior,
as created by \link[beautier]{create_mrca_prior}}

\item{mu}{per-species extinction rate}

\item{mutation_rate}{the mutation rate per base pair per time unit}

\item{nu}{the rate at which a multiple-birth specation is triggered}

\item{nu_events}{the number of nu-triggered events that have to be
present in the simulated tree}

\item{parameter_filename}{full path to a 'parameters.csv' file}

\item{parameters_filename}{full path to a 'parameters.csv' file}

\item{phylo}{a phylogeny of class \link[ape]{phylo}}

\item{phylogeny}{a phylogeny of class \link[ape]{phylo}}

\item{posterior_trees}{phylogenetic trees in a BEAST2 posterior,
of class \code{multiphylo}}

\item{precision}{define the precision of the approximation.}

\item{project_folder_name}{project folder name}

\item{root_sequence}{the DNA sequence at the root of the phylogeny.
By default, this will consist out of only adenine}

\item{sample_interval}{the interval at which the MCMC algorithm
makes a measurement}

\item{seed}{a random number generator seed}

\item{sequence_length}{the length of each DNA sequence in an alignment}

\item{sim_pars}{something}

\item{sim_phylo}{something}

\item{site_model}{a nucleotide substitution model,
as created by \link[beautier]{create_site_model}}

\item{site_models}{a list of one or more site models,
as created by \link[beautier]{create_site_model}}

\item{site_model_name}{name of a site model}

\item{sub_chain_length}{length of the sub-chain used by the Nested Sampling
algorithm to estimate the marginal likelihood}

\item{tree}{an ultrametric phylogenetic tree of class \link[ape]{phylo}}

\item{tree_filename}{name of the phylogeny file}

\item{tree_prior}{a tree prior,
as created by \link[beautier]{create_tree_prior}}

\item{tree_priors}{a list of one or more tree priors,
as created by \link[beautier]{create_tree_prior}}

\item{tree_prior_name}{name of a tree prior}

\item{trees_filename}{name of the BEAST2 posterior phylogenies file}

\item{twinning_params}{can be \code{NA} if no twinning is desired,
or can be the twinning parameters,
as can be created by \link{create_twinning_param}}

\item{type}{one or more ways to select the models used in inference:
\itemize{
  \item \code{"generative"}: pick the generative model
  \item \code{most_evidence} picks the model with most evidence
}
See \link{get_model_selections} for a list.}

\item{verbose}{if TRUE, show more output}
}
\description{
This function does nothing. It is intended to inherit is parameters'
documentation.
}
\note{
This is an internal function, so it should be marked with
  \code{@noRd}. This is not done, as this will disallow all
  functions to find the documentation parameters
}
\author{
Documentation by Giovanni Laudanno,
  use of this function by Richel J.C. Bilderbeek
}
