% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_beast2_state_files.R
\name{delete_beast2_state_files}
\alias{delete_beast2_state_files}
\title{Delete the BEAST2 state files, if present.}
\usage{
delete_beast2_state_files(beast2_optionses, verbose = FALSE)
}
\arguments{
\item{beast2_optionses}{list of one or more BEAST2 options,
as can be created by \link[beastier]{create_beast2_options}}

\item{verbose}{if TRUE, show more output}
}
\description{
Delete the BEAST2 state files, if present.
}
