% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pirouette.R
\name{enpointe}
\alias{enpointe}
\title{Random rotation tree}
\usage{
enpointe(x, y, weights = NULL, newdim = 2, prob = 1/sqrt(ncol(x)),
  gbm_control = list(n.trees = 1, interaction.depth = 2, shrinkage = 0.1,
  verbose = FALSE), ...)
}
\arguments{
\item{x}{a sparse matrix of x variables}

\item{y}{the target variable for classification or regression}

\item{weights}{option case weights}

\item{newdim}{dimensions to project into}

\item{prob}{passed to splat}

\item{gbm_control}{Control arguments to pass to the gbm call. Use 2 trees.}

\item{...}{passed to rpart}
}
\value{
an object of class enpointe
}
\description{
This is the workhorse function of the pirouette package.  It
splats the dateset and fits an rpart model to it.
}

