% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorses_to_data_frame.R
\name{errorses_to_data_frame}
\alias{errorses_to_data_frame}
\title{Convert the collect of errors to a data frame}
\usage{
errorses_to_data_frame(errorses, experiments, marg_liks)
}
\arguments{
\item{errorses}{a collection of errors (hence the}

\item{experiments}{reduplicated plural)}
}
\description{
Convert the collect of errors to a data frame
}
