% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_evidences.R
\name{est_evidences}
\alias{est_evidences}
\title{Estimate the evidences}
\usage{
est_evidences(fasta_filename, model_select_params,
  experiments = list(create_experiment()),
  evidence_filename = tempfile(".csv"))
}
\arguments{
\item{fasta_filename}{name of a FASTA file}

\item{model_select_params}{a list of one or more parameter sets
to select an inference model
(a combination of site model, clock model and tree prior),
as can be created by \link{create_model_select_param}}

\item{experiments}{a list of one or more \link{pirouette} experiments,
as can be created by \link{create_experiment}}

\item{evidence_filename}{filename to store the estimated
evidences (aka marginal likelihoods)}
}
\value{
a data frame with evidences
}
\description{
Estimate the evidences
}
\author{
Richel J.C. Bilderbeek
}
