% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_evidences.R
\name{est_evidences}
\alias{est_evidences}
\title{Estimate the evidences}
\usage{
est_evidences(fasta_filename,
  experiments = list(create_test_experiment()),
  evidence_epsilon = 1e-12, evidence_filename = tempfile(pattern =
  "evidence_", fileext = ".csv"), verbose = FALSE)
}
\arguments{
\item{fasta_filename}{name of a FASTA file}

\item{experiments}{a list of one or more \link{pirouette} experiments,
as can be created by \link{create_experiment}}

\item{evidence_epsilon}{relative error in estimating the
evidence (aka marginal likelihood).}

\item{evidence_filename}{filename to store the estimated
evidences (aka marginal likelihoods)}

\item{verbose}{if TRUE, show more output}
}
\value{
a data frame with evidences
}
\description{
Estimate the evidences
}
\author{
Richel J.C. Bilderbeek
}
