% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_evidences.R
\name{est_evidences_old_skool}
\alias{est_evidences_old_skool}
\title{Estimate the evidences old skool}
\usage{
est_evidences_old_skool(fasta_filename, model_select_params)
}
\value{
a data frame with evidences
}
\description{
Estimate the evidences old skool
}
\author{
Richel J.C. Bilderbeek
}
