% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_alignment_sequences.R
\name{get_alignment_sequences}
\alias{get_alignment_sequences}
\title{Get the sequences from an alignment}
\usage{
get_alignment_sequences(alignment)
}
\arguments{
\item{alignment}{a DNA alignment, of class \link[ape]{DNAbin}}
}
\value{
a numeric vector with the sequences
}
\description{
Get the sequences from an alignment
}
\examples{
get_alignment_sequences(
  alignment = ape::as.DNAbin(
    x = list(species_1 = strsplit("aaaa", split = "")[[1]])
  )
)
}
