% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gamma_error_function.R
\name{get_gamma_error_function}
\alias{get_gamma_error_function}
\title{Get an error function that uses the difference in gamma statistic}
\usage{
get_gamma_error_function()
}
\description{
Get an error function that uses the difference in gamma statistic
}
\author{
Richel J.C. Bilderbeek
}
