% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_selections.R
\name{get_model_selections}
\alias{get_model_selections}
\title{Get the possible ways to select an inference model}
\usage{
get_model_selections()
}
\value{
a character vector
}
\description{
Get the possible ways to select an inference model
}
\examples{
get_model_selections()
}
\author{
Richèl J.C. Bilderbeek
}
