% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pir_params_filenames.R
\name{get_pir_params_filenames}
\alias{get_pir_params_filenames}
\title{Extract the filenames from a \code{pir_params}}
\usage{
get_pir_params_filenames(pir_params)
}
\arguments{
\item{pir_params}{the parameters of \link[pirouette]{pirouette}.
They are created by \link{create_pir_params}.}
}
\description{
Extract the filenames from a \code{pir_params}
}
\author{
Richèl J.C. Bilderbeek
}
