% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/get_sim_true_alignment_with_linked_node_sub_site_model_function.R
\name{get_sim_true_alignment_with_linked_node_sub_site_model_function}
\alias{get_sim_true_alignment_with_linked_node_sub_site_model_function}
\title{Create a function to simulate a true alignment with a linked
node substitution site model}
\usage{

  get_sim_true_alignment_with_linked_node_sub_site_model_function(root_sequence)
}
\arguments{
\item{root_sequence}{the DNA sequence at the root of the phylogeny.
By default, this will consist out of an equal amount of each letter
Use \link{check_root_sequence} to check if a root sequence is valid.}
}
\description{
Create a function to simulate a true alignment with a linked
node substitution site model
}
\examples{
# Create a valid 'check_sim_true_alignment_function'
f <- get_sim_true_alignment_with_linked_node_sub_site_model_function(
  root_sequence = "acgt"
)
check_sim_true_alignment_function(f)

# Simulate a true alignment
phylogeny <- ape::read.tree(text = "((A:1, B:1):1, C:2);")
true_alignment <- f(true_phylogeny = phylogeny)
}
