% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/get_sim_true_alignment_with_unlinked_node_sub_site_model_function.R
\name{get_sim_true_alignment_with_unlinked_node_sub_site_model_function}
\alias{get_sim_true_alignment_with_unlinked_node_sub_site_model_function}
\title{Create a function to simulate a true alignment with an unlinked
node substitution site model}
\usage{

  get_sim_true_alignment_with_unlinked_node_sub_site_model_function(branch_subst_matrix = NULL,
  node_subst_matrix = 1, branch_mutation_rate = 1,
  node_mutation_rate = 1, base_frequencies = NULL, root_sequence,
  node_time = 0.001)
}
\arguments{
\item{branch_subst_matrix}{substitution matrix along the branches.
See, among others, \link[nodeSub]{sim_dual_independent} for more details}

\item{node_subst_matrix}{substitution matrix on the nodes.
See, among others, \link[nodeSub]{sim_dual_independent} for more details}

\item{branch_mutation_rate}{mutation rate along the branch.
See, among others, \link[nodeSub]{sim_dual_independent} for more details}

\item{node_mutation_rate}{mutation rate on the node.
See, among others, \link[nodeSub]{sim_dual_independent} for more details}

\item{base_frequencies}{the four base frequencies (a, c, g, t) to be
specified to create the rate matrix (i.e. Q matrix)
used to simulate alignments}

\item{root_sequence}{the DNA sequence at the root of the phylogeny.
By default, this will consist out of an equal amount of each letter
Use \link{check_root_sequence} to check if a root sequence is valid.}

\item{node_time}{amount of time spent at the nodes.
See, among others, \link[nodeSub]{sim_dual_independent} for more details}
}
\description{
Create a function to simulate a true alignment with an unlinked
node substitution site model
}
\examples{
f <- get_sim_true_alignment_with_unlinked_node_sub_site_model_function(
  root_sequence = "acgt"
)
check_sim_true_alignment_function(f)
phylogeny <- ape::read.tree(text = "((A:1, B:1):1, C:2);")
true_alignment <- f(true_phylogeny = phylogeny)
}
