% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/get_sim_twin_alignment_with_same_n_mutation_function.R
\name{get_sim_twin_alignment_with_same_n_mutation_function}
\alias{get_sim_twin_alignment_with_same_n_mutation_function}
\title{Simulate a twin alignment from a
twin phylogeny and a true alignment, of which the twin
alignment has an equal amount of mutations from the root sequence
as the true alignment has}
\usage{
get_sim_twin_alignment_with_same_n_mutation_function()
}
\value{
the function \link{sim_twin_alignment_with_same_n_mutation}
}
\description{
This is an adapter function
(see \url{https://en.wikipedia.org/wiki/Adapter_pattern}),
with the purpose of passing \link{check_sim_twin_alignment_function},
by being a function with the function arguments \code{twin_phylogeny}
and \code{true_alignment}.
}
\examples{
library(testthat)

expect_silent(
  check_sim_twin_alignment_function(
    get_sim_twin_alignment_with_same_n_mutation_function
  )
)
}
\seealso{
Use \link{get_sim_twin_alignment_with_standard_site_model_function} to
get a function that simulates a twin alignment ignoring the number
of mutations in the true alignment
}
