% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/get_sim_twin_alignment_with_standard_site_model_function.R
\name{get_sim_twin_alignment_with_standard_site_model_function}
\alias{get_sim_twin_alignment_with_standard_site_model_function}
\title{Get a function to simulate a twin alignment which uses a standard
site model}
\usage{
get_sim_twin_alignment_with_standard_site_model_function(root_sequence,
  mutation_rate, site_model = beautier::create_jc69_site_model())
}
\arguments{
\item{root_sequence}{the DNA sequence at the root of the phylogeny.
By default, this will consist out of an equal amount of each letter}

\item{mutation_rate}{the mutation rate per base pair per time unit}

\item{site_model}{a nucleotide substitution model,
which can be:
\itemize{
  \item{
    A standard nucloetide substitution model,
    as created by \link[beautier]{create_site_model}
  }
  \item{
    \code{linked_node_sub}: a linked node-substitution model
  }
  \item{
    \code{unlinked_node_sub}: an unlinked node-substitution model
  }
}}
}
\description{
Get a function to simulate a twin alignment which uses a standard
site model
}
