% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_test_alignment.R
\name{get_test_alignment}
\alias{get_test_alignment}
\title{Get an alignment for testing.}
\usage{
get_test_alignment(n_taxa = 3, sequence_length = 4)
}
\arguments{
\item{n_taxa}{number of tree tips}

\item{sequence_length}{the length of each DNA sequence in an alignment}
}
\description{
Get an alignment for testing.
}
\examples{
alignment <- get_test_alignment(
  n_taxa = 3,
  sequence_length = 4
)
check_alignment(alignment)
}
