% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tree_and_model_descriptions.R
\name{get_tree_and_model_descriptions}
\alias{get_tree_and_model_descriptions}
\title{Internal function to creat a mapping from a \link{tree_and_model}}
\usage{
get_tree_and_model_descriptions()
}
\value{
a \link[tibble]{tibble} with columns \link{tree_and_model}
  and \link{description}
}
\description{
Internal function to creat a mapping from a \link{tree_and_model}
}
\author{
Richèl J.C. Bilderbeek, Giovanni Laudanno
}
