% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_tree_types}
\alias{get_tree_types}
\title{Get the names of the tree types}
\usage{
get_tree_types()
}
\value{
the tree types
}
\description{
Get the names of the tree types
}
\examples{
library(testthat)

expect_true("true" \%in\% get_tree_types())
expect_true("twin" \%in\% get_tree_types())
expect_false("Jensen Ackles" \%in\% get_tree_types())
}
\author{
Giovanni Laudanno, Richel J.C. Bilderbeek
}
