% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_twin_methods}
\alias{get_twin_methods}
\title{Twin methods}
\usage{
get_twin_methods()
}
\value{
the twin methods
}
\description{
Twin methods
}
\examples{
get_twin_methods()
}
\author{
Giovanni Laudanno, Richèl J.C. Bilderbeek
}
