% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_twin_models}
\alias{get_twin_models}
\title{Twin models}
\usage{
get_twin_models()
}
\value{
the twin models
}
\description{
Twin models
}
\examples{

expect_true("yule" \%in\% get_twin_models())
expect_true("birth_death" \%in\% get_twin_models())
expect_true("copy_true" \%in\% get_twin_models())
expect_false("nonsense" \%in\% get_twin_models())
}
\author{
Giovanni Laudanno, Richèl J.C. Bilderbeek
}
