% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_experiment.R
\name{init_experiment}
\alias{init_experiment}
\title{Initialize the \code{experiment}.}
\usage{
init_experiment(experiment, alignment_params)
}
\arguments{
\item{experiment}{a \link{pirouette} experiment,
as can be created by \link{create_experiment}}

\item{alignment_params}{parameters to simulate an alignment,
as can be created by \link{create_alignment_params}}
}
\description{
A normal user should never need to call this function.
}
