% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_dna_seq.R
\name{is_dna_seq}
\alias{is_dna_seq}
\title{Determine if the string is a lowercase DNA sequence
of at least one base pair}
\usage{
is_dna_seq(s)
}
\arguments{
\item{s}{the string to be checked}
}
\value{
TRUE if the string is a lowercase DNA sequence
  of at least one base pair
}
\description{
Determine if the string is a lowercase DNA sequence
of at least one base pair
}
\examples{
  library(testthat)

  # OK: valid and lowercase characters
  expect_true(is_dna_seq("acgt"))

  # Must be lowercase
  expect_false(is_dna_seq("AGCT"))

  # Must be only valid characters
  expect_false(is_dna_seq("xxxx"))

  # Must have at least one nucleotide
  expect_false(is_dna_seq(""))
}
\author{
Richèl J.C. Bilderbeek
}
