% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_model_select_params.R
\name{list_model_select_params}
\alias{list_model_select_params}
\title{Make sure that model_select_params is a list
#' @inheritParams default_params_doc}
\usage{
list_model_select_params(model_select_params)
}
\description{
Make sure that model_select_params is a list
#' @inheritParams default_params_doc
}
\author{
Giovanni Laudanno
}
