% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo_to_errors.R
\name{phylo_to_errors}
\alias{phylo_to_errors}
\title{Measure the error BEAST2 makes from a true phylogeny}
\usage{
phylo_to_errors(phylogeny, alignment_params, site_model, clock_model,
  tree_prior, inference_param)
}
\description{
Measure the error BEAST2 makes from a true phylogeny
}
\author{
Richel J.C. Bilderbeek
}
