% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo_to_errors.R
\name{phylo_to_errors}
\alias{phylo_to_errors}
\title{Measure the error BEAST2 makes from a true phylogeny.}
\usage{
phylo_to_errors(phylogeny, alignment_params, inference_model,
  inference_params = create_inference_params(),
  error_measure_params = create_error_measure_params())
}
\arguments{
\item{phylogeny}{a phylogeny of class \link[ape]{phylo}}

\item{alignment_params}{parameters to simulate an alignment,
as can be created by \link{create_alignment_params}}

\item{inference_model}{an inference model, which is a combination
of site model, clock model, tree prior and BEAST2 input and
input filenames.}

\item{inference_params}{one set of parameters for the Bayesian inference,
as can be created by \link{create_inference_params}.
The Bayesian inference is handled by the \link[babette]{babette}
R package, that calls the phylogenetic tool \code{BEAST2}.}

\item{error_measure_params}{parameter set to specify how the
error between the given phylogeny and the Bayesian
posterior is determined.
Use \link{create_error_measure_params} to create such
a parameter set}
}
\description{
The supplied phylogeny (\code{phylogeny}) is the true/known phylogeny.
From the phylogeny, already an alignment is simulated and saved
as a FASTA file with name \code{alignment_params$fasta_filename}.
}
\details{
Using the setup itemized below, the posterior phylogengies are generated:
\itemize{
   \item a site model (\code{site_model})
   \item a clock model (\code{clock_model})
   \item a tree prior (\code{tree_prior})
   \item an MRCA prior (\code{inference_params$mrca_prior})
   \item an MCMC setup (\code{inference_params$mcmc})
   \item a BEAST2 RNG seed (\code{inference_params$rng_seed})
   \item a BEAST2 path (\code{inference_params$beast2_path})
}

The posterior phylogenies are compared to the true/known phylogeny
using the nLTT statistics. These nLTT statistics, all with values
betweem (including) zero and (including) one, are returned.
}
\author{
Richel J.C. Bilderbeek
}
