% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo_to_posterior.R
\name{phylo_to_posterior}
\alias{phylo_to_posterior}
\title{Creates a posterior from the phylogeny}
\usage{
phylo_to_posterior(phylogeny, alignment_params, inference_params)
}
\arguments{
\item{phylogeny}{a phylogeny of class \link[ape]{phylo}}

\item{alignment_params}{parameters to simulate an alignment,
as can be created by \link{create_alignment_params}}

\item{inference_params}{parameters for the Bayesian inference,
as can be created by \link{create_inference_params}.
The Bayesian inference is handled by the \link[babette]{babette}
R package, that calls the phylogenetic tool \code{BEAST2}.}
}
\value{
a posterior of phylogenies
}
\description{
Creates a posterior from the phylogeny
}
\author{
Richel J.C. Bilderbeek
}
