% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pir_plot.R
\name{pir_plot}
\alias{pir_plot}
\title{Plot the error BEAST2 make from the known phylogeny}
\usage{
pir_plot(pir_out)
}
\arguments{
\item{pir_out}{the output created by \code{\link{pir_run}}}
}
\value{
a ggplot2 plot
}
\description{
Plot the error BEAST2 make from the known phylogeny
}
\examples{
  pir_plot(
    create_test_pir_run_output(
      add_twin = TRUE,
      add_best = TRUE
    )
  )
}
\seealso{
Use \link{create_test_pir_run_output} to create a test output
  of \link{pir_run}
}
\author{
Richèl J.C. Bilderbeek
}
