% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pir_plot.R
\name{pir_plot}
\alias{pir_plot}
\title{Plot the error BEAST2 make from the known phylogeny}
\usage{
pir_plot(pir_out)
}
\arguments{
\item{pir_out}{the output created by \code{\link{pir_run}}}
}
\value{
a ggplot2 plot
}
\description{
Plot the error BEAST2 make from the known phylogeny
}
\examples{
  phylogeny <- ape::read.tree(text = "(((A:1, B:1):1, C:2):1, D:3);")
  alignment_params <- create_alignment_params(
    mutation_rate = 0.01
  )
  errors <- pir_run(
    phylogeny = phylogeny,
    alignment_params = alignment_params,
    model_select_params = list(
      create_gen_model_select_param(
        alignment_params = alignment_params
      )
    ),
    inference_param = create_inference_param(
      mcmc = beautier::create_mcmc(chain_length = 2000, store_every = 1000)
    )
  )
  pir_plot(errors)
}
\author{
Richel J.C. Bilderbeek
}
