% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pir_plot_from_long.R
\name{pir_plot_from_long}
\alias{pir_plot_from_long}
\title{Plot the error 'BEAST2' makes from a known phylogeny
from tidy data}
\usage{
pir_plot_from_long(df_long)
}
\arguments{
\item{df_long}{the output created by \code{\link{pir_run}} in the long form}
}
\value{
a \code{ggplot2} plot
}
\description{
Plot the error 'BEAST2' makes from a known phylogeny
from tidy data
}
\author{
Richèl J.C. Bilderbeek, Giovanni Laudanno
}
