% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pir_run.R
\name{pir_run}
\alias{pir_run}
\title{Meaure the error BEAST2 makes from a known phylogeny}
\usage{
pir_run(phylogeny, mcmc)
}
\arguments{
\item{phylogeny}{a phylogeny of class \link[ape]{phylo}}

\item{mcmc}{MCMC options, as created by \link[beautier]{create_mcmc}}
}
\value{
a data frame with errors
}
\description{
Meaure the error BEAST2 makes from a known phylogeny
}
\author{
Richel J.C. Bilderbeek
}
