% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pir_run.R
\name{pir_run}
\alias{pir_run}
\title{Measure the error BEAST2 makes from a known phylogeny.}
\usage{
pir_run(
  phylogeny,
  pir_params = create_pir_params(alignment_params = create_alignment_params(),
    twinning_params = create_twinning_params())
)
}
\arguments{
\item{phylogeny}{a phylogeny of class \link[ape]{phylo}}

\item{pir_params}{the parameters of \link[pirouette]{pirouette}.
They are created by \link{create_pir_params}.}
}
\value{
a data frame with errors, with as many rows as model selection
parameter sets. The output can be checked using \link{check_pir_out}.
}
\description{
From a phylogeny of (un)known speciation model,
an alignment is created using a known site model and clock model,
as given by \code{alignment_params}.
}
\examples{

phylogeny <- ape::read.tree(text = "(((A:1, B:1):1, C:2):1, D:3);")
pir_params <- create_test_pir_params()

errors <- NA
if (
  rappdirs::app_dir()$os != "win" &&
  is_on_ci() &&
  is_beast2_installed()
) {
  errors <- pir_run(
    phylogeny = phylogeny,
    pir_params = pir_params
  )
} else {
  errors <- create_test_pir_run_output()
}

# Return value
"tree" \%in\% names(errors))
is.factor(errors$tree))
"true" \%in\% errors$tree)

"inference_model" \%in\% names(errors))
is.factor(errors$inference_model))
"generative" \%in\% errors$inference_model)

"inference_model_weight" \%in\% names(errors))
!is.factor(errors$inference_model_weight))

"site_model" \%in\% names(errors))
is.factor(errors$site_model))
"JC69" \%in\% errors$site_model)

"clock_model" \%in\% names(errors))
is.factor(errors$clock_model))

"tree_prior" \%in\% names(errors))
is.factor(errors$tree_prior))
"birth_death" \%in\% errors$tree_prior ||
  "yule" \%in\% errors$tree_prior
)

"error_1" \%in\% names(errors))
!is.factor(errors$error_1))

# Errors more than zero
col_first_error <- which(colnames(errors) == "error_1")
col_last_error <- ncol(errors)
all(errors[, col_first_error:col_last_error] > 0.0))
n_errors <- col_last_error - col_first_error + 1
n_errors < 11) # due to burn-in
}
\seealso{
\itemize{
  \item Use \link{pir_plot} to display the output of \link{pir_run} as a
    figure.
  \item Use \link{create_test_pir_run_output} to create a test output
    of \link{pir_run}.
  \item Use \link{pir_runs} to do multiple \link{pirouette} runs.
}
}
\author{
Richèl J.C. Bilderbeek, Giovanni Laudanno
}
