% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pir_run.R
\name{pir_run}
\alias{pir_run}
\title{Meaure the error BEAST2 makes from a known phylogeny}
\usage{
pir_run(phylogeny, alignment_params, inference_params,
  model_selections = "generative")
}
\arguments{
\item{phylogeny}{a phylogeny of class \link[ape]{phylo}}

\item{alignment_params}{parameters to simulate an alignment,
as can be created by \link{create_alignment_params}}

\item{inference_params}{parameters for the Bayesian inference,
as can be created by \link{create_inference_params}.
The Bayesian inference is handled by the \link[babette]{babette}
R package, that calls the phylogenetic tool \code{BEAST2}.}

\item{model_selections}{one or more ways to select the models used in
inference, for example, \code{generative} picks the generative
model, where \code{most_evidence} picks the model with most
evidence. See \link{get_model_selections} for a list of}
}
\value{
a data frame with errors
}
\description{
Meaure the error BEAST2 makes from a known phylogeny
}
\author{
Richel J.C. Bilderbeek
}
