% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pir_run.R
\name{pir_run}
\alias{pir_run}
\title{Measure the error BEAST2 makes from a known phylogeny.}
\usage{
pir_run(phylogeny, alignment_params,
  model_select_params = create_gen_model_select_params(alignment_params),
  inference_params)
}
\arguments{
\item{phylogeny}{a phylogeny of class \link[ape]{phylo}}

\item{alignment_params}{parameters to simulate an alignment,
as can be created by \link{create_alignment_params}}

\item{model_select_params}{parameters to select a model (a combination
of site model, clock model and tree prior),
as can be created by \link{create_model_select_params}}

\item{inference_params}{parameters for the Bayesian inference,
as can be created by \link{create_inference_params}.
The Bayesian inference is handled by the \link[babette]{babette}
R package, that calls the phylogenetic tool \code{BEAST2}.}
}
\value{
a data frame with errors
}
\description{
From a phylogeny of (un)known speciation model,
an alignment is created using a known site model and clock model,
as given by \code{alignment_params}.
From the model selection parameters in \code{model_select_params},
models are selected to run BEAST2 with. One such model is to use
the generative site and clock model the alignment is created,
}
\author{
Richel J.C. Bilderbeek
}
