% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pir_run.R
\name{pir_run}
\alias{pir_run}
\title{Measure the error BEAST2 makes from a known phylogeny.}
\usage{
pir_run(phylogeny, twinning_params = NA, alignment_params,
  model_select_params = create_gen_model_select_param(alignment_params),
  inference_params = create_inference_params(),
  experiments = list(create_experiment()),
  error_measure_params = create_error_measure_params())
}
\arguments{
\item{phylogeny}{a phylogeny of class \link[ape]{phylo}}

\item{twinning_params}{can be \code{NA} if no twinning is desired,
or can be the twinning parameters,
as can be created by \link{create_twinning_params}}

\item{alignment_params}{parameters to simulate an alignment,
as can be created by \link{create_alignment_params}}

\item{model_select_params}{a list of one or more parameter sets
to select an inference model
(a combination of site model, clock model and tree prior),
as can be created by \link{create_model_select_param}}

\item{inference_params}{one set of parameters for the Bayesian inference,
as can be created by \link{create_inference_params}.
The Bayesian inference is handled by the \link[babette]{babette}
R package, that calls the phylogenetic tool \code{BEAST2}.}

\item{experiments}{a list of one or more \link{pirouette} experiments,
as can be created by \link{create_experiment}}

\item{error_measure_params}{parameter set to specify how the
error between the given phylogeny and the Bayesian
posterior is determined.
Use \link{create_error_measure_params} to create such
a parameter set}
}
\value{
a data frame with errors, with as many rows as model selection
  parameter sets.
  Tip: use \link{pir_plot} to display it
}
\description{
From a phylogeny of (un)known speciation model,
an alignment is created using a known site model and clock model,
as given by \code{alignment_params}.
}
\details{
From each of the one or more model selection parameters in
\code{model_select_params}, one inference model is concluded.
One such parameter setup is to use the generative site and clock model
the alignment is created with (\code{"generative"}), or use the model
that has the most evidence (\code{"most evidence"}).

Each inference model is used to run BEAST2 and measures the
difference between the known/true/given phylogeny and the
ones created by BEAST2.
}
\examples{
  phylogeny <- ape::read.tree(text = "(((A:1, B:1):1, C:2):1, D:3);")
  alignment_params <- create_alignment_params(
    mutation_rate = 0.01
  )
  errors <- pir_run(
    phylogeny = phylogeny,
    alignment_params = alignment_params,
    model_select_params = create_gen_model_select_param(
      alignment_params = alignment_params
    ),
    inference_params = create_inference_params(
      mcmc = beautier::create_mcmc(chain_length = 2000, store_every = 1000)
    )
  )
  pir_plot(errors)
}
\author{
Richel J.C. Bilderbeek
}
