% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pir_run.R
\name{pir_run}
\alias{pir_run}
\title{Creates a posterior from the phylogeny}
\usage{
pir_run(phylogeny, sequence_length, root_sequence = paste(rep("a",
  sequence_length), collapse = ""), mutation_rate, mcmc,
  site_models = beautier::create_jc69_site_model(),
  clock_models = beautier::create_strict_clock_model(),
  tree_priors = beautier::create_bd_tree_prior(), crown_age = NA,
  mrca_distr = NA, alignment_rng_seed = 0, beast2_rng_seed = 1,
  verbose = FALSE, beast2_path = beastier::get_default_beast2_path())
}
\arguments{
\item{phylogeny}{a phylogeny}

\item{sequence_length}{the number of basepair the simulated DNA
alignment consists of}

\item{root_sequence}{the DNA sequence at the root of the phylogeny.
By default, this will consist out of only adenine}

\item{mutation_rate}{the mutation rate per base pair per time unit}

\item{mcmc}{MCMC options, as created by \link[beautier]{create_mcmc}}

\item{site_models}{one or more nucleotide substitution models,
as created by \link[beautier]{create_site_model}}

\item{clock_models}{one or more clock models,
as created by \link[beautier]{create_clock_model}}

\item{tree_priors}{one or more tree prios,
as created by \link[beautier]{create_tree_prior}}

\item{crown_age}{the fixed crown age of the posterior. Set to NA
to let it be estimated}

\item{mrca_distr}{if MRCA prior used on all taxa.
Set to NA to not use an MRCA prior}

\item{alignment_rng_seed}{The random number generator seed used
to generate an alignment}

\item{beast2_rng_seed}{The random number generator seed used by BEAST2}

\item{verbose}{if TRUE, show more output}

\item{beast2_path}{Path to the BEAST2 binary (\code{beast})
or jar file (\code{beast.jar})}
}
\value{
a posterior of phylogenies
}
\description{
Creates a posterior from the phylogeny
}
\examples{
 # Create a phylogeny
 phylogeny <- ape::read.tree(text = "(((A:1,B:1):1,C:2):1,D:3);")

 # Create a BEAST2 posterior from this phylogeny's simulated alignment.
 # Estimate the crown age
 out <- pir_run(
   phylogeny = phylogeny,
   sequence_length = 10,
   mutation_rate = 0.1,
   mcmc = beautier::create_mcmc(chain_length = 2000)
 )
 # Trees are estimated after 0, 1000 and 2000 MCMC states
 testit::assert(length(out$trees) == 3)
 testit::assert(nrow(out$estimates) == 3)

 # Create a BEAST2 posterior from this phylogeny's simulated alignment,
 # now assume a (close-to) fixed crown age
 crown_age <- 15.0
 out <- pir_run(
   phylogeny = phylogeny,
   sequence_length = 10,
   mutation_rate = 0.1,
   mcmc = beautier::create_mcmc(chain_length = 2000),
   mrca_distr = beautier::create_normal_distr(
     mean = create_mean_param(crown_age),
     sigma = create_sigma_param(0.001)
   )
 )
 testit::assert(length(out$trees) == 3)
 testit::assert(nrow(out$estimates) == 3)
 # Estimated crown age (called 'TreeHeight' by BEAST2) is not
 # yet estimated correctly. Increase the MCMC length to achieve this
 testit::assert(!all(abs(crown_age - out$estimates$TreeHeight) == crown_age))
}
\author{
Richel J.C. Bilderbeek
}
