% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pir_run.R
\name{pir_run}
\alias{pir_run}
\title{Measure the error BEAST2 makes from a known phylogeny.}
\usage{
pir_run(phylogeny, pir_params = create_pir_params(alignment_params =
  create_alignment_params(mutation_rate = create_standard_mutation_rate),
  twinning_params = create_twinning_params(rng_seed = "same_seed")))
}
\arguments{
\item{phylogeny}{a phylogeny of class \link[ape]{phylo}}

\item{pir_params}{the parameters of \link[pirouette]{pirouette}.
They are created by \link{create_pir_params}.}
}
\value{
a data frame with errors, with as many rows as model selection
  parameter sets.
}
\description{
From a phylogeny of (un)known speciation model,
an alignment is created using a known site model and clock model,
as given by \code{alignment_params}.
}
\seealso{
Use \link{pir_plot} to display the output of \link{pir_run} as a
  figure.
  Use \link{create_test_pir_run_output} to create a test output
  of \link{pir_run}
}
\author{
Richel J.C. Bilderbeek, Giovanni Laudanno
}
