% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pir_run.R
\name{pir_run_check_inputs}
\alias{pir_run_check_inputs}
\title{Checks the \link{pir_run} inputs.}
\usage{
pir_run_check_inputs(phylogeny, alignment_params, model_select_params,
  inference_param)
}
\arguments{
\item{phylogeny}{a phylogeny of class \link[ape]{phylo}}

\item{alignment_params}{parameters to simulate an alignment,
as can be created by \link{create_alignment_params}}

\item{model_select_params}{a list of one or more parameter sets
to select an inference model
(a combination of site model, clock model and tree prior),
as can be created by \link{create_model_select_param}}

\item{inference_param}{one set of parameters for the Bayesian inference,
as can be created by \link{create_inference_param}.
The Bayesian inference is handled by the \link[babette]{babette}
R package, that calls the phylogenetic tool \code{BEAST2}.}
}
\value{
nothing
}
\description{
Calls \link{stop} if an input is invalud
}
\author{
Richel J.C. Bilderbeek
}
