% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pir_run.R
\name{pir_run_tree}
\alias{pir_run_tree}
\title{Measure the error BEAST2 makes from a phylogeny}
\usage{
pir_run_tree(phylogeny, tree_type = "true", alignment_params,
  experiments = list(create_test_experiment()),
  error_measure_params = create_error_measure_params(),
  evidence_filename = tempfile(pattern = "evidence_", fileext = ".csv"),
  verbose = FALSE)
}
\arguments{
\item{phylogeny}{a phylogeny of class \link[ape]{phylo}}

\item{tree_type}{type of tree, can be \code{true} for the true
phylogeny, and \code{twin} for its twin tree}

\item{alignment_params}{parameters to simulate an alignment,
as can be created by \link{create_alignment_params}}

\item{experiments}{a list of one or more \link{pirouette} experiments,
as can be created by \link{create_experiment}}

\item{error_measure_params}{parameter set to specify how the
error between the given phylogeny and the Bayesian
posterior is determined.
Use \link{create_error_measure_params} to create such
a parameter set}

\item{evidence_filename}{filename to store the estimated
evidences (aka marginal likelihoods)}

\item{verbose}{if TRUE, show more output}
}
\value{
a data frame with errors, with as many rows as model selection
  parameter sets
}
\description{
The phylogeny can be the true tree or its twin.
}
\author{
Richel J.C. Bilderbeek, Giovanni Laudanno
}
