% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pir_run.R
\name{pir_run_tree}
\alias{pir_run_tree}
\title{Measure the error BEAST2 makes from a phylogeny}
\usage{
pir_run_tree(phylogeny, tree_type = "true", alignment_params,
  model_select_params = list(create_gen_model_select_param(alignment_params)),
  inference_param = create_inference_param(),
  error_measure_params = create_error_measure_params())
}
\arguments{
\item{phylogeny}{a phylogeny of class \link[ape]{phylo}}

\item{tree_type}{type of tree, can be \code{true} for the true
phylogeny, and \code{twin} for its twin tree}

\item{alignment_params}{parameters to simulate an alignment,
as can be created by \link{create_alignment_params}}

\item{model_select_params}{a list of one or more parameter sets
to select an inference model
(a combination of site model, clock model and tree prior),
as can be created by \link{create_model_select_param}}

\item{inference_param}{one set of parameters for the Bayesian inference,
as can be created by \link{create_inference_param}.
The Bayesian inference is handled by the \link[babette]{babette}
R package, that calls the phylogenetic tool \code{BEAST2}.}

\item{error_measure_params}{parameter set to specify how the
error between the given phylogeny and the Bayesian
posterior is determined.
Use \link{create_error_measure_params} to create such
a parameter set}
}
\value{
a data frame with errors, with as many rows as model selection
  parameter sets
}
\description{
The phylogeny can be the true tree or its twin.
}
\author{
Richel J.C. Bilderbeek
}
