% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pir_run_true_tree.R
\name{pir_run_true_tree}
\alias{pir_run_true_tree}
\title{Measure the error BEAST2 makes from a phylogeny}
\usage{
pir_run_true_tree(true_phylogeny, pir_params = create_test_pir_params())
}
\arguments{
\item{true_phylogeny}{a phylogeny of class \link[ape]{phylo}}

\item{pir_params}{the parameters of \link[pirouette]{pirouette}.
They are created by \link{create_pir_params}.}
}
\value{
a data frame with errors, with as many rows as model selection
  parameter sets
}
\description{
The phylogeny can be the true tree or its twin.
}
\author{
Richèl J.C. Bilderbeek, Giovanni Laudanno
}
