% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pirouette.R
\name{pirouette}
\alias{pirouette}
\title{pirouette}
\usage{
pirouette(x, y, ctrl = pirouetteControl())
}
\arguments{
\item{x}{a sparse matrix of x variables}

\item{y}{the target variable for classification or regression}
}
\value{
an object of class pirouette
}
\description{
Random rotation forests for sparse data
}
\references{
\url{http://web.stanford.edu/~hastie/Papers/Ping/KDD06_rp.pdf}
}

