% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pirouette.R
\name{pirouetteControl}
\alias{pirouetteControl}
\title{pirouetteControl}
\usage{
pirouetteControl(ntrees = 100, newdim = 2, allowParallel = TRUE)
}
\arguments{
\item{ntrees}{number of models to fit}

\item{newdim}{dimensions to project into}

\item{allowParallel}{the function to run in parallel (if a foreach backend is loaded)}
}
\value{
a list
}
\description{
Define a random sparse projection
}

