% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pirouette.R
\name{pirouetteControl}
\alias{pirouetteControl}
\title{pirouetteControl}
\usage{
pirouetteControl(ntrees = 100, mtry = 2, nobs = 500,
  allowParallel = TRUE)
}
\arguments{
\item{ntrees}{number of models to fit}

\item{mtry}{dimensionality of the random projection}

\item{nobs}{maximum number of observations used per tree}

\item{allowParallel}{the function to run in parallel (if a foreach backend is loaded)}
}
\value{
a list
}
\description{
Define a random sparse projection
}

