% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_alignment_from_file.R
\name{plot_alignment_from_file}
\alias{plot_alignment_from_file}
\title{Plot an alignment stored as a file}
\usage{
plot_alignment_from_file(fasta_filename, title = "")
}
\arguments{
\item{fasta_filename}{name of a FASTA file}

\item{title}{the plot title
alignment_params <- create_test_alignment_params()
alignment <- create_tral_file(
  phylogeny = ape::rcoal(4),
  alignment_params = alignment_params
)
plot_alignment_from_file(fasta_filename = alignment_params$fasta_filename)}
}
\description{
Plot an alignment stored as a file
}
