% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_errors_csv.R
\name{read_errors_csv}
\alias{read_errors_csv}
\title{Create a \code{pir_out} from a folder}
\usage{
read_errors_csv(errors_filename)
}
\arguments{
\item{errors_filename}{baseline name for errors filenames}
}
\value{
a \code{pir_out}, as can be checked by \link{check_pir_out}
}
\description{
Create a \code{pir_out} from a folder
}
\author{
Richèl J.C. Bilderbeek
}
