% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relevel_inference_model.R
\name{relevel_inference_model}
\alias{relevel_inference_model}
\title{Internal function}
\usage{
relevel_inference_model(inference_model)
}
\arguments{
\item{inference_model}{an inference model, which is a combination
of site model, clock model, tree prior and BEAST2 input and
input filenames.}
}
\value{
a releveled \link{inference_model}
}
\description{
Internal function to relevel the \code{inference_model}, so that
\link{pir_plot} has the legend labels in the right order
}
\author{
Richèl J.C. Bilderbeek
}
