% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run}
\alias{run}
\title{Creates a posterior from the phylogeny}
\usage{
run(phylogeny, sequence_length, mutation_rate, chain_length, crown_age = NA,
  mrca_distr = NA, rng_seed = 0, verbose = FALSE,
  beast_jar_path = beastier::get_default_beast2_jar_path())
}
\arguments{
\item{phylogeny}{a phylogeny}

\item{sequence_length}{the number of basepair the simulated DNA
alignment consists of}

\item{mutation_rate}{the mutation rate per base pair per time unit}

\item{chain_length}{MCMC chain length}

\item{crown_age}{the fixed crown age of the posterior. Set to NA
to let it be estimated}

\item{mrca_distr}{if MRCA prior used on all taxa.
Set to NA to not use an MRCA prior}

\item{rng_seed}{The random number generator seed used by BEAST2}

\item{verbose}{if TRUE, show more output}

\item{beast_jar_path}{Where the jar 'beast.jar' can be found}
}
\value{
a posterior of phylogenies
}
\description{
Creates a posterior from the phylogeny
}
\author{
Richel J.C. Bilderbeek
}
