% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_candidate_evidences.R
\name{select_candidate_evidences}
\alias{select_candidate_evidences}
\title{Select the evidences for candidate experiments}
\usage{
select_candidate_evidences(experiments = list(create_test_experiment()),
  marg_liks = create_test_marg_liks())
}
\arguments{
\item{experiments}{a list of one or more \link{pirouette} experiments,
as can be created by \link{create_experiment}. If more than one experiment
is provided and a "generative" experiment is part of them, the "generative"
one has to be the first in the list.}

\item{marg_liks}{a data frame with marginal likelihoods/evidences.
A test data frame can be created by \link{create_test_marg_liks}}
}
\description{
Select the evidences for candidate experiments
}
\examples{
library(testthat)

if (rappdirs::app_dir()$os != "win") {
  experiment_1 <- create_test_gen_experiment()
  experiment_2 <- create_test_cand_experiment()
  expect_equal(experiment_1$inference_conditions$model_type, "generative")
  expect_equal(experiment_2$inference_conditions$model_type, "candidate")
  experiments <- list(experiment_1, experiment_2)

  # Experiments must have different inference models
  experiments[[1]]$inference_model$site_model <- create_gtr_site_model()

  candidate_evidences <- select_candidate_evidences(
    experiments = experiments,
    marg_liks = create_test_marg_liks()
  )
  expect_equal(1, nrow(candidate_evidences))
}
}
\author{
Richèl J.C. Bilderbeek
}
