% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_candidate_evidences.R
\name{select_candidate_evidences}
\alias{select_candidate_evidences}
\title{Select the evidences for candidate experiments}
\usage{
select_candidate_evidences(experiments = list(create_test_experiment()),
  marg_liks = create_test_marg_liks())
}
\arguments{
\item{experiments}{a list of one or more \link{pirouette} experiments,
as can be created by \link{create_experiment}. If more than one experiment
is provided and a "generative" experiment is part of them, the "generative"
one has to be the first in the list.}

\item{marg_liks}{a data frame with marginal likelihoods/evidences.
A test data frame can be created by \link{create_test_marg_liks}}
}
\description{
Select the evidences for candidate experiments
}
\author{
Richèl J.C. Bilderbeek
}
