% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_alignment.R
\name{sim_alignment}
\alias{sim_alignment}
\title{Converts a phylogeny to a random DNA alignment}
\usage{
sim_alignment(phylogeny, alignment_params)
}
\arguments{
\item{phylogeny}{a phylogeny}

\item{sequence_length}{the number of nucleotides to alignment
will have per taxon}

\item{root_sequence}{the DNA sequence at the root of the phylogeny.
By default, this will consist out of only adenine}

\item{mutation_rate}{the rate per nucleotide to change,
per million years}
}
\value{
an alignment of type \code{DNAbin}
}
\description{
Converts a phylogeny to a random DNA alignment
}
\examples{
n_taxa <- 5
n_base_pairs <- 4
alignment <- sim_alignment(
   phylogeny = ape::rcoal(n_taxa),
   alignment_params = create_alignment_params(
     root_sequence = create_blocked_dna(length = n_base_pairs),
     mutation_rate = 0.1
   )
 )
 testit::assert(class(alignment) == "DNAbin")
 testit::assert(nrow(alignment) == n_taxa)
 testit::assert(ncol(alignment) == n_base_pairs)
}
\author{
Richel Bilderbeek
}
