% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_alignment.R
\name{sim_alignment}
\alias{sim_alignment}
\title{Converts a phylogeny to a random DNA alignment}
\usage{
sim_alignment(phylogeny, sequence_length, mutation_rate = 1)
}
\arguments{
\item{phylogeny}{a phylogeny}

\item{sequence_length}{the number of nucleotides to alignment
will have per taxon}

\item{mutation_rate}{the rate per nucleotide to change,
per million years}
}
\value{
an alignment
}
\description{
Converts a phylogeny to a random DNA alignment
}
\examples{
n_taxa <- 5
n_base_pairs <- 10
alignment <- sim_alignment(
   phylogeny = ape::rcoal(n_taxa),
   sequence_length = n_base_pairs,
   mutation_rate = 1
 )
 testit::assert(class(alignment) == "DNAbin")
 testit::assert(nrow(alignment) == n_taxa)
 testit::assert(ncol(alignment) == n_base_pairs)
}
\author{
Richel Bilderbeek
}
