% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_true_alignment.R
\name{sim_true_alignment}
\alias{sim_true_alignment}
\title{Simulate the true alignment from the true phylogeny}
\usage{
sim_true_alignment(true_phylogeny,
  alignment_params = create_alignment_params(), verbose = FALSE)
}
\arguments{
\item{true_phylogeny}{a phylogeny of class \link[ape]{phylo}}

\item{alignment_params}{parameters to simulate an alignment,
as can be created by \link{create_alignment_params}}

\item{verbose}{if TRUE, show more output}
}
\value{
an alignment of type \link[ape]{DNAbin}
}
\description{
Simulate the true alignment from the true phylogeny
}
\examples{
library(testthat)

# Create the phylogeny to simulate the alignment on
n_taxa <- 5
true_phylogeny <- ape::rcoal(n_taxa)

# Use default settings to create the alignment
alignment_params <- create_alignment_params()

# Simulate the alignment
alignment <- sim_true_alignment(
   true_phylogeny = true_phylogeny,
   alignment_params = alignment_params,
 )

expect_equal(class(alignment), "DNAbin")
expect_equal(nrow(alignment), n_taxa)
expect_equal(ncol(alignment), nchar(alignment_params$root_sequence))
}
\seealso{
Use \link{sim_true_alignment_file} to save the simulated alignment
  directly to a file
}
\author{
Richèl J.C. Bilderbeek
}
