% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/sim_twin_alignment_with_linked_node_sub_site_model.R
\name{sim_twin_alignment_with_linked_node_sub_site_model}
\alias{sim_twin_alignment_with_linked_node_sub_site_model}
\title{Adapter function to simulte the twin alignment
using the \code{linked_node_sub} site model}
\usage{
sim_twin_alignment_with_linked_node_sub_site_model(twin_phylogeny,
  true_alignment = "irrelevant", root_sequence, subst_matrix = NULL,
  branch_mutation_rate = 1, node_mutation_rate = 1,
  base_frequencies = NULL, node_time = 0.001)
}
\arguments{
\item{twin_phylogeny}{a phylogeny of class \link[ape]{phylo}}

\item{true_alignment}{a DNA alignment, of class \link[ape]{DNAbin}}

\item{root_sequence}{the DNA sequence at the root of the phylogeny.
By default, this will consist out of an equal amount of each letter
Use \link{check_root_sequence} to check if a root sequence is valid.}

\item{subst_matrix}{nucleotide substitution matrix}

\item{branch_mutation_rate}{mutation rate along the branch.
See, among others, \link[nodeSub]{sim_dual_independent} for more details}

\item{node_mutation_rate}{mutation rate on the node.
See, among others, \link[nodeSub]{sim_dual_independent} for more details}

\item{base_frequencies}{the four base frequencies (a, c, g, t) to be
specified to create the rate matrix (i.e. Q matrix)
used to simulate alignments}

\item{node_time}{amount of time spent at the nodes.
See, among others, \link[nodeSub]{sim_dual_independent} for more details}
}
\value{
an alignment of type \code{DNAbin}
}
\description{
Adapter function to simulte the twin alignment
using the \code{linked_node_sub} site model
}
\seealso{
Use \link{sim_twin_alignment_with_unlinked_node_sub_site_model}
to simulate using an unlinked node substitution model.
Use \link{sim_true_alignment_with_linked_node_sub_site_model}
to simulate a true alignment.
}
