% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pirouette.R
\name{splat}
\alias{splat}
\title{splat}
\usage{
splat(x, mtry = 2, retx = TRUE)
}
\arguments{
\item{x}{a sparse matrix}

\item{mtry}{new dimensionality}

\item{retx}{If true, return the splatted matrix as well as the rotation.}
}
\value{
a splatted Matrix
}
\description{
Splat a matrix into a lower dimension.
}
\examples{
splat(matrix(runif(100), ncol=10), 2)
}

