% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_twin_filenames.R
\name{to_twin_filenames}
\alias{to_twin_filenames}
\title{Convert multiple filenames to their twin equivalent}
\usage{
to_twin_filenames(filenames)
}
\description{
Convert multiple filenames to their twin equivalent
}
