% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twin_to_bd_tree.R
\name{twin_to_bd_tree}
\alias{twin_to_bd_tree}
\title{Create an BD twin tree from a phylogeny
and save it as a file}
\usage{
twin_to_bd_tree(phylogeny, twinning_params)
}
\arguments{
\item{phylogeny}{a phylogeny of class \link[ape]{phylo}}

\item{twinning_params}{can be \code{NA} if no twinning is desired,
or can be the twinning parameters,
as can be created by \link{create_twinning_params}}
}
\value{
a twin BD tree of class \code{phylo},
  obtained from the corresponding phylogeny.
}
\description{
Create an BD twin tree from a phylogeny
and save it as a file
}
\author{
Richel J.C. Bilderbeek, Giovanni Laudanno
}
