% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twin_to_yule_tree.R
\name{twin_to_yule_tree}
\alias{twin_to_yule_tree}
\title{Create a twin tree from a phylogeny using a Yule process}
\usage{
twin_to_yule_tree(true_phylogeny, seed = 0, method = "random_tree",
  n_replicates = 10000)
}
\arguments{
\item{true_phylogeny}{a phylogeny of class \link[ape]{phylo}}

\item{seed}{a random number generator seed}

\item{method}{determines how to create the twin tree
\itemize{
    \item 'random_tree' just produces a random tree;
    \item 'max_clade_cred' simulates \code{n_replicates} trees and
      uses \link[phangorn]{maxCladeCred} to create a consensus tree;
    \item 'max_likelihood' simulates \code{n_replicates} trees
     and selects the most likely;
  }}

\item{n_replicates}{number of replicas to evaluate in order to create the
twin tree}
}
\value{
a twin Yule tree of class \link[ape]{phylo},
}
\description{
Create a twin tree from a phylogeny using a Yule process
}
\examples{
library(testthat)

phylogeny <- ape::read.tree(text = "((A:2, B:2):1, C:3);")
twinning_params <- create_twinning_params()
yule_tree <- twin_to_yule_tree(phylogeny)

expect_equal(class(yule_tree), "phylo")

# Branching times will differ, except the crown
expect_false(
  all(
    ape::branching.times(phylogeny) ==
    ape::branching.times(yule_tree)
  )
)

# Crown age stays the same
expect_equal(
  max(ape::branching.times(yule_tree)),
  max(ape::branching.times(phylogeny))
)
}
\author{
Richèl J.C. Bilderbeek, Giovanni Laudanno
}
