% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twin_to_yule_tree.R
\name{twin_to_yule_tree}
\alias{twin_to_yule_tree}
\title{Create a twin tree from a phylogeny using a Yule process}
\usage{
twin_to_yule_tree(phylogeny, twinning_params)
}
\arguments{
\item{phylogeny}{a phylogeny of class \link[ape]{phylo}}

\item{twinning_params}{can be \code{NA} if no twinning is desired,
or can be the twinning parameters,
as can be created by \link{create_twinning_params}}
}
\value{
a twin Yule tree of class \code{phylo},
}
\description{
Create a twin tree from a phylogeny using a Yule process
}
\author{
Richel J.C. Bilderbeek, Giovanni Laudanno
}
