% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/will_measure_evidence.R
\name{will_measure_evidence}
\alias{will_measure_evidence}
\title{Determine if there is an experiment in which the evidence will
be measured.}
\usage{
will_measure_evidence(pir_params)
}
\arguments{
\item{pir_params}{the parameters of \link[pirouette]{pirouette}.
They are created by \link{create_pir_params}.}
}
\description{
Determine if there is an experiment in which the evidence will
be measured.
}
\examples{
library(testthat)

if (rappdirs::app_dir()$os != "win") {
  pir_params <- create_test_pir_params_setup(has_candidate = TRUE)
  expect_true(will_measure_evidence(pir_params))
}

pir_params <- create_test_pir_params_setup(has_candidate = FALSE)
expect_true(!will_measure_evidence(pir_params))

if (rappdirs::app_dir()$os != "win") {
  pir_params$experiments[[1]]$inference_conditions$do_measure_evidence <-
    TRUE
  pir_params$evidence_filename <- get_temp_evidence_filename()
  expect_true(will_measure_evidence(pir_params))
}
}
